% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_dataset.R
\name{acs_dataset}
\alias{acs_dataset}
\title{Create the ACS/PCI bleeding/ischaemia risk dataset}
\usage{
acs_dataset(config_file = "config.yaml")
}
\arguments{
\item{config_file}{}
}
\value{
A tibble of the dataset
}
\description{
.. content for \description{} (no empty lines) ..
}
