% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{add_country_name}
\alias{add_country_name}
\title{Add short country name to referendum data}
\usage{
add_country_name(data)
}
\arguments{
\item{data}{RDB referendum data as returned by \code{\link[=rfrnds]{rfrnds()}}. A data frame that at minimum contains the column \code{country_code} (with \href{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2}{ISO 3166-1 alpha-2} or \href{https://en.wikipedia.org/wiki/ISO_3166-3}{ISO 3166-3 alpha-4} codes).}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Augments \code{data} with an additional column holding the common English name of the country in which the referendum took place.
}
\examples{
rdb::rfrnds() |>
  rdb:::add_country_name() |>
  dplyr::select(id,
                starts_with("country_"))
}
\seealso{
Other referendum data augmentation functions:
\code{\link{add_country_code_continual}()},
\code{\link{add_country_code_long}()},
\code{\link{add_country_name_long}()},
\code{\link{add_former_country_flag}()},
\code{\link{add_period}()},
\code{\link{add_turnout}()},
\code{\link{add_urls}()},
\code{\link{add_world_regions}()}
}
\concept{augment}
