% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{add_former_country_flag}
\alias{add_former_country_flag}
\title{Add \code{is_former_country} flag to referendum data}
\usage{
add_former_country_flag(data)
}
\arguments{
\item{data}{RDB referendum data as returned by \code{\link[=rfrnds]{rfrnds()}}. A data frame that at minimum contains the column \code{country_code} (with \href{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2}{ISO 3166-1 alpha-2} or \href{https://en.wikipedia.org/wiki/ISO_3166-3}{ISO 3166-3 alpha-4} codes).}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Augments \code{data} with an additional column \code{is_former_country} indicating whether or not the column \code{country_code} holds an \href{https://en.wikipedia.org/wiki/ISO_3166-3}{ISO 3166-3 alpha-4 code} referring to a historical country which ceased to exist. \code{is_former_country} being \code{FALSE} means
\code{country_code} holds an \href{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2}{ISO 3166-1 alpha-2 code} instead.
}
\examples{
rdb::rfrnds() |>
  rdb:::add_former_country_flag() |>
  dplyr::select(id,
                starts_with("country_"),
                is_former_country)
}
\seealso{
Other referendum data augmentation functions:
\code{\link{add_country_code_continual}()},
\code{\link{add_country_code_long}()},
\code{\link{add_country_name_long}()},
\code{\link{add_country_name}()},
\code{\link{add_period}()},
\code{\link{add_turnout}()},
\code{\link{add_urls}()},
\code{\link{add_world_regions}()}
}
\concept{augment}
