% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{add_period}
\alias{add_period}
\title{Add period to referendum data}
\usage{
add_period(
  data,
  period = c("week", "month", "quarter", "year", "decade", "century")
)
}
\arguments{
\item{data}{RDB referendum data as returned by \code{\link[=rfrnds]{rfrnds()}}. A data frame that at minimum contains the column \code{date}.}

\item{period}{Type of period to add. One of \code{"week"},\code{"month"},\code{"quarter"},\code{"year"},\code{"decade"} or \code{"century"}.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Augments \code{data} with an additional column holding the specified period in which the referendum took place. The new column is named after \code{period} and its
values are always of type integer.
}
\details{
Note that the \code{period} types \code{"year"}, \code{"decade"} and \code{"century"} are \emph{unique} timespans, while \code{"week"}, \code{"month"} and \code{"quarter"} are \emph{recurring} timespans (e.g. every year has a week 1).
}
\examples{
rdb::rfrnds(quiet = TRUE) |>
  rdb::add_period() |>
  dplyr::select(id, date, week)

rdb::rfrnds(quiet = TRUE) |>
  rdb::add_period("year") |>
  dplyr::select(id, date, year)
}
\seealso{
Other referendum data augmentation functions:
\code{\link{add_country_code_continual}()},
\code{\link{add_country_code_long}()},
\code{\link{add_country_name_long}()},
\code{\link{add_country_name}()},
\code{\link{add_former_country_flag}()},
\code{\link{add_turnout}()},
\code{\link{add_urls}()},
\code{\link{add_world_regions}()}
}
\concept{augment}
