% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d-GEN.R
\name{add_referendums}
\alias{add_referendums}
\title{Add new referendums to the C2D database}
\usage{
add_referendums(
  data,
  email = names(getOption("c2d.credentials")[1L]),
  password = getOption("c2d.credentials")[1L]
)
}
\arguments{
\item{data}{The new referendum data. A \link[tibble:tbl_df-class]{tibble} that must contain an \href{https://rpkg.dev/c2d/articles/codebook.html#id}{\code{id}} column
identifying the referendums plus additional \link[=codebook]{valid columns} containing the values for the corresponding database fields.}

\item{email}{The e-mail address of the C2D API user account to be used for authentication. A character scalar.}

\item{password}{The password of the C2D API user account to be used for authentication. A character scalar.}
}
\value{
\code{data}, invisibly.
}
\description{
Adds new referendum entries to the C2D database via \href{https://github.com/ccmdesign/c2d-app/blob/master/docs/services.md#3-referendum-routes}{its API}.
}
\details{
Note that the following variables are not supported, i.e. simply dropped from \code{data}: \code{files}
}
\seealso{
Other referendum data functions:
\code{\link{count_referendums}()},
\code{\link{download_file_attachment}()},
\code{\link{edit_referendums}()},
\code{\link{referendums}()},
\code{\link{referendum}()},
\code{\link{search_referendums}()},
\code{\link{validate_referendums}()}
}
\concept{referendum}
