% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d.gen.R
\name{add_rfrnds}
\alias{add_rfrnds}
\title{Add new referendums to the C2D database}
\usage{
add_rfrnds(
  data,
  email = pal::pkg_config_val(key = "api_username", pkg = this_pkg),
  password = pal::pkg_config_val(key = "api_password", pkg = this_pkg),
  use_testing_server = pal::pkg_config_val(key = "use_testing_server", pkg = this_pkg),
  quiet = FALSE
)
}
\arguments{
\item{data}{The new referendum data. A \link[tibble:tbl_df-class]{tibble} that in any case must contain the columns
\itemize{
\item \href{https://rpkg.dev/c2d/articles/codebook.html#country_code}{\code{country_code}}
\item \href{https://rpkg.dev/c2d/articles/codebook.html#level}{\code{level}}
\item \href{https://rpkg.dev/c2d/articles/codebook.html#date}{\code{date}}
\item \href{https://rpkg.dev/c2d/articles/codebook.html#title_en}{\code{title_en}}
\item \href{https://rpkg.dev/c2d/articles/codebook.html#result}{\code{result}}
\item \href{https://rpkg.dev/c2d/articles/codebook.html#electorate_total}{\code{electorate_total}}
\item \href{https://rpkg.dev/c2d/articles/codebook.html#electorate_abroad}{\code{electorate_abroad}}
\item \href{https://rpkg.dev/c2d/articles/codebook.html#votes_yes}{\code{votes_yes}}
\item \href{https://rpkg.dev/c2d/articles/codebook.html#votes_no}{\code{votes_no}}
\item \href{https://rpkg.dev/c2d/articles/codebook.html#votes_empty}{\code{votes_empty}}
\item \href{https://rpkg.dev/c2d/articles/codebook.html#votes_invalid}{\code{votes_invalid}}
\item \href{https://rpkg.dev/c2d/articles/codebook.html#is_draft}{\code{is_draft}}
\item \href{https://rpkg.dev/c2d/articles/codebook.html#type}{\code{type}}
}

plus the column \href{https://rpkg.dev/c2d/articles/codebook.html#subnational_entity_name}{\code{subnational_entity_name}} for referendums of
\href{https://rpkg.dev/c2d/articles/codebook.html#subnational_entity_name}{\code{level}} below \code{"national"}, and the column
\href{https://rpkg.dev/c2d/articles/codebook.html#municipality}{\code{municipality}} for referendums of \code{level = "local"},

plus any additional \link[=codebook]{valid} columns containing the values for the corresponding database fields.}

\item{email}{The e-mail address of the C2D API user account to be used for authentication. A character scalar.}

\item{password}{The password of the C2D API user account to be used for authentication. A character scalar.}

\item{use_testing_server}{Whether or not to use the testing servers instead of the production servers for C2D Services API calls etc.}
}
\value{
A character vector of newly created referendum IDs.
}
\description{
Adds new referendum entries to the C2D database via \href{https://github.com/ccmdesign/c2d-app/blob/master/docs/services.md#3-referendum-routes}{its API}.
}
\details{
Note that adding/editing the column \code{files} is not supported, i.e. simply dropped from \code{data}.
}
\seealso{
Other referendum data functions:
\code{\link{count_rfrnds}()},
\code{\link{delete_rfrnds}()},
\code{\link{download_file_attachment}()},
\code{\link{edit_rfrnds}()},
\code{\link{rfrnds}()},
\code{\link{rfrnd}()},
\code{\link{search_rfrnds}()},
\code{\link{validate_rfrnds}()}
}
\concept{rfrnd}
