% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d.gen.R
\name{add_turnout}
\alias{add_turnout}
\title{Add turnout}
\usage{
add_turnout(data, rough = TRUE)
}
\arguments{
\item{data}{C2D referendum data as returned by \code{\link[=rfrnds]{rfrnds()}}. A data frame that at minimum contains the columns \code{electorate_total}, \code{votes_yes}, \code{votes_no},
\code{votes_empty} and \code{votes_invalid}.}

\item{rough}{Whether to fall back on a "rough" calculation of the turnout in case any of the variables \code{votes_empty} or \code{votes_invalid} is unknown (\code{NA}), or
to be strict and return \code{NA} in such a case.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Adds an additional column \code{turnout} with the voter turnout calculated as \verb{sum(votes_*) / electorate_total}.
}
\examples{
# rough turnout numbers
c2d::rfrnds(country_code = "IT") |>
  c2d::add_turnout() |>
  dplyr::select(id, electorate_total, starts_with("votes_"), turnout)

# strict turnout numbers
c2d::rfrnds(country_code = "IT") |>
  c2d::add_turnout(rough = FALSE) |>
  dplyr::select(id, electorate_total, starts_with("votes_"), turnout)
}
\seealso{
Other referendum data augmentation functions:
\code{\link{add_period}()},
\code{\link{add_world_regions}()}
}
\concept{augment}
