% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{add_urls}
\alias{add_urls}
\title{Add various URLs to referendum data}
\usage{
add_urls(data, types = c("sudd", "swissvotes"))
}
\arguments{
\item{data}{RDB referendum data as returned by \code{\link[=rfrnds]{rfrnds()}}. A data frame that at minimum contains the column \code{id_sudd} for \code{types = "sudd"} and the columns
\code{country_code}, \code{level} and \code{id_official} for \code{types = "swissvotes"}.}

\item{types}{Type(s) of URLs to add. One of
\code{"sudd"} or \code{"swissvotes"}.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Augments \code{data} with additional columns holding URLs of the specified \code{types}. The new columns will be named after \code{types}, prefixed with \code{url_}, so
\code{types = "sudd"} will add the column \code{url_sudd} etc.
}
\examples{
rdb::rfrnds() |>
  dplyr::filter(country_code == "CH" & level == "national") |>
  rdb::add_urls() |>
  dplyr::select(id,
                country_code,
                level,
                starts_with("id_"),
                starts_with("url_"))
}
\seealso{
Other referendum data augmentation functions:
\code{\link{add_country_code_continual}()},
\code{\link{add_country_code_long}()},
\code{\link{add_country_name}()},
\code{\link{add_country_name_long}()},
\code{\link{add_former_country_flag}()},
\code{\link{add_period}()},
\code{\link{add_turnout}()},
\code{\link{add_world_regions}()}
}
\concept{augment}
