% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d.gen.R
\name{add_world_regions}
\alias{add_world_regions}
\title{Add UN world regions}
\usage{
add_world_regions(data)
}
\arguments{
\item{data}{A data frame containing at least the column \code{country_code} with two-letter \href{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2}{ISO 3166-1 alpha-2}
codes.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Augments \code{data} with information about the \href{https://en.wikipedia.org/wiki/United_Nations_geoscheme}{United Nations (UN) geoscheme} on three different
grouping tiers based on the \href{https://en.wikipedia.org/wiki/UN_M49#Code_lists}{UN M49 area code hierarchy}.

In total, six different columns are added:
\itemize{
\item \code{un_region_tier_1_code}: UN tier-1 region's M49 area code
\item \code{un_region_tier_1_name}: UN tier-1 region's English name
\item \code{un_region_tier_2_code}: UN tier-2 region's M49 area code
\item \code{un_region_tier_2_name}: UN tier-2 region's English name
\item \code{un_region_tier_3_code}: UN tier-3 region's M49 area code
\item \code{un_region_tier_3_name}: UN tier-3 region's English name
}

Tier-1 regions are the highest, i.e. most aggregated UN regions, commonly referred to as continents. Tier-2 regions are also known as "subregions" and tier-3
regions as "sub-subregions".
}
\details{
Note that only part of all UN tier-2 regions are further divided into UN tier-3 regions, meaning that not all countries are part of a UN tier-3 region. If a
country doesn't belong to any UN tier-3 region, the corresponding \verb{un_region_tier_3_*} values will simply be \code{NA}.
}
\examples{
c2d::referendum(id = "5bbbe26a92a21351232dd73f") |>
  add_world_regions() |>
  dplyr::select(country_name,
                starts_with("un_"))
}
