% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d.gen.R
\name{as_ballot_dates}
\alias{as_ballot_dates}
\title{Transform to ballot-date-level observations}
\usage{
as_ballot_dates(data)
}
\arguments{
\item{data}{C2D referendum data as returned by \code{\link[=rfrnds]{rfrnds()}}. A data frame that at minimum contains the column \code{date}.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Transforms referendum-level observations to ones on the level of ballot date and jurisdiction via \link[tidyr:nest]{nesting} of referendum-level columns. The
individual values of all the referendums on a specific ballot date in a specific jurisdiction are preserved in a list column named \code{rfrnd_data}.
}
\examples{
c2d::rfrnds(country_code = "AT") |> c2d::as_ballot_dates()
}
\seealso{
Other referendum data transformation functions:
\code{\link{n_rfrnds_per_period}()},
\code{\link{printify_col_names}()}
}
\concept{transform}
