% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras_rdb.R
\name{as_num}
\alias{as_num}
\title{Testa se o valor eh numerico. Se nao for, converte.}
\usage{
as_num(i, check = TRUE, fix_na = TRUE)
}
\arguments{
\item{i}{vetor qualquer}

\item{check}{TRUE checa a estrutura do dado}

\item{fix_na}{vetor de strings a serem removidas, caso existam.}
}
\value{
vetor numérico tratado a partir de i
}
\description{
Verifica se o conteúdo do vetor passado é de números. Quando check = TRUE, a estrutura do dado
é testada via regex e se o padrão for detectado, o vetor é definido como numérico.
}
