% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{assert_vars}
\alias{assert_vars}
\title{Assert referendum variables are present}
\usage{
assert_vars(data, var_names)
}
\arguments{
\item{data}{RDB referendum data as returned by \code{\link[=rfrnds]{rfrnds()}}.}

\item{var_names}{Names of the variables to check. A character vector.}
}
\value{
\code{data}, invisibly.
}
\description{
Asserts that the input \link[checkmate:checkDataFrame]{is a data frame} and that the specified variables are present. Depending on \code{var_names}, further
integrity checks are performed.
}
\examples{
rdb::rfrnds() |> rdb::assert_vars(var_names = "country_code")

try(
  tibble::tibble(country_code = "AN") |> rdb::assert_vars(var_names = "country_code")
)
}
\seealso{
Other data functions:
\code{\link{delete_rfrnds}()},
\code{\link{read_tbl}()},
\code{\link{rfrnd_types}()},
\code{\link{rfrnds}()},
\code{\link{rfrnds_old}()},
\code{\link{validate_rfrnds}()}
}
\concept{data}
