% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{backup_rdb}
\alias{backup_rdb}
\title{Back up RDB}
\usage{
backup_rdb(
  path = "rdb_dm.rds",
  overwrite = TRUE,
  connection = connect(user = "rdb_admin", password = pg_role_pw("rdb_admin")),
  disconnect = TRUE
)
}
\arguments{
\item{path}{Path to write the RDB backup \code{.rds} file to.}

\item{overwrite}{Whether or not to overwrite an existing file under \code{path}.}

\item{connection}{\code{\link[RPostgres:PqConnection-class]{PqConnection}}\cr
Database connection to a PostgreSQL DBMS as returned by \code{\link[=connect]{connect()}}.}

\item{disconnect}{\code{\link[base:logical]{logical(1)}}\cr
Whether or not to \link[DBI:dbDisconnect]{terminate} the database connection when finished.}
}
\value{
An object of type \code{\link[dm:dm]{dm::dm}}, invisibly.
}
\description{
Fetches the complete RDB data – excluding fully programmatically administered tables – as a \code{\link[dm:dm]{dm::dm}} object that also includes relationship information
(primary and foreign keys) and writes it to an \code{\link[=saveRDS]{.rds}} file.
}
\seealso{
Other administration functions:
\code{\link{clean_attachment_bucket}()},
\code{\link{config_nocodb_tbls}()},
\code{\link{connect}()},
\code{\link{create_nocodb_tbl_cols}()},
\code{\link{create_nocodb_tbls}()},
\code{\link{delete_rfrnds}()},
\code{\link{dm}()},
\code{\link{notify_postgrest}()},
\code{\link{purge_nocodb}()},
\code{\link{reset_nocodb}()},
\code{\link{reset_rdb}()},
\code{\link{restart_neon_ep}()},
\code{\link{restart_postgrest}()},
\code{\link{restore_rdb}()},
\code{\link{update_administrative_units}()},
\code{\link{update_countries}()},
\code{\link{update_langs}()},
\code{\link{update_municipalities}()},
\code{\link{update_nocodb_users}()},
\code{\link{update_rfrnds}()},
\code{\link{update_subnational_entities}()},
\code{\link{update_tbl}()},
\code{\link{update_topics}()}
}
\concept{admin}
