% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{clean_attachment_bucket}
\alias{clean_attachment_bucket}
\title{Clean referendum attachments}
\usage{
clean_attachment_bucket(
  s3_bucket = s3_bucket_attachments,
  s3_endpoint = s3_endpoint_url,
  s3_access_key = pal::pkg_config_val("nocodb_s3_access_key"),
  s3_access_secret = pal::pkg_config_val("nocodb_s3_access_secret"),
  run_dry = TRUE
)
}
\arguments{
\item{s3_bucket}{Name of the S3-compatible object storage bucket where the referendum attachments reside.}

\item{s3_endpoint}{S3-compatible endpoint URL where \code{s3_bucket} is located.}

\item{s3_access_key}{S3-compatible access key ID with access to \code{s3_bucket}.}

\item{s3_access_secret}{S3-compatible access secret with access to \code{s3_bucket}.}

\item{run_dry}{Whether or not to display the files to be deleted on the console instead of actually deleting any file.}
}
\value{
A character vector of deleted file attachment URLs.
}
\description{
Deletes obsolete attachment files from the specified S3-compatible object storage bucket. An attachment is considered obsolete if it isn't linked to any
referendum in the RDB.
}
\seealso{
Other referendum file attachment functions:
\code{\link{download_attachments}()}

Other administration functions:
\code{\link{backup_rdb}()},
\code{\link{config_nocodb_tbls}()},
\code{\link{connect}()},
\code{\link{create_nocodb_tbl_cols}()},
\code{\link{create_nocodb_tbls}()},
\code{\link{delete_rfrnds}()},
\code{\link{dm}()},
\code{\link{notify_postgrest}()},
\code{\link{purge_nocodb}()},
\code{\link{reset_nocodb}()},
\code{\link{reset_rdb}()},
\code{\link{restart_neon_ep}()},
\code{\link{restart_postgrest}()},
\code{\link{restore_rdb}()},
\code{\link{update_administrative_units}()},
\code{\link{update_countries}()},
\code{\link{update_langs}()},
\code{\link{update_municipalities}()},
\code{\link{update_nocodb_users}()},
\code{\link{update_rfrnds}()},
\code{\link{update_subnational_entities}()},
\code{\link{update_tbl}()},
\code{\link{update_topics}()}
}
\concept{admin}
\concept{attachments}
