% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{clean_rfrnd_attachments}
\alias{clean_rfrnd_attachments}
\title{Clean referendum attachments}
\usage{
clean_rfrnd_attachments(
  s3_endpoint = s3_endpoint_url,
  s3_access_key = pal::pkg_config_val(key = "nocodb_s3_access_key", pkg = this_pkg),
  s3_access_secret = pal::pkg_config_val(key = "nocodb_s3_access_secret", pkg = this_pkg),
  run_dry = TRUE
)
}
\arguments{
\item{s3_endpoint}{S3-compatible endpoint URL where RDB's object storage bucket \code{rdb-attachments} resides.}

\item{s3_access_key}{S3-compatible access key ID with access to the \code{rdb-attachments} bucket.}

\item{s3_access_secret}{S3-compatible access secret with access to the \code{rdb-attachments} bucket.}

\item{run_dry}{Whether or not to display the files to be deleted on the console instead of actually deleting any file.}
}
\value{
A character vector of deleted file attachment URLs.
}
\description{
Deletes obsolete attachment files from the \code{rdb-attachments} object storage bucket. An attachment is considered obsolete if it
isn't linked to any referendum in the RDB.
}
\seealso{
Other administration functions:
\code{\link{config_nocodb_tbls}()},
\code{\link{connect}()},
\code{\link{create_nocodb_tbls}()},
\code{\link{delete_rfrnds}()},
\code{\link{dm}()},
\code{\link{notify_postgrest}()},
\code{\link{purge_nocodb}()},
\code{\link{reset_nocodb}()},
\code{\link{reset_rdb}()},
\code{\link{restart_postgrest}()},
\code{\link{update_countries}()},
\code{\link{update_langs}()},
\code{\link{update_municipalities}()},
\code{\link{update_rfrnds}()},
\code{\link{update_subnational_entities}()},
\code{\link{update_tbl}()},
\code{\link{update_topics}()}
}
\concept{admin}
