% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_utils.R
\name{code_groups}
\alias{code_groups}
\title{Read the code groups defined in the supplied file.}
\usage{
code_groups(file)
}
\arguments{
\item{file}{The code group file to read}
}
\value{
A named list mapping group names to a tibble of codes
}
\description{
Read the code groups defined in the supplied file.
}
