% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{config_ncdb}
\alias{config_ncdb}
\title{Configure NocoDB}
\usage{
config_ncdb(
  base_id = ncdb_base_id(),
  auth_token = pal::pkg_config_val(key = "ncdb_token", pkg = this_pkg)
)
}
\arguments{
\item{base_id}{NocoDB base identifier as returned by \code{\link[=ncdb_base_id]{ncdb_base_id()}}. A character scalar.}

\item{auth_token}{NocoDB API authentication token. A character scalar.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Configures the RDB NocoDB server according to our needs. Currently, this includes:
\itemize{
\item \link[=set_ncdb_display_vals]{Setting the correct display value columm} for all tables.
\item \link[=set_ncdb_tbl_metadata]{Setting the proper metadata} for all tables..
}
}
\seealso{
Other NocoDB API functions
\code{\link{ncdb_api}()},
\code{\link{ncdb_base_id}()},
\code{\link{ncdb_bases}()},
\code{\link{ncdb_tbl}()},
\code{\link{ncdb_tbl_col}()},
\code{\link{ncdb_tbl_col_id}()},
\code{\link{ncdb_tbl_cols}()},
\code{\link{ncdb_tbl_id}()},
\code{\link{ncdb_tbls}()},
\code{\link{reorder_ncdb_tbl}()},
\code{\link{set_ncdb_display_val}()},
\code{\link{set_ncdb_display_vals}()},
\code{\link{set_ncdb_tbl_metadata}()},
\code{\link{update_ncdb_tbl}()},
\code{\link{update_ncdb_tbl_col}()}

Other administration functions:
\code{\link{connect}()},
\code{\link{delete_rfrnds}()},
\code{\link{update_countries}()},
\code{\link{update_langs}()},
\code{\link{update_municipalities}()},
\code{\link{update_rfrnds}()},
\code{\link{update_subnational_entities}()}
}
\concept{admin}
\concept{ncdb}
