% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{config_nocodb}
\alias{config_nocodb}
\title{Configure NocoDB}
\usage{
config_nocodb(
  base_id = nocodb::base_id(title = pal::pkg_config_val(key = nocodb_base_name, pkg =
    this_pkg), hostname = pal::pkg_config_val(key = "nocodb_hostname", pkg = this_pkg),
    auth_token = pal::pkg_config_val(key = "nocodb_api_token", pkg = this_pkg)),
  hostname = pal::pkg_config_val(key = "nocodb_hostname", pkg = this_pkg),
  auth_token = pal::pkg_config_val(key = "nocodb_api_token", pkg = this_pkg),
  quiet = FALSE
)
}
\arguments{
\item{base_id}{NocoDB base identifier as returned by \code{\link[nocodb:base_id]{base_id()}}. A character scalar.}

\item{hostname}{NocoDB server \href{https://en.wikipedia.org/wiki/Hostname}{hostname}. A character scalar.}

\item{auth_token}{NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API authentication token}. A character scalar.}

\item{quiet}{Whether or not to suppress printing status output from internal processing.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Configures the RDB NocoDB server according to our needs. Currently, this includes:
\itemize{
\item Setting the correct display value columm for all tables via \link[nocodb:set_display_vals]{nocodb::set_display_vals}.
\item Setting the proper metadata for all tables via \link[nocodb:set_tbl_metadata]{nocodb::set_tbl_metadata}.
}
}
\seealso{
Other administration functions:
\code{\link{connect}()},
\code{\link{delete_rfrnds}()},
\code{\link{dm}()},
\code{\link{update_countries}()},
\code{\link{update_langs}()},
\code{\link{update_municipalities}()},
\code{\link{update_rfrnds}()},
\code{\link{update_subnational_entities}()},
\code{\link{update_tbl}()}
}
\concept{admin}
\concept{nocodb}
