% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{config_nocodb_tbl_cols}
\alias{config_nocodb_tbl_cols}
\title{Configure NocoDB table columns}
\usage{
config_nocodb_tbl_cols(
  hostname = pal::pkg_config_val(key = "nocodb_host", pkg = this_pkg),
  quiet = FALSE
)
}
\arguments{
\item{hostname}{NocoDB server \href{https://en.wikipedia.org/wiki/Hostname}{hostname}. A character scalar.}

\item{quiet}{Whether or not to suppress printing status output from internal processing.}
}
\value{
\code{hostname}, invisibly.
}
\description{
Configures the RDB table columns on the specified NocoDB server. Currently, this includes:
\itemize{
\item Setting ...
}
}
\seealso{
Other NocoDB functions
\code{\link{config_nocodb_tbls}()},
\code{\link{create_nocodb_tbl_cols}()},
\code{\link{create_nocodb_tbls}()},
\code{\link{purge_nocodb}()},
\code{\link{reset_nocodb}()},
\code{\link{reset_rdb}()}

Other administration functions:
\code{\link{clean_attachment_bucket}()},
\code{\link{config_nocodb_tbls}()},
\code{\link{connect}()},
\code{\link{create_nocodb_tbl_cols}()},
\code{\link{create_nocodb_tbls}()},
\code{\link{delete_rfrnds}()},
\code{\link{dm}()},
\code{\link{notify_postgrest}()},
\code{\link{purge_nocodb}()},
\code{\link{reset_nocodb}()},
\code{\link{reset_rdb}()},
\code{\link{restart_neon_ep}()},
\code{\link{restart_postgrest}()},
\code{\link{update_administrative_units}()},
\code{\link{update_countries}()},
\code{\link{update_langs}()},
\code{\link{update_municipalities}()},
\code{\link{update_rfrnds}()},
\code{\link{update_subnational_entities}()},
\code{\link{update_tbl}()},
\code{\link{update_topics}()}
}
\concept{admin}
\concept{nocodb}
\keyword{internal}
