% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d-GEN.R
\name{count_referendums}
\alias{count_referendums}
\title{Count number of referendums}
\usage{
count_referendums(
  is_draft = FALSE,
  country_code = NULL,
  subnational_entity_name = NULL,
  municipality = NULL,
  level = NULL,
  type = NULL,
  date = NULL,
  date_time_created_min = NULL,
  date_time_created_max = NULL,
  query_filter = NULL
)
}
\arguments{
\item{is_draft}{\code{TRUE} means to include only referendum entries with \emph{draft} status, \code{FALSE} to include only normal entries. Set to \code{NULL} in order to
include both draft and normal entries.}

\item{country_code}{The \code{country_code}(s) to be included. A character vector.}

\item{subnational_entity_name}{The \code{subnational_entity_name}(s) to be included. A character vector.}

\item{municipality}{The \code{municipality}(s) to be included. A character vector.}

\item{level}{The \code{level}(s) to be included. A character vector.}

\item{type}{The \code{type}(s) to be included. A character vector.}

\item{date}{The \code{date}(s) to be included. A character vector.}

\item{date_time_created_min}{The minimum \code{date_time_created} to be included. A \link[base:Dates]{date}, a \link[base:DateTimeClasses]{datetime}, or something
coercible to (like \code{"2006-01-02"} or \code{"2006-01-02T15:04:05"}).}

\item{date_time_created_max}{The maximum \code{date_time_created} to be included. A \link[base:Dates]{date}, a \link[base:DateTimeClasses]{datetime}, or something
coercible to (like \code{"2006-01-02"} or \code{"2006-01-02T15:04:05"}).}

\item{query_filter}{A valid \href{https://docs.mongodb.com/manual/core/document/#query-filter-documents}{MongoDB JSON query filter document} which allows for
maximum control over what data is included. This takes precedence over all the other parameters, i.e. if \code{query_filter} is provided, all the other
parameters are ignored.}
}
\value{
A named list with \code{level} as names and referendum counts as values.
}
\description{
Counts the number of referendums in the C2D Database.
}
\examples{
# the whole database (excl. drafts)
c2d::count_referendums()

# only Swiss and Austrian referendums
c2d::count_referendums(country_code = c("CH", "AT"))

# only Swiss referendums between 2020 and 2021
c2d::count_referendums(country_code = "CH",
                       date_time_created_min = "2020-01-01",
                       date_time_created_max = "2021-01-01")
}
