% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d.gen.R
\name{count_rfrnds}
\alias{count_rfrnds}
\title{Count number of referendums}
\usage{
count_rfrnds(
  is_draft = FALSE,
  country_code = NULL,
  subnational_entity_name = NULL,
  municipality = NULL,
  level = NULL,
  type = NULL,
  date_min = NULL,
  date_max = NULL,
  date_time_created_min = NULL,
  date_time_created_max = NULL,
  date_time_last_edited_min = NULL,
  date_time_last_edited_max = NULL,
  query_filter = NULL,
  use_testing_server = pal::pkg_config_val(key = "use_testing_server", pkg = this_pkg)
)
}
\arguments{
\item{is_draft}{\code{TRUE} means to include only referendum entries with \emph{draft} status, \code{FALSE} to include only normal entries. Set to \code{NULL} in order to
include both draft and normal entries.}

\item{country_code}{The \code{country_code}(s) to be included. A character vector.}

\item{subnational_entity_name}{The \code{subnational_entity_name}(s) to be included. A character vector.}

\item{municipality}{The \code{municipality}(s) to be included. A character vector.}

\item{level}{The \code{level}(s) to be included. A character vector.}

\item{type}{The \code{type}(s) to be included. A character vector.}

\item{date_min}{The minimum \code{date} to be included. A \link[=Date]{date} or something coercible to.}

\item{date_max}{The maximum \code{date} to be included. A \link[=Date]{date} or something coercible to.}

\item{date_time_created_min}{The minimum \code{date_time_created} to be included. A \link[base:DateTimeClasses]{datetime}, or something coercible to (like
\code{"2006-01-02"} or \code{"2006-01-02T15:04:05Z"}; assumed to be in UTC if no timezone is given).}

\item{date_time_created_max}{The maximum \code{date_time_created} to be included. A \link[base:DateTimeClasses]{datetime}, or something coercible to (like
\code{"2006-01-02"} or \code{"2006-01-02T15:04:05Z"}; assumed to be in UTC if no timezone is given).}

\item{date_time_last_edited_min}{The minimum \code{date_time_last_edited} to be included. A \link[base:DateTimeClasses]{datetime}, or something coercible to (like
\code{"2006-01-02"} or \code{"2006-01-02T15:04:05Z"}; assumed to be in UTC if no timezone is given).}

\item{date_time_last_edited_max}{The maximum \code{date_time_last_edited} to be included. A \link[base:DateTimeClasses]{datetime}, or something coercible to (like
\code{"2006-01-02"} or \code{"2006-01-02T15:04:05Z"}; assumed to be in UTC if no timezone is given).}

\item{query_filter}{A valid \href{https://docs.mongodb.com/manual/core/document/#query-filter-documents}{MongoDB JSON query filter document} which allows for
maximum control over what data is included. This takes precedence over all of the above listed parameters, i.e. if \code{query_filter} is provided, the
parameters \code{country_code}, \code{subnational_entity_name}, \code{municipality}, \code{level}, \code{type}, \code{date_min}, \code{date_max}, \code{is_draft}, \code{date_time_created_min}, \code{date_time_created_max}, \code{date_time_last_edited_min} and \code{date_time_last_edited_max} are ignored.}

\item{use_testing_server}{Whether or not to use the testing servers instead of the production servers for C2D Services API calls etc.}
}
\value{
A named list with \code{level} as names and referendum counts as values.
}
\description{
Counts the number of referendums per \href{https://rpkg.dev/c2d/articles/codebook.html#level}{\code{level}} in the C2D Database.
}
\examples{
# the whole database (excl. drafts)
c2d::count_rfrnds()

# only Swiss and Austrian referendums
c2d::count_rfrnds(country_code = c("CH", "AT"))

# only Swiss referendums created between 2020 and 2021
c2d::count_rfrnds(country_code = "CH",
                  date_time_created_min = "2020-01-01",
                  date_time_created_max = "2021-01-01")
}
\seealso{
Other referendum data functions:
\code{\link{add_rfrnds}()},
\code{\link{delete_rfrnds}()},
\code{\link{download_file_attachment}()},
\code{\link{edit_rfrnds}()},
\code{\link{rfrnd_exists}()},
\code{\link{rfrnds}()},
\code{\link{rfrnd}()},
\code{\link{search_rfrnds}()},
\code{\link{validate_rfrnds}()}
}
\concept{rfrnd}
