% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{create_ncdb_base}
\alias{create_ncdb_base}
\title{Create NocoDB base}
\usage{
create_ncdb_base(
  title = ncdb_base_name,
  description = NULL,
  color = color_rdb_main,
  show_null_and_empty_in_filter = TRUE,
  auth_token = pal::pkg_config_val(key = "ncdb_token", pkg = this_pkg)
)
}
\arguments{
\item{title}{Title of the new base. A character scalar.}

\item{description}{Description of the new base. A character scalar.}

\item{color}{Color to use for the base icon. A hexadecimal color code like \code{"#3f4eff"}.}

\item{show_null_and_empty_in_filter}{Whether or not to distinguish \code{NULL}s from empty values in column filters. If \code{FALSE}, \code{NULL} and empty fields are
treated alike.}

\item{auth_token}{NocoDB API authentication token. A character scalar.}
}
\value{
A \link[tibble:tbl_df-class]{tibble} containing metadata about the newly created NocoDB base.
}
\description{
Adds a new base on the RDB NocoDB server via the
\href{https://meta-apis-v2.nocodb.com/#tag/Base/operation/base-create}{\code{POST /api/v2/meta/bases}} API endpoint.
}
\seealso{
Other NocoDB API functions
\code{\link{config_ncdb}()},
\code{\link{create_ncdb_user}()},
\code{\link{ncdb_api}()},
\code{\link{ncdb_base}()},
\code{\link{ncdb_base_id}()},
\code{\link{ncdb_bases}()},
\code{\link{ncdb_tbl}()},
\code{\link{ncdb_tbl_col}()},
\code{\link{ncdb_tbl_col_id}()},
\code{\link{ncdb_tbl_cols}()},
\code{\link{ncdb_tbl_id}()},
\code{\link{ncdb_tbls}()},
\code{\link{ncdb_users}()},
\code{\link{reorder_ncdb_tbl}()},
\code{\link{set_ncdb_display_val}()},
\code{\link{set_ncdb_display_vals}()},
\code{\link{set_ncdb_tbl_metadata}()},
\code{\link{update_ncdb_base}()},
\code{\link{update_ncdb_tbl}()},
\code{\link{update_ncdb_tbl_col}()},
\code{\link{upload_ncdb_attachments}()}
}
\concept{ncdb}
\keyword{internal}
