% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{create_ncdb_user}
\alias{create_ncdb_user}
\title{Create NocoDB user}
\usage{
create_ncdb_user(
  email,
  role = c("no-access", "commenter", "editor", "guest", "owner", "viewer", "creator"),
  base_id = ncdb_base_id(),
  auth_token = pal::pkg_config_val(key = "ncdb_token", pkg = this_pkg),
  quiet = FALSE
)
}
\arguments{
\item{email}{E-mail address of the new user.}

\item{role}{Base role to assign to the new user. One of \code{"no-access"}, \code{"commenter"}, \code{"editor"}, \code{"guest"}, \code{"owner"}, \code{"viewer"} or \code{"creator"}.}

\item{base_id}{NocoDB base identifier as returned by \code{\link[=ncdb_base_id]{ncdb_base_id()}}. A character scalar.}

\item{auth_token}{NocoDB API authentication token. A character scalar.}
}
\value{
\code{email}, invisibly.
}
\description{
Adds a new user account to the specified base on the RDB NocoDB server via the
\href{https://meta-apis-v2.nocodb.com/#tag/Auth/operation/auth-base-user-add}{POST \verb{/api/v2/meta/bases/\{base_id\}/users}} API endpoint.
}
\seealso{
Other NocoDB API functions
\code{\link{config_ncdb}()},
\code{\link{ncdb_api}()},
\code{\link{ncdb_base_id}()},
\code{\link{ncdb_bases}()},
\code{\link{ncdb_tbl}()},
\code{\link{ncdb_tbl_col}()},
\code{\link{ncdb_tbl_col_id}()},
\code{\link{ncdb_tbl_cols}()},
\code{\link{ncdb_tbl_id}()},
\code{\link{ncdb_tbls}()},
\code{\link{ncdb_users}()},
\code{\link{reorder_ncdb_tbl}()},
\code{\link{set_ncdb_display_val}()},
\code{\link{set_ncdb_display_vals}()},
\code{\link{set_ncdb_tbl_metadata}()},
\code{\link{update_ncdb_tbl}()},
\code{\link{update_ncdb_tbl_col}()},
\code{\link{upload_ncdb_attachments}()}
}
\concept{ncdb}
\keyword{internal}
