% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{create_nocodb_tbl_cols}
\alias{create_nocodb_tbl_cols}
\title{Create NocoDB table meta columns}
\usage{
create_nocodb_tbl_cols(origin = pal::pkg_config_val("nocodb_origin"))
}
\arguments{
\item{origin}{\code{\link[base:character]{character(1)}}\cr
NocoDB server \href{https://developer.mozilla.org/docs/Glossary/Origin}{origin}.}
}
\value{
\code{origin}, invisibly.
}
\description{
Adds the PostgreSQL meta columns \code{created_by}, \code{updated_by}, \code{created_at} and \code{updated_at} to all RDB tables via NocoDB's API.
}
\seealso{
Other NocoDB functions
\code{\link{config_nocodb_tbls}()},
\code{\link{create_nocodb_tbls}()},
\code{\link{purge_nocodb}()},
\code{\link{reset_nocodb}()},
\code{\link{reset_rdb}()}

Other administration functions:
\code{\link{backup_rdb}()},
\code{\link{clean_attachment_bucket}()},
\code{\link{config_nocodb_tbls}()},
\code{\link{connect}()},
\code{\link{create_nocodb_tbls}()},
\code{\link{delete_rfrnds}()},
\code{\link{dm}()},
\code{\link{notify_postgrest}()},
\code{\link{purge_nocodb}()},
\code{\link{reset_nocodb}()},
\code{\link{reset_rdb}()},
\code{\link{restart_neon_ep}()},
\code{\link{restart_postgrest}()},
\code{\link{restore_rdb}()},
\code{\link{update_administrative_units}()},
\code{\link{update_countries}()},
\code{\link{update_langs}()},
\code{\link{update_municipalities}()},
\code{\link{update_nocodb_users}()},
\code{\link{update_rfrnds}()},
\code{\link{update_subnational_entities}()},
\code{\link{update_tbl}()},
\code{\link{update_topics}()}
}
\concept{admin}
\concept{nocodb}
\keyword{internal}
