% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\docType{data}
\name{data_iso_3166_1}
\alias{data_iso_3166_1}
\title{ISO 3166-1 data}
\format{
A \link[tibble:tbl_df-class]{tibble}.
}
\usage{
data_iso_3166_1
}
\description{
\href{https://en.wikipedia.org/wiki/ISO_3166-1}{ISO 3166-1} data, corrected and extended by unofficial information for countries which are not covered by the ISO
standard yet. See also the \href{https://www.statcan.gc.ca/en/subjects/standard/sccai/2019/sccai2019-iso3166-12013}{Differences between SCCAI 2019 and ISO 3166-1:2013} for inspiration on name corrections.
}
\examples{
rdb::data_iso_3166_1
}
\seealso{
Other auxiliary data:
\code{\link{data_iso_3166_2}},
\code{\link{data_iso_3166_3}},
\code{\link{data_iso_639_1}},
\code{\link{data_municipalities}},
\code{\link{data_topics}}
}
\concept{aux_data}
\keyword{datasets}
