% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\docType{data}
\name{data_iso_3166_2}
\alias{data_iso_3166_2}
\title{ISO 3166-2 data}
\format{
A \link[tibble:tbl_df-class]{tibble}.
}
\usage{
data_iso_3166_2
}
\description{
\href{https://en.wikipedia.org/wiki/ISO_3166-2}{ISO 3166-2} data, extended by
\itemize{
\item unofficial information for subnational entities which ceased to exist before the ISO standard was first published in 1998 – and hence don't have an
official ISO 3166-2 code assigned. Such unofficial codes are of the form \code{AB-CUSTOM-1}, \code{AB-CUSTOM-2}, etc., where \code{AB} is the official \href{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2}{ISO 3166-1 alpha-2 country code} and \code{CUSTOM-1}, \code{CUSTOM-2} etc. are sequential subdivision codes. Note that these custom
codes are assigned in random order and have no further meaning.
\item information on the period of validity. Due to lack of official revocation information in ISO 3166-2, we construct our own (delayed) \code{valid_from} and
\code{valid_to} date columns based on detected changes in \code{\link[ISOcodes:ISO_3166]{ISOcodes::ISO_3166_2}}.
}
}
\examples{
rdb::data_iso_3166_2
}
\seealso{
Other auxiliary data:
\code{\link{data_iso_3166_1}},
\code{\link{data_iso_3166_3}},
\code{\link{data_iso_639_1}},
\code{\link{data_municipalities}},
\code{\link{data_topics}}
}
\concept{aux_data}
\keyword{datasets}
