% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\docType{data}
\name{data_iso_3166_3}
\alias{data_iso_3166_3}
\title{ISO 3166-3 data}
\format{
A \link[tibble:tbl_df-class]{tibble}.
}
\usage{
data_iso_3166_3
}
\description{
\href{https://en.wikipedia.org/wiki/ISO_3166-3}{ISO 3166-3} data, extended by some deduced information.
}
\examples{
rdb::data_iso_3166_3
}
\seealso{
Other auxiliary data:
\code{\link{data_iso_3166_1}},
\code{\link{data_iso_3166_2}},
\code{\link{data_iso_639_1}},
\code{\link{data_municipalities}}
}
\concept{aux_data}
\keyword{datasets}
