% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\docType{data}
\name{data_iso_639_1}
\alias{data_iso_639_1}
\title{ISO 639-1 data}
\format{
A \link[tibble:tbl_df-class]{tibble}.
}
\usage{
data_iso_639_1
}
\description{
\href{https://en.wikipedia.org/wiki/ISO_639-1}{ISO 639-1} data.
}
\examples{
rdb::data_iso_639_1
}
\seealso{
Other auxiliary data:
\code{\link{data_iso_3166_1}},
\code{\link{data_iso_3166_2}},
\code{\link{data_iso_3166_3}},
\code{\link{data_municipalities}},
\code{\link{data_topics}}
}
\concept{aux_data}
\keyword{datasets}
