% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\docType{data}
\name{data_municipalities}
\alias{data_municipalities}
\title{Municipality data}
\format{
A \link[tibble:tbl_df-class]{tibble}.
}
\usage{
data_municipalities
}
\description{
Municipality data...
}
\details{
Currently, municipalities in the following countries are covered by this dataset:
\itemize{
\item \href{https://en.wikipedia.org/wiki/Community_Identification_Number#Switzerland}{Switzerland}
}
}
\examples{
rdb::data_municipalities
}
\seealso{
Other auxiliary data:
\code{\link{data_iso_3166_1}},
\code{\link{data_iso_3166_2}},
\code{\link{data_iso_3166_3}},
\code{\link{data_iso_639_1}}
}
\concept{aux_data}
\keyword{datasets}
