% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d.gen.R
\docType{data}
\name{data_topics}
\alias{data_topics}
\title{Topic hierarchy}
\format{
A \link[tibble:tbl_df-class]{tibble}.
}
\usage{
data_topics
}
\description{
A tibble reflecting the complete \href{https://rpkg.dev/c2d/articles/codebook.html#topics}{referendum topics hierarchy}.
}
\examples{
c2d::data_topics
}
\seealso{
Other referendum topic functions:
\code{\link{hierarchize_topics_fast}()},
\code{\link{hierarchize_topics}()},
\code{\link{infer_topics}()},
\code{\link{topics}()}
}
\concept{topics}
\keyword{datasets}
