% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{delete_rfrnds}
\alias{delete_rfrnds}
\title{Delete referendums from the RDB}
\usage{
delete_rfrnds(data, connection = connect(), disconnect = TRUE)
}
\arguments{
\item{data}{Referendum data to be deleted. A \link[tibble:tbl_df-class]{tibble} that at least must contain the column
\code{id}.}

\item{connection}{Database connection to \link[=connect]{RDB's PostgreSQL DBMS}. An S4 object that inherits from \code{\link[DBI:DBIConnection-class]{DBI::DBIConnection}}.}

\item{disconnect}{Whether or not to \link[DBI:dbDisconnect]{terminate} the database connection when finished.}
}
\value{
The updated \code{referendums} table data as a tibble, invisibly.
}
\description{
Deletes \code{data} from the RDB's \code{referendums} table.
}
\seealso{
Other administration functions:
\code{\link{config_ncdb}()},
\code{\link{connect}()},
\code{\link{dm}()},
\code{\link{update_countries}()},
\code{\link{update_langs}()},
\code{\link{update_municipalities}()},
\code{\link{update_rfrnds}()},
\code{\link{update_subnational_entities}()}

Other data functions:
\code{\link{assert_vars}()},
\code{\link{download_file_attachment}()},
\code{\link{read_tbl}()},
\code{\link{rfrnd_types}()},
\code{\link{rfrnds}()},
\code{\link{rfrnds_old}()},
\code{\link{validate_rfrnds}()}
}
\concept{admin}
\concept{data}
