% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{dm}
\alias{dm}
\title{Get RDB data model}
\usage{
dm(
  connection = connect(user = Sys.getenv("R_RDB_PG_OWNER"), password =
    Sys.getenv("R_RDB_PG_OWNER_PASSWORD"))
)
}
\arguments{
\item{connection}{Database connection to \link[=connect]{RDB's PostgreSQL DBMS}. An S4 object that inherits from \code{\link[DBI:DBIConnection-class]{DBI::DBIConnection}}.}
}
\value{
A \link[dm:dm]{data model} object.
}
\description{
Get RDB data model
}
\examples{
rdb::dm() |> dm::dm_draw(view_type = "all")
}
\seealso{
Other administration functions:
\code{\link{config_ncdb}()},
\code{\link{connect}()},
\code{\link{delete_rfrnds}()},
\code{\link{update_countries}()},
\code{\link{update_langs}()},
\code{\link{update_municipalities}()},
\code{\link{update_rfrnds}()},
\code{\link{update_subnational_entities}()},
\code{\link{update_tbl}()}
}
\concept{admin}
