% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{dm_excl_gen_always_cols}
\alias{dm_excl_gen_always_cols}
\title{Remove \verb{GENERATED ALWAYS} columns from \code{dm}}
\usage{
dm_excl_gen_always_cols(dm, connection = dm::dm_get_con(dm))
}
\arguments{
\item{dm}{A \link[dm:dm]{data model} object.}
}
\value{
A \link[dm:dm]{data model} object.
}
\description{
Removes \verb{GENERATED ALWAYS} columns from a data model object.
}
\examples{
rdb::dm() |> rdb:::dm_excl_gen_always_cols()
}
\keyword{internal}
