% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{download_attachments}
\alias{download_attachments}
\title{Download file attachments}
\usage{
download_attachments(
  data,
  path = ".",
  filename_from_title = FALSE,
  fileext_from_mimetype = !filename_from_title
)
}
\arguments{
\item{data}{RDB referendum data as returned by \code{\link[=rfrnds]{rfrnds()}}. A data frame that at minimum contains the columns \code{id} and \code{attachments}.}

\item{path}{Path to the directory where the downloaded file(s) are to be written to.}

\item{filename_from_title}{Whether to use an attachment's \code{title} field value as the filename. Note that attachment titles are \strong{not guaranteed to be
unique}, i.e. there can be multiple file attachments with the same title (but differing content, of course). If \code{FALSE}, the referendum identifier plus a
sequential number is used as an attachment's filename.}

\item{fileext_from_mimetype}{Whether to add a file extension suffix to the filename which is determined from the attachment's \code{mimetype} field.}
}
\value{
\code{data}, invisibly.
}
\description{
Downloads file attachments from the Referendum Database (RDB). ... the \code{attachments} list column returned by \code{\link[=rfrnds]{rfrnds()}}.
}
\examples{
# get the first three referendums which contain attachments
d <-
  rdb::rfrnds() |>
  dplyr::filter(purrr::map_lgl(attachments, \(x) !is.null(x))) |>
  dplyr::filter(dplyr::row_number() \%in\% 1:3)

# download the attachments to the current directory
d |> rdb::download_attachments()

# and delete them again
d$attachments |>
  purrr::map(\(x) x$title) |>
  unlist() |>
  fs::file_delete()
}
\seealso{
Other referendum file attachment functions:
\code{\link{clean_attachment_bucket}()}
}
\concept{attachments}
