% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{exec_sql_file}
\alias{exec_sql_file}
\title{Executes SQL script}
\usage{
exec_sql_file(.path, .connection = connect(), .disconnect = TRUE, ...)
}
\arguments{
\item{.path}{Path to a file containing SQL statements.}

\item{.connection}{Database connection as returned by \code{\link[=connect]{connect()}}. An S4 object that inherits from \code{\link[DBI:DBIConnection-class]{DBI::DBIConnection}}.}

\item{...}{Further (named) arguments used as temporary variables available for substitution while interpolating the SQL statements via
\code{\link[DBI:sqlInterpolate]{DBI::sqlInterpolate()}}.}
}
\value{
\code{path}, invisibly.
}
\description{
Reads an SQL script using \code{\link[=read_sql_file]{read_sql_file()}} and executes it as an atomic SQL transaction using \code{\link[DBI:dbWithTransaction]{DBI::dbWithTransaction()}}.
}
\seealso{
Other PostgreSQL functions
\code{\link{pg_col_metadata}()},
\code{\link{pg_drop_tbls}()},
\code{\link{pg_has_tbl}()},
\code{\link{pg_init_aux_tbls}()},
\code{\link{pg_init_db}()},
\code{\link{pg_init_main_tbls}()},
\code{\link{pg_pk}()},
\code{\link{pg_reset_db}()},
\code{\link{pg_tbl_del}()},
\code{\link{pg_tbl_keep}()},
\code{\link{pg_tbl_update}()},
\code{\link{read_sql_file}()},
\code{\link{split_sql_str}()}
}
\concept{pg}
\keyword{internal}
