% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_utils.R
\name{find_code_duplicates}
\alias{find_code_duplicates}
\title{Return a tibble of duplicated codes in code groups}
\usage{
find_code_duplicates(code_groups)
}
\arguments{
\item{code_groups}{A named list of code groups (named
by the name of the group}
}
\value{
A tibble of duplicated codes
}
\description{
Returns a flat tibble of duplicated codes and descriptions,
one row per instance of the code in a group, which a column
"duplicate_count" indicating how many times a code has been
duplicated. The returned tibble is arranged by the code
column to see duplicated groups together.
}
