% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{ggplot_streamgraph}
\alias{ggplot_streamgraph}
\title{Period streamgraph}
\usage{
ggplot_streamgraph(
  data,
  period = c("week", "month", "quarter", "year", "decade", "century"),
  by,
  bandwidth = 0.75
)
}
\arguments{
\item{data}{RDB referendum data as returned by \code{\link[=rfrnds]{rfrnds()}}. A data frame that at minimum contains the column specified in \code{period} or the column \code{date} (to
compute the \link[=add_period]{period column}), plus the one(s) specified via \code{by} (if any).}

\item{period}{Type of period to count referendums by. One of \code{"week"},\code{"month"},\code{"quarter"},\code{"year"},\code{"decade"} or \code{"century"}.}

\item{by}{\code{data} column to group by before counting number of referendums.}

\item{bandwidth}{Kernel density estimation bandwidth. A numeric scalar.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2} object.
}
\description{
Creates a ggplot2 \href{https://en.wikipedia.org/wiki/Streamgraph}{streamgraph} based on \code{\link[ggstream:geom_stream]{ggstream::geom_stream()}} using the specified \code{period} as time
resolution.
}
\examples{
rdb::rfrnds(country_code = "CH",
            level = "national",
            quiet = TRUE,
            max_cache_age = "1 year") |>
  rdb::ggplot_streamgraph(period = "year",
                          by = topics_tier_1)
}
\seealso{
Other visualization functions:
\code{\link{plot_rfrnd_share_per_period}()},
\code{\link{plot_topic_segmentation}()},
\code{\link{plot_topic_share_per_period}()}
}
\concept{visualize}
