% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d.gen.R
\name{hierarchize_tags_fast}
\alias{hierarchize_tags_fast}
\title{Hierarchize tags (fast)}
\usage{
hierarchize_tags_fast(
  tags_tier_1 = character(),
  tags_tier_2 = character(),
  tags_tier_3 = character()
)
}
\arguments{
\item{tags_tier_1}{First-tier tags. A character vector.}

\item{tags_tier_2}{Second-tier tags. A character vector.}

\item{tags_tier_3}{Third-tier tags. A character vector.}
}
\value{
A \link[tibble:tbl_df-class]{tibble} with the columns \code{tag_tier_1}, \code{tag_tier_2} and \code{tag_tier_3}.
}
\description{
Reconstructs the hierarchical relations between the three tag variables \code{tags_tier_1}, \code{tags_tier_2} and \code{tags_tier_3}. Other than \code{\link[=hierarchize_tags]{hierarchize_tags()}}, this
function assumes that the three tag variables are always \emph{complete}, i.e. that no (grand)parent tags of lower-tier tags are missing. This assumption is met
by \code{\link[=rfrnds]{rfrnds()}} and \code{\link[=rfrnd]{rfrnd()}}.
}
\examples{
library(magrittr)

c2d::rfrnd(id = "5bbbe26a92a21351232dd73f") \%$\%
  c2d:::hierarchize_tags_fast(unlist(tags_tier_1),
                              unlist(tags_tier_2),
                              unlist(tags_tier_3))

# hierarchize the tags of all Austrian referendums
c2d::rfrnds(country_code = "AT") |>
  dplyr::group_split(id) |>
  purrr::map(~ c2d:::hierarchize_tags_fast(unlist(.x$tags_tier_1),
                                           unlist(.x$tags_tier_2),
                                           unlist(.x$tags_tier_3)))
}
\seealso{
Other tag-related functions:
\code{\link{hierarchize_tags}()},
\code{\link{infer_tags}()},
\code{\link{tags}()}
}
\concept{tags}
