% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{hierarchize_topics}
\alias{hierarchize_topics}
\title{Hierarchize topics}
\usage{
hierarchize_topics(x)
}
\arguments{
\item{x}{The topics to hierarchize. Either a character vector of topics or a single-row data frame containing at least the columns \code{topics_tier_1},
\code{topics_tier_2} and \code{topics_tier_3}.}
}
\value{
A \link[tibble:tbl_df-class]{tibble} with the columns \code{topic_tier_1}, \code{topic_tier_2} and \code{topic_tier_3}.
}
\description{
Reconstructs the hierarchical relations between the three topic variables \code{topics_tier_1}, \code{topics_tier_2} and \code{topics_tier_3}. Can also be used to simply
determine the parent topic(s) of any topic.
}
\examples{
rdb::hierarchize_topics("territorial questions")

# hierarchize the topics of all Austrian referendums
rdb::rfrnds() |>
  dplyr::filter(country_code == "AT") |>
  dplyr::group_split(id) |>
  purrr::map(rdb::hierarchize_topics)
}
\seealso{
Other referendum topic functions:
\code{\link{data_topics}},
\code{\link{hierarchize_topics_fast}()},
\code{\link{infer_topics}()},
\code{\link{topics}()}
}
\concept{topics}
