% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{hierarchize_topics_fast}
\alias{hierarchize_topics_fast}
\title{Hierarchize topics (fast)}
\usage{
hierarchize_topics_fast(
  topics_tier_1 = character(),
  topics_tier_2 = character(),
  topics_tier_3 = character()
)
}
\arguments{
\item{topics_tier_1}{First-tier topics. A character vector.}

\item{topics_tier_2}{Second-tier topics. A character vector.}

\item{topics_tier_3}{Third-tier topics. A character vector.}
}
\value{
A \link[tibble:tbl_df-class]{tibble} with the columns \code{topic_tier_1}, \code{topic_tier_2} and \code{topic_tier_3}.
}
\description{
Reconstructs the hierarchical relations between the three topic variables \code{topics_tier_1}, \code{topics_tier_2} and \code{topics_tier_3}. Other than
\code{\link[=hierarchize_topics]{hierarchize_topics()}}, this function assumes that the three topic variables are always \emph{complete}, i.e. that no (grand)parent topics of lower-tier topics
are missing. This assumption is met by the data returned by \code{\link[=rfrnds]{rfrnds()}}.
}
\examples{
# hierarchize the topics of all Austrian referendums
rdb::rfrnds() |>
  dplyr::filter(country_code == "AT") |>
  dplyr::group_split(id) |>
  purrr::map(~ rdb::hierarchize_topics_fast(unlist(.x$topics_tier_1),
                                            unlist(.x$topics_tier_2),
                                            unlist(.x$topics_tier_3)))
}
\seealso{
Other referendum topic functions:
\code{\link{data_topics}},
\code{\link{hierarchize_topics}()},
\code{\link{infer_topics}()},
\code{\link{topics}()}
}
\concept{topics}
