% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d-GEN.R
\name{infer_tags}
\alias{infer_tags}
\title{Infer higher-tier tags}
\usage{
infer_tags(tags, tier = 1L)
}
\arguments{
\item{tags}{The tags to determine the corresponding (grand)parent tags of. A factor or character vector.}

\item{tier}{The tier of the inferred tags. Should be >= than the tier of the \code{tags}. An integer scalar.}
}
\value{
A character vector.
}
\description{
Determines the top-tier (\code{tier = 1L}) or second-tier (\code{tier = 2L}) tags that are the parents or grandparents of \code{tags} in the
\href{https://rpkg.dev/c2d/articles/codebook.html#tags}{hierarchy}.
}
\examples{
c2d::infer_tags(tags = c("EU", "animal protection"),
                tier = 1L)
c2d::infer_tags(tags = c("EU", "animal protection"),
                tier = 2L)

# tags of different tiers can mixed in `tags`
c2d::infer_tags(tags = c("EU", "environment"),
                tier = 2L)

# but `tags` of a higher tier than `tier` will be ignored
c2d::infer_tags(tags = "foreign policy",
                tier = 2L)
}
