% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras_rdb.R
\name{is_date}
\alias{is_date}
\title{Testa se o valor eh data}
\usage{
is_date(i, check = TRUE, fix_na = TRUE)
}
\arguments{
\item{i}{vetor qualqer}

\item{check}{se TRUE (padão) avalia o conteúdo de i e infere se a variável total é de data}

\item{fix_na}{se tiver algum tipo de na no vetor será removido}
}
\value{
TRUE se o vetor é das classes de data
}
\description{
Permite avaliar, através do tipo ou do conteúdo, se um vetor de dados recebido é ou não no tipo data
}
