% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras_rdb.R
\name{is_int64}
\alias{is_int64}
\title{Verifica se eh int64}
\usage{
is_int64(i)
}
\arguments{
\item{i}{vetor qualqer}
}
\value{
TRUE se o vetor i eh de classe integer64
}
\description{
Verifica se eh int64
}
