% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras_rdb.R
\name{is_num}
\alias{is_num}
\title{Testa se o valor eh numerico}
\usage{
is_num(i, check = TRUE, fix_na = FALSE)
}
\arguments{
\item{i}{vetor qualqer numérico}

\item{check}{se TRUE (padão) avalia o conteúdo de i e infere se a variável é de data}

\item{fix_na}{se tiver algum tipo de NA no vetor será removido}
}
\value{
TRUE se o vetor i é das classes numéricas
}
\description{
Permite avaliar, através do tipo ou do conteúdo, se um vetor de dados recebido é ou não no tipo numérico
}
