% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{is_online}
\alias{is_online}
\title{Test RDB API availability}
\usage{
is_online(
  use_testing_server = pal::pkg_config_val(key = "use_testing_server", pkg = this_pkg),
  quiet = FALSE
)
}
\arguments{
\item{use_testing_server}{Whether or not to use the testing servers instead of the production servers for RDB Services API calls etc.}

\item{quiet}{Whether or not to suppress printing a warning in case the API is unavailable.}
}
\value{
A logical scalar.
}
\description{
Checks if the RDB API server is online and operational.
}
\concept{api_status}
