% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d.gen.R
\name{list_sudd_rfrnds}
\alias{list_sudd_rfrnds}
\title{List referendum data from \href{https://sudd.ch/}{sudd.ch}}
\usage{
list_sudd_rfrnds(
  mode = c("by_date", "by_mod_date", "filter", "random"),
  order = c("ascending", "descending"),
  filter = list(territory_name_de = NULL, title_de = NULL, year_min = NULL, year_max =
    NULL),
  use_cache = TRUE,
  cache_lifespan = "1 week"
)
}
\arguments{
\item{mode}{The listing mode. One of
\itemize{
\item \code{"by_date"}: Lists \href{https://sudd.ch/list.php?mode=allrefs}{all referendums in the sudd.ch database \strong{by \code{date}}}, together with their \code{id_sudd},
\code{country_code}, \code{territory_name_de} and \code{title_de}. Specifying the sorting order of the results via the \code{order} parameter is supported.
\item \code{"by_mod_date"}: Lists \href{https://sudd.ch/list.php?mode=moddate}{all referendums in the sudd.ch database \strong{by \code{date_last_edited}}}, together with their
\code{id_sudd}, \code{country_code}, \code{territory_name_de}, \code{date} and \code{title_de}.
\item \code{"filter"}: Allows to provide additional arguments (via the \code{filter} parameter) that limit the results accordingly. Specifying the sorting order of the
results via the \code{order} parameter is supported.
\item \code{"random"}: Lists the \code{id_sudd} of five randomly selected referendums, together with their \code{country_code}.
}}

\item{order}{The sorting order of the results. Only relevant if \code{mode} is either \code{"by_date"} or \code{"filter"}.}

\item{filter}{A list with valid filtering arguments. Only relevant if \code{mode = "filter"}. The supported filtering arguments include
\itemize{
\item \code{territory_name_de}: A string that must be (partially) matched by the referendums' \code{territory_name_de}. Matching is case-insensitive and no \href{https://en.wikipedia.org/wiki/Approximate_string_matching}{fuzzy matching} is supported.
\item \code{title_de}: A string that must be (partially) matched by the referendums' \code{title_de}. Matching is case-insensitive and no \href{https://en.wikipedia.org/wiki/Approximate_string_matching}{fuzzy matching} is supported.
\item \code{year_min}: The lower year limit of the referendums' \code{date}. A positive integer.
\item \code{year_max}: The upper year limit of the referendums' \code{date}. A positive integer.
}}

\item{use_cache}{Return cached results if possible. If \code{FALSE}, results are always newly fetched regardless of \code{cache_lifespan}.}

\item{cache_lifespan}{The duration after which cached results are refreshed (i.e. newly fetched). A valid \link[lubridate:as.duration]{lubridate duration}. Only relevant if \code{use_cache = TRUE}.}
}
\value{
A \link[tibble:tbl_df-class]{tibble} containing at least an \code{id_sudd} column.
}
\description{
Lists the referendum data from \href{https://sudd.ch/}{sudd.ch} in various ways its \href{https://sudd.ch/list.php}{\code{list.php}} endpoint allows. The output of this
function can be directly fed to \code{\link[=sudd_rfrnds]{sudd_rfrnds()}}.
}
\section{About \href{https://sudd.ch/}{sudd.ch}}{
\strong{sudd} stands for \emph{\strong{Su}chmaschine für \strong{d}irekte \strong{D}emokratie} (German) and is operated by \href{mailto:beat@sudd.ch}{Beat Müller}. Its database content is
licensed under \href{https://creativecommons.org/licenses/by-nc-sa/4.0/}{Creative Commons Attribution-NonCommercial-ShareAlike 4.0 International (CC BY-NC-SA 4.0)}.
}

\examples{
# list all referendums by modification date
c2d::list_sudd_rfrnds(mode = "by_mod_date")

# list all referendums whose title matches "AHV"
c2d::list_sudd_rfrnds(mode = "filter",
                           filter = list(title_de = "AHV"))

# get sudd.ch referendum data from all referendums from 2020 onwards
c2d::list_sudd_rfrnds(mode = "filter",
                      filter = list(year_min = 2020)) |>
  c2d::sudd_rfrnds()

# get sudd.ch referendum data from five randomly picked referendums
c2d::list_sudd_rfrnds(mode = "random") |>
  c2d::sudd_rfrnds()
}
\seealso{
Other sudd.ch scraping functions:
\code{\link{list_sudd_territories}()},
\code{\link{list_sudd_titles}()},
\code{\link{sudd_rfrnds}()}
}
\concept{sudd}
