% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d.gen.R
\name{list_sudd_territories}
\alias{list_sudd_territories}
\title{List referendum territories from \href{https://sudd.ch/}{sudd.ch}}
\usage{
list_sudd_territories()
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Lists \href{https://sudd.ch/list.php?mode=allareas}{all referendum territories from sudd.ch}, which means each \code{country_name_de} together with all the associated
\code{territory_name_de}, their search URL and their number of occurrences.
}
\details{
Note that the values in the \code{territory_name_de} column returned by this function can differ from those in the \code{territory_name_de} column of
\code{\link[=sudd_referendums]{sudd_referendums()}} and \code{\link[=list_sudd_referendums]{list_sudd_referendums()}}. The latter is often more extensive and usually includes the \code{country_name_de} (in parentheses) for
subnational referendums.
}
\section{About \href{https://sudd.ch/}{sudd.ch}}{
\strong{sudd} stands for \emph{\strong{Su}chmaschine für \strong{d}irekte \strong{D}emokratie} (German) and is operated by \href{mailto:beat@sudd.ch}{Beat Müller}. Its database content is
licensed under \href{https://creativecommons.org/licenses/by-nc-sa/4.0/}{Creative Commons Attribution-NonCommercial-ShareAlike 4.0 International (CC BY-NC-SA 4.0)}.
}

\seealso{
Other sudd.ch scraping functions:
\code{\link{list_sudd_referendums}()},
\code{\link{list_sudd_titles}()},
\code{\link{sudd_referendums}()}
}
\concept{sudd}
