% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d-GEN.R
\name{list_sudd_titles}
\alias{list_sudd_titles}
\title{List referendum titles from \href{https://sudd.ch/}{sudd.ch}}
\usage{
list_sudd_titles()
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Lists \href{https://sudd.ch/list.php?mode=alltopics}{all referendum titles from sudd.ch}, together with their search URLs and number of occurrences.
_referendums()].
}
\section{About \href{https://sudd.ch/}{sudd.ch}}{
\strong{sudd} stands for \emph{\strong{Su}chmaschine für \strong{d}irekte \strong{D}emokratie} (German) and is operated by \href{mailto:beat@sudd.ch}{Beat Müller}. Its database content is
licensed under \href{https://creativecommons.org/licenses/by-nc-sa/4.0/}{Creative Commons Attribution-NonCommercial-ShareAlike 4.0 International (CC BY-NC-SA 4.0)}.
}

\examples{
c2d::list_sudd_titles() |> dplyr::filter(n > 1)
}
\seealso{
Other sudd.ch scraping functions:
\code{\link{list_sudd_referendums}()},
\code{\link{list_sudd_territories}()},
\code{\link{sudd_referendums}()}
}
\concept{sudd}
