% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_dataset.R
\name{load_acs_dataset}
\alias{load_acs_dataset}
\title{Load the ACS dataset from the database or a file}
\usage{
load_acs_dataset(config_path = "config.yaml")
}
\arguments{
\item{config_path}{The path to the YAML configuration file}
}
\value{
A tibble of the dataset
}
\description{
Loading the full dataset from the database takes a long time
(about 20 minutes), so this function is a wrapper around
make_acs_dataset() which saves the result to a file and reads
from the file if specified in the config file.
}
\details{
If the dataset is loaded from the database, then it is stored
in the location specified in the file block of the config file.
}
