% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras_rdb.R
\name{mon_read_bf_rd}
\alias{mon_read_bf_rd}
\title{Ler dados do pacote readr}
\usage{
mon_read_bf_rd(file, control = mon_read_control(type = "readr"))
}
\arguments{
\item{file}{Nome do arquivo a ser importado.}

\item{control}{Controle com todos argumentos necessários para importar dados para uso inteno com a mon_read()}
}
\description{
Ler dados do pacote readr
}
