% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras_rdb.R
\name{mon_read_bf_rio}
\alias{mon_read_bf_rio}
\title{Ler dados do pacote rio}
\usage{
mon_read_bf_rio(file, ...)
}
\arguments{
\item{file}{Nome do arquivo a ser importado}

\item{...}{Argumentos}
}
\description{
Ler dados do pacote rio
}
