% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras_rdb.R
\name{mon_read_bf_vr}
\alias{mon_read_bf_vr}
\title{Ler dados do pacote vroom}
\usage{
mon_read_bf_vr(file, control = mon_read_control(type = "vroom"))
}
\arguments{
\item{file}{Nome do arquivo a ser importado.}

\item{control}{Controle com todos argumentos necessários para importar dados para uso inteno com a mon_read()}
}
\description{
Ler dados do pacote vroom
}
