% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras_rdb.R
\name{mon_rm_accent}
\alias{mon_rm_accent}
\title{Remove acentos de caracteres.}
\usage{
mon_rm_accent(str)
}
\arguments{
\item{str}{vetor de characters.}
}
\value{
Vetor de characters sem acentos e sem cedilhas.
}
\description{
Remove acentos (agudos, circunflexos, tremas e crases) e cedilhas de caracteres e strings.
}
\examples{
\dontrun{
palavras <- c("\u00e1gua", "op\u00e7\u00e3o", "\u00ff\u00c0\u00ea")
mon_rm_accent(palavras)
}
}
