% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras_rdb.R
\name{mon_rm_na}
\alias{mon_rm_na}
\title{Remove NA.}
\usage{
mon_rm_na(x, from = NULL, to = NULL)
}
\arguments{
\item{x}{Vetor qualquer}

\item{from}{Qual o codigo que quer remover.}

\item{to}{Qual o valor a substituir}
}
\value{
Vetor x com as devidas alteracoes
}
\description{
Remove acentos agudos, circunflexos, tremas e crases e cedilhas de caracteres e strings.
}
\details{
Quando from = NULL, padrao, busca em numerico os valores  'NA, NaN, NULL, -Inf, Inf' e
quando character busca por '','.','NA','na','N/A','n/a','NaN','nan' e '<NA>' e troca pelo valor de
to. Quando to = NULL, padrao, substitui pelo padrao do R que eh NA.
}
