% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MsRDB.R
\name{msrdb}
\alias{msrdb}
\title{Multiscale adaptive differential abundance analysis}
\usage{
msrdb(
  seqtab,
  Z,
  X = NULL,
  seqdist = NULL,
  disttype = "align",
  knn = NULL,
  alpha = 0.1,
  fdr = TRUE
)
}
\arguments{
\item{seqtab}{A numerical matrix for ASV/OTU data, each row represents a sample and each column represents a sequence.}

\item{Z}{A binary vector, 1 means treated group and 0 means control group.}

\item{X}{A numerical matrix for observed covariates, each row represents a sample and each column represents a covariates.}

\item{seqdist}{A distance matrix that measures the distance between sequences.}

\item{disttype}{Type of sequence distance. "align" means pairwise alignment distance, while "kmer" means kmer based distance.}

\item{knn}{Size of neighborhood.}

\item{alpha}{A numerical value, indicating the significance level of FWER or FDR.}

\item{fdr}{A boolean value, indicating if FDR is the measure of type I error. TRUE means FDR is used and FALSE means FWER is used.}
}
\value{
A list contains three parts; Sig is an indicator vector, where each entry correponds to each sample. TRUE means it is differential component; seqdistK is a list containing sequence clustering result; filtering is the sequence matrix after filtering.
}
\description{
\code{MsRDB} is used to conduct robust differential abundance analysis for ASV/OTU data.
}
\details{
MsRDB test embeds the sequences into a metric space and integrates a multiscale adaptive strategy for utilizing spatial structure to identify differentially abundant microbes.
}
\examples{
data(vangay)
msrdb(seqtabInterest,Z,knn = 20)

}
