% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{n_rfrnds}
\alias{n_rfrnds}
\title{Count number of referendums}
\usage{
n_rfrnds(data, by = NULL, complete_fcts = TRUE)
}
\arguments{
\item{data}{RDB referendum data as returned by \code{\link[=rfrnds]{rfrnds()}}. A data frame that at minimum contains the columns specified in \code{by} (if any).}

\item{by}{Optional \code{data} column(s) to group by before counting number of referendums. \link[dplyr:dplyr_tidy_select]{Tidy selections} are supported.}

\item{complete_fcts}{Whether or not to complete the result with implicitly missing combinations of those columns specified in \code{by} which are of type factor.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Counts the number of RDB referendums, optionally by additional columns specified via \code{by}.
}
\examples{
rdb::rfrnds(country_code = "AT",
            quiet = TRUE) |>
  rdb::n_rfrnds(by = level)

# count ballot dates instead of referendums
rdb::rfrnds(country_code = "AT",
            quiet = TRUE) |>
  rdb::as_ballot_dates() |>
  rdb::n_rfrnds(by = level)
}
\seealso{
Other referendum data transformation functions:
\code{\link{as_ballot_dates}()},
\code{\link{n_rfrnds_per_period}()},
\code{\link{printify_col_names}()},
\code{\link{unnest_var}()}
}
\concept{transform}
