% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d.gen.R
\name{n_rfrnds_per_period}
\alias{n_rfrnds_per_period}
\title{Count number of referendums per period}
\usage{
n_rfrnds_per_period(
  data,
  period = c("week", "month", "quarter", "year", "decade", "century"),
  fill_gaps = TRUE,
  by = NULL
)
}
\arguments{
\item{data}{C2D referendum data as returned by \code{\link[=rfrnds]{rfrnds()}}. A data frame that at minimum contains the column specified in \code{period} or the column \code{date} (to
compute the \link[=add_period]{period column}), plus the one(s) specified via \code{by} (if any).}

\item{period}{Type of period to count referendums by. One of \code{"week"},\code{"month"},\code{"quarter"},\code{"year"},\code{"decade"} or \code{"century"}.}

\item{fill_gaps}{Whether or not to add zero-value rows to the result for \code{period} gaps in \code{data}.}

\item{by}{Optional \code{data} column(s) to group by before counting number of referendums.
\link[dplyr:dplyr_tidy_select]{Tidy selections} are supported.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Counts the number of C2D referendums per desired period, optionally by additional columns specified via \code{by}.
}
\details{
Note that the \code{period} types \code{"year"}, \code{"decade"} and \code{"century"} are \emph{unique} timespans, while \code{"week"}, \code{"month"} and \code{"quarter"} are not, as they are
\emph{recurring} by nature (e.g. every year has a week 1).
}
\examples{
c2d::rfrnds(country_code = "AT") |> c2d::n_rfrnds_per_period()
c2d::rfrnds(country_code = "AT") |> c2d::n_rfrnds_per_period(by = level)

# without filling gaps
c2d::rfrnds(country_code = "AT") |>
  c2d::n_rfrnds_per_period(by = level,
                           fill_gaps = FALSE)

# per decade and by multiple columns
c2d::rfrnds(country_code = "AT") |>
  c2d::n_rfrnds_per_period(period = "decade",
                           by = c(level, type))
}
\seealso{
Other referendum data transformation functions:
\code{\link{prettify_col_names}()}
}
\concept{transform}
