% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{ncdb_tbl_col_id}
\alias{ncdb_tbl_col_id}
\title{Get NocoDB column ID}
\usage{
ncdb_tbl_col_id(
  tbl_id,
  col_name = NULL,
  col_title = NULL,
  auth_token = pal::pkg_config_val(key = "ncdb_token", pkg = this_pkg)
)
}
\arguments{
\item{tbl_id}{NocoDB table identifier as returned by \code{\link[=ncdb_tbl_id]{ncdb_tbl_id()}}. A character scalar.}

\item{col_name}{NocoDB column name. A character scalar.}

\item{col_title}{NocoDB column title. A character scalar.}

\item{auth_token}{NocoDB API authentication token. A character scalar.}
}
\value{
A character scalar.
}
\description{
Returns the identifier of the column with the specified \code{col_name} or \code{col_title} in the table with the specified \code{tbl_id} on the RDB NocoDB server.
}
\seealso{
Other NocoDB API functions
\code{\link{config_ncdb}()},
\code{\link{create_ncdb_user}()},
\code{\link{ncdb_api}()},
\code{\link{ncdb_base_id}()},
\code{\link{ncdb_bases}()},
\code{\link{ncdb_tbl}()},
\code{\link{ncdb_tbl_col}()},
\code{\link{ncdb_tbl_cols}()},
\code{\link{ncdb_tbl_id}()},
\code{\link{ncdb_tbls}()},
\code{\link{ncdb_users}()},
\code{\link{reorder_ncdb_tbl}()},
\code{\link{set_ncdb_display_val}()},
\code{\link{set_ncdb_display_vals}()},
\code{\link{set_ncdb_tbl_metadata}()},
\code{\link{update_ncdb_tbl}()},
\code{\link{update_ncdb_tbl_col}()},
\code{\link{upload_ncdb_attachments}()}
}
\concept{ncdb}
\keyword{internal}
