% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{nest_topics}
\alias{nest_topics}
\title{Nest topics hierarchically}
\usage{
nest_topics(data)
}
\arguments{
\item{data}{Political topics like \code{\link{data_topics}}.}
}
\value{
A \link[tibble:tbl_df-class]{tibble} with the columns \code{topic} and \code{parent_topic}.
}
\description{
Takes a data frame with the three columns \code{topic_tier_1}, \code{topic_tier_2} and \code{topic_tier_3} and transforms it into a nested data frame with the
two columns \code{topic} and \code{parent_topic}.
}
\examples{
rdb::data_topics |> rdb::nest_topics()
}
\seealso{
Other referendum topic functions:
\code{\link{data_topics}},
\code{\link{hierarchize_topics}()},
\code{\link{hierarchize_topics_fast}()},
\code{\link{infer_topics}()},
\code{\link{topics}()},
\code{\link{update_topics}()}
}
\concept{topics}
