% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{notify_postgrest}
\alias{notify_postgrest}
\title{Notify RDB PostgREST about schema changes}
\usage{
notify_postgrest(connection = connect(), disconnect = TRUE)
}
\arguments{
\item{connection}{\code{\link[RPostgres:PqConnection-class]{PqConnection}}\cr
Database connection to a PostgreSQL DBMS as returned by \code{\link[=connect]{connect()}}.}

\item{disconnect}{\code{\link[base:logical]{logical(1)}}\cr
Whether or not to \link[DBI:dbDisconnect]{terminate} the database connection when finished.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Sends a \verb{reload schema} message on the \code{pgrst} PostgreSQL \href{https://www.postgresql.org/docs/16/sql-notify.html}{notification channel}, which triggers the RDB
PostgREST server to \href{https://postgrest.org/en/latest/references/schema_cache.html#schema-cache-reloading}{reload its schema cache}.
}
\details{
Note that \code{notify_postgrest()} only works if the PostgREST server is actively \code{LISTEN}ing on the \code{pgrst} channel, which \href{https://postgrest.org/en/v12/references/listener.html#listener-on-read-replicas}{requires a connection to a read-write PostgreSQL instance}.
}
\seealso{
Other PostgREST functions
\code{\link{restart_postgrest}()}

Other administration functions:
\code{\link{backup_rdb}()},
\code{\link{clean_attachment_bucket}()},
\code{\link{config_nocodb_tbls}()},
\code{\link{connect}()},
\code{\link{create_nocodb_tbl_cols}()},
\code{\link{create_nocodb_tbls}()},
\code{\link{delete_rfrnds}()},
\code{\link{dm}()},
\code{\link{purge_nocodb}()},
\code{\link{reset_nocodb}()},
\code{\link{reset_rdb}()},
\code{\link{restart_neon_ep}()},
\code{\link{restart_postgrest}()},
\code{\link{restore_rdb}()},
\code{\link{update_administrative_units}()},
\code{\link{update_countries}()},
\code{\link{update_langs}()},
\code{\link{update_municipalities}()},
\code{\link{update_nocodb_users}()},
\code{\link{update_rfrnds}()},
\code{\link{update_subnational_entities}()},
\code{\link{update_tbl}()},
\code{\link{update_topics}()}
}
\concept{admin}
\concept{postgrest}
\keyword{internal}
