% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{pg_has_tbl}
\alias{pg_has_tbl}
\title{Test whether PostreSQL table exists}
\usage{
pg_has_tbl(tbl_name, schema = pg_schema, connection = connect())
}
\arguments{
\item{tbl_name}{Table name. A character scalar.}

\item{schema}{Schema name. A character scalar.}

\item{connection}{Database connection to \link[=connect]{RDB's PostgreSQL DBMS}. An S4 object that inherits from \code{\link[DBI:DBIConnection-class]{DBI::DBIConnection}}.}
}
\value{
A logical scalar.
}
\description{
Test whether PostreSQL table exists
}
\seealso{
Other PostgreSQL functions
\code{\link{pg_col_metadata}()},
\code{\link{pg_init_aux_tbls}()},
\code{\link{pg_init_db}()},
\code{\link{pg_init_main_tbls}()},
\code{\link{pg_pk}()},
\code{\link{pg_tbl_del}()},
\code{\link{pg_tbl_keep}()},
\code{\link{pg_tbl_update}()},
\code{\link{split_sql_str}()}
}
\concept{pg}
\keyword{internal}
