% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{pg_init_main_tbls}
\alias{pg_init_main_tbls}
\title{Initialize main RDB PostgreSQL tables}
\usage{
pg_init_main_tbls(
  connection = connect(user = "rdb_admin", password = pg_role_pw("rdb_admin"))
)
}
\arguments{
\item{connection}{Database connection to \link[=connect]{RDB's PostgreSQL DBMS}. An S4 object that inherits from \code{\link[DBI:DBIConnection-class]{DBI::DBIConnection}}.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Initialize main RDB PostgreSQL tables
}
\seealso{
Other PostgreSQL functions
\code{\link{pg_col_metadata}()},
\code{\link{pg_has_tbl}()},
\code{\link{pg_init_aux_tbls}()},
\code{\link{pg_init_db}()},
\code{\link{pg_pk}()},
\code{\link{pg_tbl_del}()},
\code{\link{pg_tbl_keep}()},
\code{\link{pg_tbl_update}()},
\code{\link{split_sql_str}()}
}
\concept{pg}
\keyword{internal}
