% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{pg_pk}
\alias{pg_pk}
\title{Get primary key column names(s) from PostgreSQL}
\usage{
pg_pk(tbl_name, schema = pg_schema, connection = connect())
}
\arguments{
\item{tbl_name}{Table name. A character scalar.}

\item{schema}{Schema name. A character scalar.}

\item{connection}{Database connection to \link[=connect]{RDB's PostgreSQL DBMS}. An S4 object that inherits from \code{\link[DBI:DBIConnection-class]{DBI::DBIConnection}}.}
}
\value{
A character scalar if \code{tbl_name} exists. Otherwise an empty character vector.
}
\description{
Returns the primary key column names(s) from RDB's DBMS, a PostgreSQL server.
}
\seealso{
Other PostgreSQL functions
\code{\link{pg_col_metadata}()},
\code{\link{pg_has_tbl}()},
\code{\link{pg_tbl_del}()},
\code{\link{pg_tbl_keep}()}
}
\concept{pg}
\keyword{internal}
