% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{pg_reset_db}
\alias{pg_reset_db}
\title{Reset RDB PostgreSQL database}
\usage{
pg_reset_db(
  connection = connect(dbname = "postgres", user = "rdb_admin", password =
    pg_role_pw("rdb_admin")),
  disconnect = TRUE
)
}
\arguments{
\item{connection}{\code{\link[RPostgres:PqConnection-class]{PqConnection}}\cr
Database connection to a PostgreSQL DBMS as returned by \code{\link[=connect]{connect()}}.}

\item{disconnect}{\code{\link[base:logical]{logical(1)}}\cr
Whether or not to \link[DBI:dbDisconnect]{terminate} the database connection when finished.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Drops and recreates the RDB PostgreSQL database.
}
\details{
Remember to restart the corresponding read-only Neon.tech compute endpoint (e.g. via \code{\link[=restart_neon_ep]{restart_neon_ep()}}) after runnning \code{pg_reset_db()} to render it
accessible again.

If the \href{https://stackoverflow.com/a/71279781/7196903}{following error} occurs, try deleting the \code{rdb} database manually via \href{https://console.neon.tech/}{Neon console}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{permission denied to terminate process
DETAIL: Only roles with privileges of the role whose process is being terminated or with privileges of the "pg_signal_backend" role may terminate this (...)
}\if{html}{\out{</div>}}
}
\seealso{
Other PostgreSQL functions
\code{\link{exec_sql_file}()},
\code{\link{pg_add_default_data}()},
\code{\link{pg_col_metadata}()},
\code{\link{pg_col_nr_to_name}()},
\code{\link{pg_drop_tbls}()},
\code{\link{pg_has_rls}()},
\code{\link{pg_has_tbl}()},
\code{\link{pg_init_aux_tbls}()},
\code{\link{pg_init_db}()},
\code{\link{pg_init_main_tbls}()},
\code{\link{pg_pk}()},
\code{\link{pg_post_nocodb}()},
\code{\link{pg_role_pw}()},
\code{\link{pg_tbl_constraints}()},
\code{\link{pg_tbl_del}()},
\code{\link{pg_tbl_keep}()},
\code{\link{pg_tbl_read}()},
\code{\link{pg_tbl_update}()},
\code{\link{pg_update_seq}()},
\code{\link{read_sql_file}()},
\code{\link{split_sql_str}()}
}
\concept{pg}
\keyword{internal}
