% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{pg_role_pw}
\alias{pg_role_pw}
\title{Get PostgreSQL role password}
\usage{
pg_role_pw(
  role,
  path = pal::pkg_config_val(key = "pg_roles_csv_file", pkg = this_pkg)
)
}
\arguments{
\item{role}{Role name to return the password for.}
}
\value{
\code{role}'s password as a character scalar.
}
\description{
Extracts the specified \code{role}'s password from the CSV file under \code{path}.
}
\seealso{
Other PostgreSQL functions
\code{\link{exec_sql_file}()},
\code{\link{pg_col_metadata}()},
\code{\link{pg_drop_tbls}()},
\code{\link{pg_has_tbl}()},
\code{\link{pg_init_aux_tbls}()},
\code{\link{pg_init_db}()},
\code{\link{pg_init_main_tbls}()},
\code{\link{pg_pk}()},
\code{\link{pg_reset_db}()},
\code{\link{pg_tbl_constraints}()},
\code{\link{pg_tbl_del}()},
\code{\link{pg_tbl_keep}()},
\code{\link{pg_tbl_update}()},
\code{\link{read_sql_file}()},
\code{\link{split_sql_str}()}
}
\concept{pg}
\keyword{internal}
