% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{pg_tbl_del}
\alias{pg_tbl_del}
\title{Delete specified data from PostgreSQL table}
\usage{
pg_tbl_del(tbl, data, pk_col_name)
}
\arguments{
\item{tbl}{Table from which data is to be deleted. A data frame, \link[tibble:tbl_df-class]{tibble} or tibble extension like \code{\link[dbplyr:tbl_lazy]{tbl_lazy}}.}

\item{data}{Table data to delete. A data frame that must contain the table's primary key column. Additional columns are ignored.}

\item{pk_col_name}{Primary key column name. A character scalar.}
}
\value{
The updated table data as a \link[tibble:tbl_df-class]{tibble}, invisibly.
}
\description{
Delete specified data from PostgreSQL table
}
\seealso{
Other PostgreSQL functions
\code{\link{pg_col_metadata}()},
\code{\link{pg_has_tbl}()},
\code{\link{pg_pk}()},
\code{\link{pg_tbl_keep}()}
}
\concept{pg}
\keyword{internal}
