% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{pg_tbl_read}
\alias{pg_tbl_read}
\title{Read PostgreSQL table}
\usage{
pg_tbl_read(tbl_name, connection = connect(), disconnect = TRUE)
}
\arguments{
\item{tbl_name}{Table name, passed on to \code{\link[DBI:dbQuoteIdentifier]{DBI::dbQuoteIdentifier()}}. Options are:
\itemize{
\item A character string with the unquoted DBMS table name, e.g. \code{"table_name"}.
\item A call to \code{\link[DBI:Id]{DBI::Id()}} with components to the fully qualified table name, e.g. \code{DBI::Id(schema = "my_schema", table = "table_name")}.
\item A call to \code{\link[DBI:SQL]{DBI::SQL()}} with the quoted and fully qualified table name given verbatim, e.g. \code{DBI::SQL('"my_schema"."table_name"')}.
}}

\item{connection}{\code{\link[RPostgres:PqConnection-class]{PqConnection}}\cr
Database connection to a PostgreSQL DBMS as returned by \code{\link[=connect]{connect()}}.}

\item{disconnect}{\code{\link[base:logical]{logical(1)}}\cr
Whether or not to \link[DBI:dbDisconnect]{terminate} the database connection when finished.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Reads in a PostgreSQL table and returns it as a \link[tibble:tbl_df-class]{tibble}.
}
\seealso{
Other PostgreSQL functions
\code{\link{exec_sql_file}()},
\code{\link{pg_add_default_data}()},
\code{\link{pg_col_metadata}()},
\code{\link{pg_col_nr_to_name}()},
\code{\link{pg_drop_tbls}()},
\code{\link{pg_has_rls}()},
\code{\link{pg_has_tbl}()},
\code{\link{pg_init_aux_tbls}()},
\code{\link{pg_init_db}()},
\code{\link{pg_init_main_tbls}()},
\code{\link{pg_pk}()},
\code{\link{pg_post_nocodb}()},
\code{\link{pg_reset_db}()},
\code{\link{pg_role_pw}()},
\code{\link{pg_tbl_constraints}()},
\code{\link{pg_tbl_del}()},
\code{\link{pg_tbl_keep}()},
\code{\link{pg_tbl_update}()},
\code{\link{pg_tbls}()},
\code{\link{pg_update_seq}()},
\code{\link{read_sql_file}()},
\code{\link{split_sql_str}()}
}
\concept{pg}
\keyword{internal}
