% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{pg_tbl_update}
\alias{pg_tbl_update}
\title{Update PostgreSQL table}
\usage{
pg_tbl_update(tbl, data, sweep = FALSE, tbl_name, connection = connect())
}
\arguments{
\item{tbl}{Table to be updated. A data frame, \link[tibble:tbl_df-class]{tibble} or tibble extension like \code{\link[dbplyr:tbl_lazy]{tbl_lazy}}.}

\item{data}{New or updated table data. A data frame that must contain at least the table's primary key column plus any additional columns with new values to
update the corresponding database fields with.}

\item{sweep}{Whether or not to also sweep the table, i.e. delete rows that are not contained in \code{data}.}

\item{tbl_name}{Table name. One of \code{"actors"}, \code{"options"}, \code{"legal_bases"}, \code{"referendums"}, \code{"referendum_titles"}, \code{"referendum_questions"}, \code{"referendum_positions"}, \code{"referendum_votes"}, \code{"referendum_sub_votes"}, \code{"countries"}, \code{"subnational_entities"}, \code{"municipalities"} and \code{"languages"}.}

\item{connection}{Database connection to \link[=connect]{RDB's PostgreSQL DBMS}. An S4 object that inherits from \code{\link[DBI:DBIConnection-class]{DBI::DBIConnection}}.}
}
\value{
The updated table data as a \link[tibble:tbl_df-class]{tibble}, invisibly.
}
\description{
Update PostgreSQL table
}
\keyword{internal}
