% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{pg_update_seq}
\alias{pg_update_seq}
\title{Update}
\usage{
pg_update_seq(
  tbl_name,
  col_name = pg_pk(tbl_name = tbl_name, schema = schema, connection = connection),
  schema = pg_schema,
  connection = connect()
)
}
\arguments{
\item{tbl_name}{Table name. A character scalar.}

\item{col_name}{Name of the column whose sequence is to be updated.}

\item{schema}{Schema name. A character scalar.}

\item{connection}{Database connection to a PostgreSQL DBMS as returned by \code{\link[=connect]{connect()}}. An S4 object that inherits from \code{\link[DBI:DBIConnection-class]{DBI::DBIConnection}}.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Update
}
\seealso{
Other PostgreSQL functions
\code{\link{exec_sql_file}()},
\code{\link{pg_add_default_data}()},
\code{\link{pg_col_metadata}()},
\code{\link{pg_col_nr_to_name}()},
\code{\link{pg_drop_tbls}()},
\code{\link{pg_has_rls}()},
\code{\link{pg_has_tbl}()},
\code{\link{pg_init_aux_tbls}()},
\code{\link{pg_init_db}()},
\code{\link{pg_init_main_tbls}()},
\code{\link{pg_pk}()},
\code{\link{pg_post_nocodb}()},
\code{\link{pg_reset_db}()},
\code{\link{pg_role_pw}()},
\code{\link{pg_tbl_constraints}()},
\code{\link{pg_tbl_del}()},
\code{\link{pg_tbl_keep}()},
\code{\link{pg_tbl_update}()},
\code{\link{read_sql_file}()},
\code{\link{split_sql_str}()}
}
\concept{pg}
\keyword{internal}
