% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{pk}
\alias{pk}
\title{Get primary key column names(s)}
\usage{
pk(tbl_name = tbl_metadata$name)
}
\arguments{
\item{tbl_name}{Table name. A character scalar.}
}
\value{
A character scalar.
}
\description{
Returns the primary key column names(s) of the specified table, based on PostgreSQL metadata obtained at package-build time. This function is faster
than \code{\link[=pg_pk]{pg_pk()}} since it doesn't perform network requests. But it can return outdated information if changes were made to the RDB PostgreSQL schema \emph{after}
the package was last built.
}
\keyword{internal}
