% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{plot_rfrnd_share_per_period}
\alias{plot_rfrnd_share_per_period}
\title{Referendum share per period stacked area chart}
\usage{
plot_rfrnd_share_per_period(
  data,
  by,
  period = c("week", "month", "quarter", "year", "decade", "century")
)
}
\arguments{
\item{data}{RDB referendum data as returned by \code{\link[=rfrnds]{rfrnds()}}. A data frame that at minimum contains the column specified in \code{period} or the column \code{date} (to
compute the \link[=add_period]{period column}), plus the column specified in \code{by}.}

\item{by}{\code{data} column to group by before counting number of referendums. \link[dplyr:dplyr_tidy_select]{Tidy selections} are supported.}

\item{period}{Type of period to count referendums by. One of \code{"week"},\code{"month"},\code{"quarter"},\code{"year"},\code{"decade"} or \code{"century"}.}
}
\value{
A \link[plotly:plot_ly]{plotly object}.
}
\description{
Creates a \href{https://plotly.com/r/filled-area-plots/#stacked-area-chart-with-cumulative-values}{Plotly stacked area chart} that visualizes the share of
referendums per period, grouped by another column.
}
\details{
Note that the \code{period} types \code{"year"}, \code{"decade"} and \code{"century"} are \emph{unique} timespans, while \code{"week"}, \code{"month"} and \code{"quarter"} are \emph{recurring} timespans (e.g. every year has a week 1).
}
\examples{
rdb::rfrnds(quiet = TRUE,
            max_cache_age = "1 year") |>
  rdb::plot_rfrnd_share_per_period(by = "level")
}
\seealso{
Other visualization functions:
\code{\link{ggplot_streamgraph}()},
\code{\link{plot_topic_segmentation}()},
\code{\link{plot_topic_share_per_period}()}
}
\concept{visualize}
