% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d.gen.R
\name{plot_tag_segmentation}
\alias{plot_tag_segmentation}
\title{Tag segmentation sunburst chart}
\usage{
plot_tag_segmentation(
  data,
  method = c("per_rfrnd", "per_tag_lineage", "naive")
)
}
\arguments{
\item{data}{C2D referendum data as returned by \code{\link[=rfrnds]{rfrnds()}}. A data frame that at minimum contains the columns \code{tags_tier_1}, \code{tags_tier_2} and
\code{tags_tier_3}.}

\item{method}{Applied method to count the number of tag occurences. One of
\itemize{
\item \strong{\code{"per_rfrnd"}}: All \emph{referendums} have the same weight. For a referendum with n different tags of the same tier, every tag is counted 1/n.
\item \strong{\code{"per_tag_lineage"}}: All \emph{tag lineages} have the same weight. For a referendum with n different tags of the same tier, every tag is fully counted,
meaning that e.g. a referendum with three different tier-3 tags has a tripled impact on the result compared to a referendum that only has a single one.
Noticeably faster than \code{"per_rfrnd"}.
\item \strong{\code{"naive"}}: Naive procedure which doesn't properly reflect tag proportions on tier 2 and 3. Based on the (wrong) assumptions that a) all referendums
have the same number of tag lineages assigned and b) tags are not deduplicated per tier. By far the fastest method, though.
}}
}
\value{
A \link[plotly:plot_ly]{plotly object}.
}
\description{
Creates a \href{https://plotly.com/r/sunburst-charts/}{Plotly sunburst chart} that visualizes the hierarchical segmentation of referendum tag occurences.
}
\details{
A \emph{tag lineage} is the hierarchical compound of a \code{tag_tier_1} and optionally a grandchild \code{tag_tier_3} and/or a child \code{tag_tier_2}.

Note that tags can be assigned on any tier to referendums (i.e. in one case, a \code{tag_tier_1} plus a child \code{tag_tier_2} is assigned, and in another case only a
\code{tag_tier_1} without any further child tag).

Furthermore, it should be noted that not every \code{tag_tier_2} has potential child \code{tag_tier_3}s. See the \href{https://rpkg.dev/c2d/articles/codebook.html#tags}{full tag hierarchy} for details.
}
\examples{
# count each referendum equally
c2d::rfrnds(country_code = "AT") |> c2d::plot_tag_segmentation(method = "per_rfrnd")

# count each tag lineage equally
c2d::rfrnds(country_code = "AT") |> c2d::plot_tag_segmentation(method = "per_tag_lineage")

# naive count (way faster, but with misleading proportions on tier 2 and 3)
c2d::rfrnds(country_code = "AT") |> c2d::plot_tag_segmentation(method = "naive")
}
\seealso{
Other visualization functions:
\code{\link{plot_tag_share_per_period}()}
}
\concept{visualize}
