% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d.gen.R
\name{plot_tag_share_per_period}
\alias{plot_tag_share_per_period}
\title{Tag share per period stacked area chart}
\usage{
plot_tag_share_per_period(
  data,
  tier = 1L,
  period = c("week", "month", "quarter", "year", "decade", "century"),
  weight_by_n_rfrnds = TRUE
)
}
\arguments{
\item{data}{C2D referendum data as returned by \code{\link[=rfrnds]{rfrnds()}}. A data frame that at minimum contains the column \code{tags_tier_#} of the specified \code{tier}.}

\item{tier}{Tier of the tags variable to plot. \code{1L}, \code{2L} or \code{3L}.}

\item{period}{Type of period to count tags by. One of \code{"week"},\code{"month"},\code{"quarter"},\code{"year"},\code{"decade"} or \code{"century"}.}

\item{weight_by_n_rfrnds}{Whether or not to weight tag occurences by number of referendums. If \code{TRUE}, for a referendum with n different tags of the same
\code{tier}, every tag is counted 1/n.}
}
\value{
A \link[plotly:plot_ly]{plotly object}.
}
\description{
Creates a \href{https://plotly.com/r/filled-area-plots/#stacked-area-chart-with-cumulative-values}{Plotly stacked area chart} that visualizes the share of
referendum tag occurences per period.
}
\details{
Note that the \code{period} types \code{"year"}, \code{"decade"} and \code{"century"} are \emph{unique} timespans, while \code{"week"}, \code{"month"} and \code{"quarter"} are \emph{recurring} timespans (e.g. every year has a week 1).
}
\examples{
c2d::rfrnds(country_code = "AT") |> c2d::plot_tag_share_per_period()
}
\seealso{
Other visualization functions:
\code{\link{plot_rfrnd_share_per_period}()},
\code{\link{plot_tag_segmentation}()}
}
\concept{visualize}
