% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d.gen.R
\name{plot_topic_segmentation}
\alias{plot_topic_segmentation}
\title{Topic segmentation sunburst chart}
\usage{
plot_topic_segmentation(
  data,
  method = c("per_rfrnd", "per_topic_lineage", "naive")
)
}
\arguments{
\item{data}{C2D referendum data as returned by \code{\link[=rfrnds]{rfrnds()}}. A data frame that at minimum contains the columns \code{topics_tier_1}, \code{topics_tier_2} and
\code{topics_tier_3}.}

\item{method}{Applied method to count the number of topic occurences. One of
\itemize{
\item \strong{\code{"per_rfrnd"}}: All \emph{referendums} have the same weight. For a referendum with n different topics of the same tier, every topic is counted 1/n.
\item \strong{\code{"per_topic_lineage"}}: All \emph{topic lineages} have the same weight. For a referendum with n different topics of the same tier, every topic is fully
counted, meaning that e.g. a referendum with three different tier-3 topics has a tripled impact on the result compared to a referendum that only has a
single one. Noticeably faster than \code{"per_rfrnd"}.
\item \strong{\code{"naive"}}: Naive procedure which doesn't properly reflect topic proportions on tier 2 and 3. Based on the (wrong) assumptions that a) all referendums
have the same number of topic lineages assigned and b) topics are not deduplicated per tier. By far the fastest method, though.
}}
}
\value{
A \link[plotly:plot_ly]{plotly object}.
}
\description{
Creates a \href{https://plotly.com/r/sunburst-charts/}{Plotly sunburst chart} that visualizes the hierarchical segmentation of referendum topic occurences.
}
\details{
A \emph{topic lineage} is the hierarchical compound of a \code{topic_tier_1} and optionally a grandchild \code{topic_tier_3} and/or a child \code{topic_tier_2}.

Note that topics can be assigned on any tier to referendums (i.e. in one case, a \code{topic_tier_1} plus a child \code{topic_tier_2} is assigned, and in another case
only a \code{topic_tier_1} without any further child topic).

Furthermore, it should be noted that not every \code{topic_tier_2} has potential child \code{topic_tier_3}s. See the \href{https://rpkg.dev/c2d/articles/codebook.html#topics}{full topic hierarchy} for details.
}
\examples{
# count each referendum equally
c2d::rfrnds(country_code = "AT") |> c2d::plot_topic_segmentation(method = "per_rfrnd")

# count each topic lineage equally
c2d::rfrnds(country_code = "AT") |> c2d::plot_topic_segmentation(method = "per_topic_lineage")

# naive count (way faster, but with misleading proportions on tier 2 and 3)
c2d::rfrnds(country_code = "AT") |> c2d::plot_topic_segmentation(method = "naive")
}
\seealso{
Other visualization functions:
\code{\link{plot_rfrnd_share_per_period}()},
\code{\link{plot_topic_share_per_period}()}
}
\concept{visualize}
