% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{prettify_col_names}
\alias{prettify_col_names}
\title{Prettify referendum data column names}
\usage{
prettify_col_names(data)
}
\arguments{
\item{data}{RDB referendum data as returned by \code{\link[=rfrnds]{rfrnds()}}.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Renames referendum data column names to be ready for publication. Useful e.g. to create tables or visualizations.
}
\details{
Note that
\itemize{
\item column names unknown to this function are not changed.
\item column \emph{labels} are \link[labelled:remove_labels]{removed} so they aren't inadvertently used instead of the column \emph{names} (i.a. relevant for \link[gt:gt]{gt}
\href{https://gt.rstudio.com/news/index.html#minor-improvements-and-bug-fixes-0-9-0}{>= 0.9.0}).
}
}
\examples{
rdb::rfrnds(quiet = TRUE) |>
  rdb::prettify_col_names()
}
\seealso{
Other referendum data transformation functions:
\code{\link{as_ballot_dates}()},
\code{\link{n_rfrnds}()},
\code{\link{n_rfrnds_per_period}()},
\code{\link{unnest_var}()}

Other functions to prettify names:
\code{\link{prettify_var_names}()}
}
\concept{prettify}
\concept{transform}
