% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{prettify_var_names}
\alias{prettify_var_names}
\title{Prettify referendum data variables}
\usage{
prettify_var_names(var_names)
}
\arguments{
\item{var_names}{Variable name(s). Those not present in \code{\link{data_codebook}} remain untouched. A character vector.}
}
\value{
A character vector of the same length as \code{var_names}.
}
\description{
Converts referendum data variable names to their ready-for-publication version. Variable names that are unknown, i.e. not present in \code{\link{data_codebook}}), are
left untouched.
}
\examples{
rdb::prettify_var_names("topics_tier_1")

# also supports unnested var names
rdb::prettify_var_names("topic_tier_1")

# unknown var names are left untouched
rdb::prettify_var_names(var_names = c("topic_tier_1", "topic_tier_99"))
}
\seealso{
\code{\link[=prettify_col_names]{prettify_col_names()}}

Other referendum metadata functions:
\code{\link{data_codebook}},
\code{\link{val_lbls}()},
\code{\link{val_scale}()},
\code{\link{var_name_unnested}()},
\code{\link{var_vals}()}
}
\concept{metadata}
