% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d.gen.R
\name{printify_col_names}
\alias{printify_col_names}
\title{Printify referendum data column names}
\usage{
printify_col_names(data)
}
\arguments{
\item{data}{C2D referendum data as returned by \code{\link[=rfrnds]{rfrnds()}}.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Renames referendum data column names to be ready to print. Useful e.g. to create publication tables.
}
\details{
Column names unknown to this function are not changed.
}
\examples{
c2d::rfrnds(country_code = "IT") |> c2d::printify_col_names()
}
\seealso{
Other referendum data transformation functions:
\code{\link{n_rfrnds_per_period}()}
}
\concept{transform}
