% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d.gen.R
\name{printify_v_names}
\alias{printify_v_names}
\title{Convert referendum data variables to their printable version}
\usage{
printify_v_names(v_names)
}
\arguments{
\item{v_names}{A character vector of C2D referendum data variable name(s). They must be either present in \code{\link{data_codebook}} or added by one of the \href{https://rpkg.dev/c2d/reference/#section-augmentation}{data augmentation functions}.}
}
\value{
A character vector of the same length as \code{v_names}.
}
\description{
Converts referendum data variables to their ready-to-print version.
}
\examples{
c2d::printify_v_names("type")
}
\seealso{
\code{\link[=printify_col_names]{printify_col_names()}}

Other referendum metadata functions:
\code{\link{data_codebook}},
\code{\link{v_vals}()},
\code{\link{val_lbls}()}
}
\concept{metadata}
