% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_dataset.R
\name{processed_acs_dataset}
\alias{processed_acs_dataset}
\title{Convert the acs dataset to the form needed by the analysis scripts}
\usage{
processed_acs_dataset(config_file = "config.yaml")
}
\arguments{
\item{config_file}{The path to the config file for the dataset}
}
\value{
A tibble of the acs dataset
}
\description{
This functions creates an ischaemia_after column by adding up the various
ischaemia-related columns. In addition, the unix timestamp columns are
converted to lubridate
}
