% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{purge_nocodb}
\alias{purge_nocodb}
\title{Purge NocoDB}
\usage{
purge_nocodb(
  fly_app_name = "rdb-nocodb",
  b2_bucket_name = fly_app_name,
  quiet = FALSE
)
}
\arguments{
\item{fly_app_name}{Fly.io app name.}

\item{b2_bucket_name}{Backblaze B2 bucket name.}

\item{quiet}{Whether or not to suppress printing status output from internal processing.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Stops the \code{fly_app_name} NocoDB instance, purges NocoDB's internal database (the Litestream replication data on the \code{b2_bucket_name/nocodb} folder) and
then restarts the NocoDB instance, thereby resetting it to factory settings. Beware that this results in \strong{loss of all existing NocoDB-specific metadata} (\href{https://docs.nocodb.com/records/expand-record/#record-comment}{comments}, \href{https://docs.nocodb.com/records/expand-record/#record-audit}{audit} \href{https://docs.nocodb.com/data-sources/actions-on-data-sources/#audit-logs}{logs} etc.).
}
\seealso{
Other NocoDB functions
\code{\link{config_nocodb_tbls}()},
\code{\link{create_nocodb_tbls}()},
\code{\link{reset_nocodb}()},
\code{\link{reset_rdb}()}

Other administration functions:
\code{\link{config_nocodb_tbls}()},
\code{\link{connect}()},
\code{\link{create_nocodb_tbls}()},
\code{\link{delete_rfrnds}()},
\code{\link{dm}()},
\code{\link{notify_postgrest}()},
\code{\link{reset_nocodb}()},
\code{\link{reset_rdb}()},
\code{\link{restart_postgrest}()},
\code{\link{update_countries}()},
\code{\link{update_langs}()},
\code{\link{update_municipalities}()},
\code{\link{update_rfrnds}()},
\code{\link{update_subnational_entities}()},
\code{\link{update_tbl}()}
}
\concept{admin}
\concept{nocodb}
\keyword{internal}
