% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb-package.R
\docType{package}
\name{rdb-package}
\alias{rdb}
\alias{rdb-package}
\title{rdb: Download Data from the Referendum Database (RDB), Which Covers Direct Democratic Votes Worldwide}
\description{
rdb provides access to the Referendum Database (RDB) from R. This database aims to record all direct democratic votes worldwide and is operated by the Centre for Democracy Studies Aarau (ZDA) at the University of Zurich, Switzerland.
}
\section{Package configuration}{


Some of rdb's functionality is controlled via package-specific global configuration which can either be set via \href{https://rdrr.io/r/base/options.html}{R options} or \href{https://en.wikipedia.org/wiki/Environment_variable}{environment variables} (the former take precedence). This configuration includes:\tabular{llll}{
   \strong{Description} \tab \strong{R option} \tab \strong{Environment variable} \tab \strong{Default value} \cr
   RDB Services API username \tab \code{rdb.api_username} \tab \code{R_RDB_API_USERNAME} \tab  \cr
   RDB Services API password \tab \code{rdb.api_password} \tab \code{R_RDB_API_PASSWORD} \tab  \cr
   Maximal timespan to preserve the package's \href{https://pkgpins.rpkg.dev/}{pkgpins} cache. Cache entries older than this will be deleted upon package loading. \tab \code{rdb.global_max_cache_age} \tab \code{R_RDB_GLOBAL_MAX_CACHE_AGE} \tab \code{"30 days"} \cr
   Whether or not to use the testing servers instead of the production servers for RDB Services API calls etc. \tab \code{rdb.use_testing_server} \tab \code{R_RDB_USE_TESTING_SERVER} \tab \code{FALSE} \cr
   Whether or not to run the tests that use the testing servers for RDB Services API calls etc. during \code{devtools::test()}. \tab \code{rdb.test_testing_server} \tab \code{R_RDB_TEST_TESTING_SERVER} \tab \code{FALSE} \cr
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://rdb.rpkg.dev}
  \item \url{https://gitlab.com/zdaarau/rpkgs/rdb}
  \item Report bugs at \url{https://gitlab.com/zdaarau/rpkgs/rdb/issues}
}

}
\author{
\strong{Maintainer}: Salim Brüggemann \email{salim-b@pm.me} (\href{https://orcid.org/0000-0002-5329-5987}{ORCID})

Authors:
\itemize{
  \item Uwe Serdült \email{uwe.serduelt@zda.uzh.ch} (\href{https://orcid.org/0000-0002-2383-3158}{ORCID})
}

Other contributors:
\itemize{
  \item  Centre for Democracy Studies Aarau (ZDA) \email{info@zdaarau.ch} [copyright holder]
}

}
\keyword{internal}
