% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.R
\name{rdb}
\alias{rdb}
\title{Robust Differential Abundance Test}
\usage{
rdb(P, Z, X = NULL, alpha = 0.1)
}
\arguments{
\item{P}{A numerical matrix for compositional data, each row represents a sample (the sume should be 1) and each column represents a component.}

\item{Z}{A binary vector, 1 means treated group and 0 means control group.}

\item{X}{A numerical matrix for observed covariates, each row represents a sample and each column represents a covariates.}

\item{alpha}{A numerical value, indicating the asymptotical level of FWER.}
}
\value{
an indicator vector, where each entry correponds to each column of P. TRUE means it is differential component
}
\description{
\code{rdb} is used to conduct robust differential abundance analysis for compositional data
}
\details{
This function returns an indicator vector, where each entry correponds to each component of compoositional data.
}
\examples{
m=50
d=100 
P=matrix(runif(m*d),nrow=m,ncol=d)
Z=rep(0,m)
Z[1:(m/2)]=1
rdb(P,Z)

}
\author{
Shulei Wang
}
