% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras_rdb.R
\name{rdb_clean_brackets}
\alias{rdb_clean_brackets}
\title{Remove colchetes, parentesis e chaves}
\usage{
rdb_clean_brackets(x)
}
\arguments{
\item{x}{string de dados}
}
\description{
Remove colchetes, parentesis e chaves
}
