% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras_rdb.R
\name{rdb_limpa_email}
\alias{rdb_limpa_email}
\title{Limpa texto de e-mail baseado em regex}
\usage{
rdb_limpa_email(email, rm_accent = TRUE, fix_na = TRUE)
}
\arguments{
\item{email}{vetor de characters.}

\item{rm_accent}{lógico indicando se deve buscar e remover acentos em email}

\item{fix_na}{se tiver algum tipo de na no vetor "email" será removido}
}
\value{
Vetor de characters detectados como e-mail
}
\description{
Aplica-se uma regex bem formatada para checar se as strings contendo dados de e-mails
estão corretas. Se houver alguma alguma string que não bate com a regex então esta é removida
e tratada como não e-mail ou e-mail inválido.
}
