% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras_rdb.R
\name{rdb_read}
\alias{rdb_read}
\title{Leitura de arquivos grandes (Big Files)}
\usage{
rdb_read(
  file,
  type = "fread",
  delim = NULL,
  dec = NULL,
  try_append = TRUE,
  sep = NULL,
  quote = "\\"",
  escape_backslash = FALSE,
  escape_double = TRUE,
  col_names = TRUE,
  col_select = NULL,
  col_types = NULL,
  encoding = NULL,
  na = c("", "NA"),
  comment = "",
  trim_ws = FALSE,
  skip = 0,
  n_max = Inf,
  guess_max = min(1000, n_max),
  altrep = TRUE,
  progress = NULL,
  skip_empty_rows = TRUE,
  ...
)
}
\arguments{
\item{file}{Vetor ou lista de strings com nomes de arquivos a serem importados da maioria das extensoes. Arquivos copactados em .gz e .bgz}

\item{type}{Nome do mecanisco de leitura. Podendo sem um em "fread","readr","vroom" ou "rio". Veja os detalhes abaixo para mais detalhes.}

\item{delim}{Delimitador de colunas, se não souber deixe NULL. Pode ser, ",", ";", "|" e outros conforme seu arquivo.}

\item{dec}{Deparador decimal.}

\item{try_append}{Se TRUE (padrão) tenta apendar (empilhar) os dados. Se FALSE retorna lista de tabelas, uma para cada arquivo importado.}

\item{sep}{Agrupador de campos para formatação de números}

\item{quote}{Regex para quotation. Serve para não confundir com aspas que delimita strings.}

\item{escape_backslash}{Utilizar contrabarra para saltar Caractéres especiais}

\item{escape_double}{Se TRUE o valor '""' representará um unico \"}

\item{col_names}{TRUE, FALSE, um vetor de nomes ou números que serao tratados como índices para selecionar colunas}

\item{col_select}{Vetor de nomes ou índices a selecionar.}

\item{col_types}{Se NULL trabalha com os padroes do pacote readr. Se data.table, utiliza os padroes da funão \code{\link[data.table]{fread}}}

\item{encoding}{Encoding do arquivo. Como padrão pega as configurações do Sistema Operacional}

\item{na}{Tipo de valor para substituir NA}

\item{comment}{Caractére a ser tratado como comentario. padrão e hashtag}

\item{trim_ws}{Limpar espacos em branco a direita e esquerda das colunas}

\item{skip}{número de linhas a saltar na leitura dos dados}

\item{n_max}{número de linhas que quer extrair. Se Inf retorna o maximo de linhas}

\item{guess_max}{Total de linhas por partes lidas (chunksize)}

\item{altrep}{Controle dos tipos de dados na leitura por \code{\link[vroom]{vroom}}}

\item{progress}{Mostrar progresso da leitura dos dados}

\item{skip_empty_rows}{Ignorar linhas vazias}

\item{...}{Outros argumentos de funções quando trabalhando com o pacote \code{\link[rio]{import}}}
}
\description{
Esta função é um ensemble que junta muitas funções otimizadas de pacotes especiais do
R. Entre eles, readr, data.table, vroom e rio.
}
\details{
Para obter detalhes completos destas funções e seus argumentos
consulte a documentacao original dos seus desenvolvedores conforme lista abaixo:
\itemize{
\item Para type = "fread" veja \code{\link[data.table]{data.table}} e \code{\link[data.table]{fread}}.
\item Para type = "readr" veja \code{\link[readr]{readr}} e \code{\link[readr]{read_delim}}.
\item Para type = "vroom" veja \code{\link[vroom]{vroom}}.
\item Para type = "rio"   veja \code{\link[rio]{rio}} e \code{\link[rio]{import}}.
}
}
\examples{
\dontrun{
library("csmon")
f2 <- list.files(pattern = "1-c000.csv.gz$", full.names = TRUE)
a <- rdb_read(file = f2, type = "fread")
b <- rdb_read(file = f2, type = "readr", delim = ",", col_names = FALSE)
d <- rdb_read(file = f2, type = "vroom", delim = ",", col_names = FALSE)
e <- rdb_read(file = "/teste_02.xls", type = "rio", sheet = 1, range = "a3:j23")
}
}
