% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras_rdb.R
\name{rdb_read_bf_dt}
\alias{rdb_read_bf_dt}
\title{Leitura de dados com pacote data.table}
\usage{
rdb_read_bf_dt(file, control = rdb_read_control(type = "fread"))
}
\arguments{
\item{file}{Nome do arquivo a ser importado.}

\item{control}{Controle com todos argumentos necessários para importar dados para uso inteno com a rdb_read()}
}
\description{
Este é apenas um wraper que deixa a função \code{\link[data.table]{fread}}
melhor formatada para passar na função \code{\link{rdb_read}}
}
