% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras_rdb.R
\name{rdb_read_bf_rio}
\alias{rdb_read_bf_rio}
\title{Leitura de dados com pacote rio}
\usage{
rdb_read_bf_rio(file, ...)
}
\arguments{
\item{file}{Nome do arquivo a ser importado}

\item{...}{Argumentos}
}
\description{
Este é apenas um wraper que deixa a função \code{\link[rio]{import}}
melhor formatada para passar na função \code{\link{rdb_read}}
}
