% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras_rdb.R
\name{rdb_read_bf_vr}
\alias{rdb_read_bf_vr}
\title{Leitura de dados com pacote vroom}
\usage{
rdb_read_bf_vr(file, control = rdb_read_control(type = "vroom"))
}
\arguments{
\item{file}{Nome do arquivo a ser importado.}

\item{control}{Controle com todos argumentos necessários para importar dados para uso inteno com a rdb_read()}
}
\description{
Este é apenas um wraper que deixa a função \code{\link[vroom]{vroom}}
melhor formatada para passar na função \code{\link{rdb_read}}
}
