% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras_rdb.R
\name{rdb_try_fix_encoding}
\alias{rdb_try_fix_encoding}
\title{Corrige UTF-8 encoding}
\usage{
rdb_try_fix_encoding(x, enc = "latin1")
}
\arguments{
\item{x}{vetor string}

\item{enc}{flag destino do encoding.}
}
\description{
Recebe um vetor string de tipos em formado UTF-8 e tentar corrigir para latin-1
desta forma, conservando letras como c cidilhado e acentos. Outros formatos de encoding de
saida são suportados.
}
\seealso{
\code{\link{Encoding}}
}
