% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{read_sql_file}
\alias{read_sql_file}
\title{Read SQL script as statements}
\usage{
read_sql_file(.path, ..., .connection = connect())
}
\arguments{
\item{.path}{Path to a file containing SQL statements.}

\item{...}{Further (named) arguments used as temporary variables available for substitution while interpolating the SQL statements via
\code{\link[DBI:sqlInterpolate]{DBI::sqlInterpolate()}}.}

\item{.connection}{Database connection as returned by \code{\link[=connect]{connect()}}. An S4 object that inherits from \code{\link[DBI:DBIConnection-class]{DBI::DBIConnection}}.}
}
\value{
A list of \code{\link[DBI:SQL]{DBI::SQL}} objects.
}
\description{
Reads in an SQL script and returns it as individual interpolated SQL statements using \code{\link[=split_sql_str]{split_sql_str()}}.
}
\seealso{
Other PostgreSQL functions
\code{\link{exec_sql_file}()},
\code{\link{pg_add_default_data}()},
\code{\link{pg_col_metadata}()},
\code{\link{pg_col_nr_to_name}()},
\code{\link{pg_drop_tbls}()},
\code{\link{pg_has_rls}()},
\code{\link{pg_has_tbl}()},
\code{\link{pg_init_aux_tbls}()},
\code{\link{pg_init_db}()},
\code{\link{pg_init_main_tbls}()},
\code{\link{pg_pk}()},
\code{\link{pg_post_nocodb}()},
\code{\link{pg_reset_db}()},
\code{\link{pg_role_pw}()},
\code{\link{pg_tbl_constraints}()},
\code{\link{pg_tbl_del}()},
\code{\link{pg_tbl_keep}()},
\code{\link{pg_tbl_update}()},
\code{\link{pg_update_seq}()},
\code{\link{split_sql_str}()}
}
\concept{pg}
\keyword{internal}
