% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d.gen.R
\name{read_toml}
\alias{read_toml}
\title{Read in a TOML file}
\usage{
read_toml(path)
}
\arguments{
\item{path}{Path to the TOML file. A \link[fs:fs_path]{path} or something coercible to.}
}
\value{
The TOML file's content as a \link[xfun:strict_list]{strict list}.
}
\description{
Read in a TOML file
}
\keyword{internal}
