% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d.gen.R
\name{referendum}
\alias{referendum}
\title{Get a single referendum's data}
\usage{
referendum(id, incl_archive = FALSE, tidy = TRUE)
}
\arguments{
\item{id}{The referendum's unique \href{https://rpkg.dev/c2d/articles/codebook.html#id}{identifier}.}

\item{incl_archive}{Whether or not to include an \code{archive} column containing data from an earlier, obsolete state of the C2D database.}

\item{tidy}{Whether or not to tidy the referendum data, i.e. apply various data cleansing tasks and add additional variables. If \code{FALSE}, the raw MongoDB
referendum data will only be modified just enough to be able to return it as a \link[tibble:tbl_df-class]{tibble}. Note that untidy data doesn't conform to the
\link[=data_codebook]{codebook} (i.a. different variable names).}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Downloads a single referendum's data from the C2D Database. See the \code{\link{codebook}} for a detailed description of all variables.
}
\examples{
c2d::referendum(id = "5bbbe26a92a21351232dd73f")
}
\seealso{
Other referendum data functions:
\code{\link{add_referendums}()},
\code{\link{count_referendums}()},
\code{\link{download_file_attachment}()},
\code{\link{edit_referendums}()},
\code{\link{referendums}()},
\code{\link{search_referendums}()},
\code{\link{validate_referendums}()}
}
\concept{referendum}
