% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d-GEN.R
\name{referendums}
\alias{referendums}
\title{Get referendum data}
\usage{
referendums(use_cache = TRUE, cache_lifespan = "1 week", incl_drafts = FALSE)
}
\arguments{
\item{use_cache}{Return cached results if possible. If \code{FALSE}, results are always newly fetched regardless of \code{cache_lifespan}.}

\item{cache_lifespan}{The duration after which cached results are refreshed (i.e. newly fetched). A valid \link[lubridate:as.duration]{lubridate duration}. Only relevant if \code{use_cache = TRUE}.}

\item{incl_drafts}{Whether or not to include referendum entries with \emph{draft} status in the returned data.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Downloads the complete data about all referendums covered by the C2D Database.
}
\details{
Currently, this function only allows to download the C2D referendums database in its entirety.
}
