% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{reorder_ncdb_tbl}
\alias{reorder_ncdb_tbl}
\title{Re-order NocoDB table}
\usage{
reorder_ncdb_tbl(
  tbl_id = ncdb_tbl_id(),
  order = 1L,
  auth_token = pal::pkg_config_val(key = "ncdb_token", pkg = this_pkg)
)
}
\arguments{
\item{tbl_id}{NocoDB table identifier as returned by \code{\link[=ncdb_tbl_id]{ncdb_tbl_id()}}. A character scalar.}

\item{auth_token}{NocoDB API authentication token. A character scalar.}
}
\value{
\code{tbl_id}, invisibly.
}
\description{
Sets the numeric order of the specified table on the RDB NocoDB server via the
\href{https://meta-apis-v2.nocodb.com/#tag/DB-Table/operation/db-table-reorder}{POST \verb{/api/v2/meta/tables/\{tbl_id\}/reorder}} API endpoint.
}
\details{
Lower numbers place the table higher up in the UI and vice versa. The current order of all the tables in a base can be determined via \code{\link[=ncdb_tbls]{ncdb_tbls()}}.
}
\seealso{
Other NocoDB API functions
\code{\link{config_ncdb}()},
\code{\link{create_ncdb_base}()},
\code{\link{create_ncdb_user}()},
\code{\link{ncdb_api}()},
\code{\link{ncdb_base}()},
\code{\link{ncdb_base_id}()},
\code{\link{ncdb_bases}()},
\code{\link{ncdb_tbl}()},
\code{\link{ncdb_tbl_col}()},
\code{\link{ncdb_tbl_col_id}()},
\code{\link{ncdb_tbl_cols}()},
\code{\link{ncdb_tbl_id}()},
\code{\link{ncdb_tbls}()},
\code{\link{ncdb_users}()},
\code{\link{set_ncdb_display_val}()},
\code{\link{set_ncdb_display_vals}()},
\code{\link{set_ncdb_tbl_metadata}()},
\code{\link{update_ncdb_base}()},
\code{\link{update_ncdb_tbl}()},
\code{\link{update_ncdb_tbl_col}()},
\code{\link{upload_ncdb_attachments}()}
}
\concept{ncdb}
\keyword{internal}
