% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{reset_rdb}
\alias{reset_rdb}
\title{Reset RDB}
\usage{
reset_rdb(
  origin_nocodb = pal::pkg_config_val("nocodb_origin"),
  hostname_pg = pal::pkg_config_val("pg_host"),
  reset_db = FALSE,
  reset_tbls = reset_db,
  keep_data = FALSE,
  ask = TRUE,
  disconnect = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{origin_nocodb}{NocoDB server \href{https://developer.mozilla.org/docs/Glossary/Origin}{origin}. A character scalar.}

\item{hostname_pg}{PostgreSQL server hostname. A character scalar.}

\item{reset_db}{Whether or not to first fully recreate the RDB PostgreSQL database using \code{\link[=pg_reset_db]{pg_reset_db()}} and \code{\link[=pg_init_db]{pg_init_db()}}. If \code{FALSE}, only the RDB
PostgreSQL tables are recreated.}

\item{reset_tbls}{Whether or not to recreate the RDB PostgreSQL tables.}

\item{keep_data}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}\cr
Whether or not to back up the contents of existing tables and restore them after the reset. Only relevant if \code{reset_db = TRUE} or \code{reset_tbls = TRUE}.}

\item{ask}{Whether or not to ask for confirmation before proceeding. Only relevant if run \link[=interactive]{interactively}.}

\item{disconnect}{\code{\link[base:logical]{logical(1)}}\cr
Whether or not to \link[DBI:dbDisconnect]{terminate} the database connection when finished.}

\item{quiet}{Whether or not to suppress printing status output from internal processing.}
}
\value{
\code{origin_nocodb}, invisibly.
}
\description{
Resets the RDB NocoDB server, optionally including the whole RDB PostgreSQL database and/or tables.
}
\examples{
\dontrun{
# reset production PGSQL DB branch and NocoDB instance
rdb::reset_rdb(reset_db = TRUE)

# reset testing PGSQL DB branch and NocoDB instance
rdb::reset_rdb(origin_nocodb = Sys.getenv("R_RDB_NOCODB_ORIGIN_TESTING"),
               hostname_pg = Sys.getenv("R_RDB_PG_HOST_TESTING"),
               reset_db = TRUE)}
}
\seealso{
Other administration functions:
\code{\link{backup_rdb}()},
\code{\link{clean_attachment_bucket}()},
\code{\link{config_nocodb_tbls}()},
\code{\link{connect}()},
\code{\link{create_nocodb_tbl_cols}()},
\code{\link{create_nocodb_tbls}()},
\code{\link{delete_rfrnds}()},
\code{\link{dm}()},
\code{\link{notify_postgrest}()},
\code{\link{purge_nocodb}()},
\code{\link{reset_nocodb}()},
\code{\link{restart_neon_ep}()},
\code{\link{restart_postgrest}()},
\code{\link{restore_rdb}()},
\code{\link{update_administrative_units}()},
\code{\link{update_countries}()},
\code{\link{update_langs}()},
\code{\link{update_municipalities}()},
\code{\link{update_nocodb_users}()},
\code{\link{update_rfrnds}()},
\code{\link{update_subnational_entities}()},
\code{\link{update_tbl}()},
\code{\link{update_topics}()}

Other NocoDB functions
\code{\link{config_nocodb_tbls}()},
\code{\link{create_nocodb_tbl_cols}()},
\code{\link{create_nocodb_tbls}()},
\code{\link{purge_nocodb}()},
\code{\link{reset_nocodb}()}
}
\concept{admin}
\concept{nocodb}
