% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{restart_neon_ep}
\alias{restart_neon_ep}
\title{Restart Neon.tech PostgreSQL endpoint}
\usage{
restart_neon_ep(
  project_id = neon_project_id,
  endpoint_id = hostname_to_ep(pal::pkg_config_val(key = "pg_host", pkg = this_pkg)),
  api_key = pal::pkg_config_val(key = "neon_api_key", pkg = this_pkg),
  max_tries = 3L
)
}
\arguments{
\item{project_id}{Neon.tech project identifier. A character scalar.}

\item{endpoint_id}{Neon.tech compute endpoint identifier. A character scalar.}

\item{api_key}{\href{https://neon.tech/docs/manage/api-keys}{API key} with sufficient access to the Neon.tech \code{project_id}.}

\item{max_tries}{Maximum number of attempts to retry in case of an HTTP error. An integerish scalar.}
}
\value{
The HTTP response body as a list.
}
\description{
Restarts the Neon.tech PostgreSQL \href{https://neon.tech/docs/reference/glossary#compute-endpoint}{compute endpoint} via the
\href{https://api-docs.neon.tech/reference/restartprojectendpoint}{\verb{/projects/\{project_id\}/endpoints/\{endpoint_id\}/restart} API endpoint}.
}
\seealso{
Other administration functions:
\code{\link{backup_rdb}()},
\code{\link{clean_attachment_bucket}()},
\code{\link{config_nocodb_tbls}()},
\code{\link{connect}()},
\code{\link{create_nocodb_tbl_cols}()},
\code{\link{create_nocodb_tbls}()},
\code{\link{delete_rfrnds}()},
\code{\link{dm}()},
\code{\link{notify_postgrest}()},
\code{\link{purge_nocodb}()},
\code{\link{reset_nocodb}()},
\code{\link{reset_rdb}()},
\code{\link{restart_postgrest}()},
\code{\link{restore_rdb}()},
\code{\link{update_administrative_units}()},
\code{\link{update_countries}()},
\code{\link{update_langs}()},
\code{\link{update_municipalities}()},
\code{\link{update_nocodb_users}()},
\code{\link{update_rfrnds}()},
\code{\link{update_subnational_entities}()},
\code{\link{update_tbl}()},
\code{\link{update_topics}()}
}
\concept{admin}
\concept{neon}
\keyword{internal}
