% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{restart_postgrest}
\alias{restart_postgrest}
\title{Restart RDB PostgREST server}
\usage{
restart_postgrest(quiet = FALSE)
}
\arguments{
\item{quiet}{Whether or not to suppress printing status output from internal processing.}
}
\value{
\code{NULL}, invisibly
}
\description{
Restarts the RDB PostgREST server VM (a \url{Fly.io} app).
}
\seealso{
Other PostgrREST functions
\code{\link{notify_postgrest}()}

Other administration functions:
\code{\link{config_nocodb_tbls}()},
\code{\link{connect}()},
\code{\link{create_nocodb_tbls}()},
\code{\link{delete_rfrnds}()},
\code{\link{dm}()},
\code{\link{notify_postgrest}()},
\code{\link{reset_nocodb}()},
\code{\link{reset_rdb}()},
\code{\link{update_countries}()},
\code{\link{update_langs}()},
\code{\link{update_municipalities}()},
\code{\link{update_rfrnds}()},
\code{\link{update_subnational_entities}()},
\code{\link{update_tbl}()}
}
\concept{admin}
\concept{postgrest}
\keyword{internal}
