% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{restore_rdb}
\alias{restore_rdb}
\title{Restore RDB}
\usage{
restore_rdb(
  dm = readRDS(file = "rdb_dm.rds"),
  sweep = FALSE,
  connection = connect(),
  disconnect = TRUE
)
}
\arguments{
\item{dm}{RDB data and relationships (primary and foreign keys) as a \code{\link[dm:dm]{dm::dm}} object.}

\item{sweep}{Whether or not to also sweep existing RDB data, i.e. delete rows that are not contained in \code{dm}'s tables.}

\item{connection}{\code{\link[RPostgres:PqConnection-class]{PqConnection}}\cr
Database connection to a PostgreSQL DBMS as returned by \code{\link[=connect]{connect()}}.}

\item{disconnect}{\code{\link[base:logical]{logical(1)}}\cr
Whether or not to \link[DBI:dbDisconnect]{terminate} the database connection when finished.}
}
\value{
\code{dm}, invisibly.
}
\description{
Restores RDB data – excluding fully programmatically administered tables – from a \code{\link[dm:dm]{dm::dm}} object that is expected to include relationship information
(primary and foreign keys).
}
\seealso{
Other administration functions:
\code{\link{backup_rdb}()},
\code{\link{clean_attachment_bucket}()},
\code{\link{config_nocodb_tbls}()},
\code{\link{connect}()},
\code{\link{create_nocodb_tbl_cols}()},
\code{\link{create_nocodb_tbls}()},
\code{\link{delete_rfrnds}()},
\code{\link{dm}()},
\code{\link{notify_postgrest}()},
\code{\link{purge_nocodb}()},
\code{\link{reset_nocodb}()},
\code{\link{reset_rdb}()},
\code{\link{restart_neon_ep}()},
\code{\link{restart_postgrest}()},
\code{\link{update_administrative_units}()},
\code{\link{update_countries}()},
\code{\link{update_langs}()},
\code{\link{update_municipalities}()},
\code{\link{update_rfrnds}()},
\code{\link{update_subnational_entities}()},
\code{\link{update_tbl}()},
\code{\link{update_topics}()}
}
\concept{admin}
