% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{rfrnd_exists}
\alias{rfrnd_exists}
\title{Test if referendum ID exists}
\usage{
rfrnd_exists(
  id,
  .use_testing_server = pal::pkg_config_val(key = "use_testing_server", pkg = this_pkg)
)
}
\arguments{
\item{id}{Referendum's unique \href{https://rdb.rpkg.dev/articles/codebook.html#id}{identifier}.}

\item{.use_testing_server}{Whether or not to use the testing servers instead of the production servers for RDB Services API calls etc.}
}
\value{
A logical scalar.
}
\description{
Tests whether the referendum with the supplied \code{id} exists or not.
}
\examples{
rdb::rfrnd_exists("6303a4cba52c3995043a8c24")
}
\seealso{
Other referendum data functions:
\code{\link{add_rfrnds}()},
\code{\link{assert_vars}()},
\code{\link{count_rfrnds}()},
\code{\link{delete_rfrnds}()},
\code{\link{download_file_attachment}()},
\code{\link{edit_rfrnds}()},
\code{\link{rfrnd}()},
\code{\link{rfrnds}()},
\code{\link{rfrnds_old}()},
\code{\link{search_rfrnds}()},
\code{\link{validate_rfrnds}()}
}
\concept{rfrnd}
