% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{rfrnd_types}
\alias{rfrnd_types}
\title{Get legal bases data}
\usage{
rfrnd_types(
  connection = connect(),
  disconnect = TRUE,
  incl_drafts = FALSE,
  use_cache = TRUE,
  max_cache_age = "1 week"
)
}
\arguments{
\item{connection}{Database connection to a PostgreSQL DBMS as returned by \code{\link[=connect]{connect()}}. An S4 object that inherits from \code{\link[DBI:DBIConnection-class]{DBI::DBIConnection}}.}

\item{disconnect}{Whether or not to \link[DBI:dbDisconnect]{terminate} the database connection when finished.}

\item{incl_drafts}{Whether or not to include database entries with \emph{draft} status.}

\item{use_cache}{Whether or not to return cached results if possible. If \code{FALSE}, results are always newly fetched regardless of \code{max_cache_age}.}

\item{max_cache_age}{Duration after which cached results are refreshed (i.e. newly fetched). A valid \link[lubridate:as.duration]{lubridate duration}. Use \code{Inf} to disable cache expiry. Only relevant if \code{use_cache = TRUE}.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Downloads the legal bases data from the Referendum Database (RDB). See the \code{\link{codebook}} for a detailed description of all variables.
}
\examples{
# get all legal bases
rdb::rfrnd_types()
}
\seealso{
Other data functions:
\code{\link{assert_vars}()},
\code{\link{delete_rfrnds}()},
\code{\link{download_file_attachments}()},
\code{\link{read_tbl}()},
\code{\link{rfrnds}()},
\code{\link{rfrnds_old}()},
\code{\link{validate_rfrnds}()}
}
\concept{data}
