% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{rfrnds_old}
\alias{rfrnds_old}
\title{Get \emph{old} referendum data}
\usage{
rfrnds_old(
  is_draft = FALSE,
  incl_archive = FALSE,
  use_cache = TRUE,
  max_cache_age = "1 week",
  quiet = FALSE
)
}
\arguments{
\item{incl_archive}{Whether or not to include an \code{archive} column containing data from an earlier, obsolete state of the Referendum Database (RDB).}

\item{use_cache}{Whether or not to return cached results if possible. If \code{FALSE}, results are always newly fetched regardless of \code{max_cache_age}.}

\item{max_cache_age}{Duration after which cached results are refreshed (i.e. newly fetched). A valid \link[lubridate:as.duration]{lubridate duration}. Use \code{Inf} to disable cache expiry. Only relevant if \code{use_cache = TRUE}.}

\item{quiet}{Whether or not to suppress printing status output from internal processing.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Downloads the latest backup of the old referendum data from the Referendum Database (RDB). See the \code{\link{codebook}} for a detailed description of all
variables.
}
\examples{
# get all referendums (excl. drafts)
rdb::rfrnds_old()
}
\seealso{
Other data functions:
\code{\link{assert_vars}()},
\code{\link{delete_rfrnds}()},
\code{\link{download_file_attachment}()},
\code{\link{read_tbl}()},
\code{\link{rfrnd_types}()},
\code{\link{rfrnds}()},
\code{\link{validate_rfrnds}()}
}
\concept{data}
\keyword{internal}
