% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d.gen.R
\name{rfrnds_per_period}
\alias{rfrnds_per_period}
\title{Count number of referendums per period}
\usage{
rfrnds_per_period(
  data,
  period = c("week", "month", "quarter", "year", "decade", "century"),
  group_by = NULL
)
}
\arguments{
\item{data}{C2D referendum data as returned by \code{\link[=rfrnds]{rfrnds()}}. A data frame that at minimum contains the column specified in \code{period} or the column \code{date} (to
compute the \link[=add_period]{period column}), plus the one specified via \code{group_by} (if any).}

\item{period}{Type of period to count referendums by. One of \code{"week"},\code{"month"},\code{"quarter"},\code{"year"},\code{"decade"} or \code{"century"}.}

\item{group_by}{Optional \code{data} column to \link[dplyr:group_by]{group by} before counting number of referendums. A \link[=is.symbol]{symbol}.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Counts the number of C2D referendums per desired period, optionally by an additional column specified via \code{group_by}.
}
\examples{
c2d::rfrnds(country_code = "AT") |> c2d::rfrnds_per_period()
c2d::rfrnds(country_code = "AT") |> c2d::rfrnds_per_period(group_by = level)
c2d::rfrnds(country_code = "AT") |> c2d::rfrnds_per_period(period = "decade", group_by = level)
}
\seealso{
Other referendum data transformation functions:
\code{\link{prettify_col_names}()}
}
\concept{transform}
