% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d.gen.R
\name{rfrnds_per_year}
\alias{rfrnds_per_year}
\title{Count number of referendums per year}
\usage{
rfrnds_per_year(data, group_by = NULL)
}
\arguments{
\item{data}{C2D referendum data as returned by \code{\link[=rfrnds]{rfrnds()}}. A data frame that at minimum contains the column \code{date} plus the one specified via \code{group_by}
(if any).}

\item{group_by}{Optional \code{data} column to \link[dplyr:group_by]{group by} before counting number of referendums. A \link[=is.symbol]{symbol}.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Counts number of C2D referendums per year, optionally by an additional column specified via \code{group_by}.
}
\examples{
c2d::rfrnds(country_code = "AT") |> c2d::rfrnds_per_year()
c2d::rfrnds(country_code = "AT") |> c2d::rfrnds_per_year(group_by = level)
}
\concept{transform}
