% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{s3_auth}
\alias{s3_auth}
\title{Authenticate S3-compatible object store}
\usage{
s3_auth(
  s3_endpoint = s3_endpoint_url,
  s3_access_key = pal::pkg_config_val(key = "nocodb_s3_access_key", pkg = this_pkg),
  s3_access_secret = pal::pkg_config_val(key = "nocodb_s3_access_secret", pkg = this_pkg)
)
}
\arguments{
\item{s3_endpoint}{S3-compatible endpoint URL where RDB's object storage bucket \code{rdb-attachments} resides.}

\item{s3_access_key}{S3-compatible access key ID with access to the \code{rdb-attachments} bucket.}

\item{s3_access_secret}{S3-compatible access secret with access to the \code{rdb-attachments} bucket.}
}
\value{
An object of class \code{\link[s3fs:S3FileSystem]{s3fs::S3FileSystem}}, invisibly.
}
\description{
Executes \code{\link[s3fs:s3_file_system]{s3fs::s3_file_system()}} to authenticate against RDB's S3-compatible object storage provider used to store file attachments.
}
\concept{s3}
\keyword{internal}
