% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d.gen.R
\name{search_referendums}
\alias{search_referendums}
\title{Search in English referendum titles}
\usage{
search_referendums(term)
}
\arguments{
\item{term}{The Search term. A character scalar.}
}
\value{
A character vector of English referendum titles matching the search \code{term}.
}
\description{
Allows to use the C2D API's primitive search functionality. The search is not case-sensitive and no \href{https://en.wikipedia.org/wiki/Approximate_string_matching}{fuzzy search} is performed (i.e. only exact matches are returned).
}
\details{
Note that this function is probably not of much use since it doesn't return any additional information about the matched referendums but only the English
titles.
}
\examples{
c2d::search_referendums("freedom")
}
\seealso{
Other referendum data functions:
\code{\link{add_referendums}()},
\code{\link{count_referendums}()},
\code{\link{download_file_attachment}()},
\code{\link{edit_referendums}()},
\code{\link{referendums}()},
\code{\link{referendum}()},
\code{\link{validate_referendums}()}
}
\concept{referendum}
