% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{search_rfrnds}
\alias{search_rfrnds}
\title{Search in English referendum titles}
\usage{
search_rfrnds(
  term,
  use_testing_server = pal::pkg_config_val(key = "use_testing_server", pkg = this_pkg)
)
}
\arguments{
\item{term}{Search term. A character scalar.}

\item{use_testing_server}{Whether or not to use the testing servers instead of the production servers for RDB Services API calls etc.}
}
\value{
A character vector of English referendum titles matching the search \code{term}.
}
\description{
Allows to use the RDB API's primitive search functionality. Note that
\itemize{
\item the search is not case-sensitive and no \href{https://en.wikipedia.org/wiki/Approximate_string_matching}{fuzzy search} is performed (i.e. only exact matches are
returned).
\item only up to the first \strong{5} matching results will be returned.
}
}
\details{
Note that this function is probably not of much use since it doesn't return any additional information about the matched referendums but only the English
titles.
}
\examples{
rdb::search_rfrnds("freedom")
}
\seealso{
Other referendum data functions:
\code{\link{add_rfrnds}()},
\code{\link{assert_vars}()},
\code{\link{count_rfrnds}()},
\code{\link{delete_rfrnds}()},
\code{\link{download_file_attachment}()},
\code{\link{edit_rfrnds}()},
\code{\link{rfrnd_exists}()},
\code{\link{rfrnds_old}()},
\code{\link{rfrnds}()},
\code{\link{rfrnd}()},
\code{\link{validate_rfrnds}()}
}
\concept{rfrnd}
