% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{set_ncdb_display_val}
\alias{set_ncdb_display_val}
\title{Set column as NocoDB display value}
\usage{
set_ncdb_display_val(
  col_id,
  auth_token = pal::pkg_config_val(key = "ncdb_token", pkg = this_pkg)
)
}
\arguments{
\item{col_id}{NocoDB column identifier as returned by \code{\link[=ncdb_tbl_col_id]{ncdb_tbl_col_id()}}. A character scalar.}

\item{auth_token}{NocoDB API authentication token. A character scalar.}
}
\value{
\code{TRUE}, invisibly.
}
\description{
Sets a column as the corresponding table's \href{https://docs.nocodb.com/fields/display-value/}{display value} on the RDB NocoDB server via the \href{https://meta-apis-v2.nocodb.com/#tag/DB-Table-Column/operation/db-table-column-primary-column-set}{POST \verb{/api/v2/meta/columns/\{col_id\}/primary}} API endpoint.
}
\seealso{
Other NocoDB API functions
\code{\link{config_ncdb}()},
\code{\link{create_ncdb_user}()},
\code{\link{ncdb_api}()},
\code{\link{ncdb_base_id}()},
\code{\link{ncdb_bases}()},
\code{\link{ncdb_tbl}()},
\code{\link{ncdb_tbl_col}()},
\code{\link{ncdb_tbl_col_id}()},
\code{\link{ncdb_tbl_cols}()},
\code{\link{ncdb_tbl_id}()},
\code{\link{ncdb_tbls}()},
\code{\link{ncdb_users}()},
\code{\link{reorder_ncdb_tbl}()},
\code{\link{set_ncdb_display_vals}()},
\code{\link{set_ncdb_tbl_metadata}()},
\code{\link{update_ncdb_tbl}()},
\code{\link{update_ncdb_tbl_col}()},
\code{\link{upload_ncdb_attachments}()}
}
\concept{ncdb}
\keyword{internal}
