% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{set_ncdb_tbl_metadata}
\alias{set_ncdb_tbl_metadata}
\title{Set metadata for all NocoDB tables}
\usage{
set_ncdb_tbl_metadata(
  base_id = ncdb_base_id(),
  auth_token = pal::pkg_config_val(key = "ncdb_token", pkg = this_pkg)
)
}
\arguments{
\item{base_id}{NocoDB base identifier as returned by \code{\link[=ncdb_base_id]{ncdb_base_id()}}. A character scalar.}

\item{auth_token}{NocoDB API authentication token. A character scalar.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Sets the proper table metadata for all tables on the RDB NocoDB server. Currently, this includes:
\itemize{
\item Setting the order of the tables in the base.
\item Setting our desired \href{https://docs.nocodb.com/tables/actions-on-table/#change-table-icon}{table icons} (emojis).
}
}
\seealso{
Other NocoDB API functions
\code{\link{config_ncdb}()},
\code{\link{create_ncdb_user}()},
\code{\link{ncdb_api}()},
\code{\link{ncdb_base_id}()},
\code{\link{ncdb_bases}()},
\code{\link{ncdb_tbl}()},
\code{\link{ncdb_tbl_col}()},
\code{\link{ncdb_tbl_col_id}()},
\code{\link{ncdb_tbl_cols}()},
\code{\link{ncdb_tbl_id}()},
\code{\link{ncdb_tbls}()},
\code{\link{ncdb_users}()},
\code{\link{reorder_ncdb_tbl}()},
\code{\link{set_ncdb_display_val}()},
\code{\link{set_ncdb_display_vals}()},
\code{\link{update_ncdb_tbl}()},
\code{\link{update_ncdb_tbl_col}()}
}
\concept{ncdb}
\keyword{internal}
