% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{split_sql_str}
\alias{split_sql_str}
\title{Split SQL string into statements}
\usage{
split_sql_str(.text, ..., .connection = connect())
}
\arguments{
\item{.text}{SQL statements to split as a character scalar.}

\item{...}{Further (named) arguments used as temporary variables available for substitution while interpolating the SQL statements via
\code{\link[DBI:sqlInterpolate]{DBI::sqlInterpolate()}}.}

\item{.connection}{Database connection as returned by \code{\link[=connect]{connect()}}. An S4 object that inherits from \code{\link[DBI:DBIConnection-class]{DBI::DBIConnection}}.}
}
\value{
A list of \code{\link[DBI:SQL]{DBI::SQL}} objects.
}
\description{
Splits a string of SQL into individual SQL statements and interpolates them using \code{\link[DBI:sqlInterpolate]{DBI::sqlInterpolate()}}. Useful to prepare the contents of an \code{.sql} script
to be able to feed it to \code{\link[DBI:dbSendStatement]{DBI::dbSendStatement()}} (which only accepts a single statement per invocation).
}
\seealso{
Other PostgreSQL functions
\code{\link{exec_sql_file}()},
\code{\link{pg_add_default_data}()},
\code{\link{pg_col_metadata}()},
\code{\link{pg_col_nr_to_name}()},
\code{\link{pg_drop_tbls}()},
\code{\link{pg_has_rls}()},
\code{\link{pg_has_tbl}()},
\code{\link{pg_init_aux_tbls}()},
\code{\link{pg_init_db}()},
\code{\link{pg_init_main_tbls}()},
\code{\link{pg_pk}()},
\code{\link{pg_post_nocodb}()},
\code{\link{pg_reset_db}()},
\code{\link{pg_role_pw}()},
\code{\link{pg_tbl_constraints}()},
\code{\link{pg_tbl_del}()},
\code{\link{pg_tbl_keep}()},
\code{\link{pg_tbl_read}()},
\code{\link{pg_tbl_update}()},
\code{\link{pg_tbls}()},
\code{\link{pg_update_seq}()},
\code{\link{read_sql_file}()}
}
\concept{pg}
\keyword{internal}
