% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d-GEN.R
\name{sudd_referendum}
\alias{sudd_referendum}
\title{Get a single referendum's data from \href{https://sudd.ch/}{sudd.ch}}
\usage{
sudd_referendum(id_sudd)
}
\arguments{
\item{id_sudd}{The referendum's identifier assigned by \href{https://sudd.ch/}{sudd.ch}.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}. The column names are aligned with those of \code{\link[=referendums]{referendums()}} as closely as possible.
}
\description{
Downloads a single referendum's data from \href{https://sudd.ch/}{sudd.ch}.
}
\examples{
c2d::referendum(id = "5bbc045192a21351232e596f")$id_sudd |> c2d::sudd_referendum()
}
\concept{sudd}
