% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d-GEN.R
\name{sudd_referendums}
\alias{sudd_referendums}
\title{Get referendum data from \href{https://sudd.ch/}{sudd.ch}}
\usage{
sudd_referendums(ids_sudd)
}
\arguments{
\item{ids_sudd}{The referendum identifiers assigned by \href{https://sudd.ch/}{sudd.ch}. Either as a character vector or a data frame containing a column
\code{id_sudd}. \code{NA}s are ignored.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}. The column names are aligned with those of \code{\link[=referendums]{referendums()}} as closely as possible.
}
\description{
Downloads referendum data from \href{https://sudd.ch/}{sudd.ch}.
}
\details{
\strong{sudd} stands for \emph{\strong{Su}chmaschine für \strong{d}irekte \strong{D}emokratie} (German) and is operated by \href{mailto:beat@sudd.ch}{Beat Müller}. Its database content is
licensed under \href{https://creativecommons.org/licenses/by-nc-sa/4.0/}{Creative Commons Attribution-NonCommercial-ShareAlike 4.0 International (CC BY-NC-SA 4.0)}.
}
\examples{
c2d::referendum(id = "5bbc045192a21351232e596f")$id_sudd |> c2d::sudd_referendums()

c2d::referendums(country_code = "AT") |> c2d::sudd_referendums()
}
\concept{sudd}
