% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{sudd_rfrnds}
\alias{sudd_rfrnds}
\title{Get referendum data from \href{https://sudd.ch/}{sudd.ch}}
\usage{
sudd_rfrnds(
  ids_sudd,
  use_cache = TRUE,
  max_cache_age = "1 week",
  quiet = FALSE
)
}
\arguments{
\item{ids_sudd}{The referendum identifiers assigned by \href{https://sudd.ch/}{sudd.ch}. Either as a character vector or a data frame containing a column
\code{id_sudd}. \code{NA}s are ignored.}

\item{use_cache}{Whether or not to return cached results if possible. If \code{FALSE}, results are always newly fetched regardless of \code{max_cache_age}.}

\item{max_cache_age}{Duration after which cached results are refreshed (i.e. newly fetched). A valid \link[lubridate:as.duration]{lubridate duration}. Use \code{Inf} to disable cache expiry. Only relevant if \code{use_cache = TRUE}.}

\item{quiet}{Whether or not to suppress printing status output from internal processing.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}. The column names are aligned with those of \code{\link[=rfrnds]{rfrnds()}} as closely as possible.
}
\description{
Downloads referendum data from \href{https://sudd.ch/}{sudd.ch}.
}
\section{About \href{https://sudd.ch/}{sudd.ch}}{
\strong{sudd} stands for \emph{\strong{Su}chmaschine für \strong{d}irekte \strong{D}emokratie} (German) and is operated by \href{mailto:beat@sudd.ch}{Beat Müller}. Its database content is
licensed under \href{https://creativecommons.org/licenses/by-nc-sa/4.0/}{Creative Commons Attribution-NonCommercial-ShareAlike 4.0 International (CC BY-NC-SA 4.0)}.
}

\examples{
rdb::rfrnd(id = "5bbc045192a21351232e596f")$id_sudd |> rdb::sudd_rfrnds()

rdb::rfrnds(quiet = TRUE) |>
  dplyr::filter(country_code == "AT") |>
  rdb::sudd_rfrnds()
}
\seealso{
Other sudd.ch scraping functions:
\code{\link{list_sudd_rfrnds}()},
\code{\link{list_sudd_territories}()},
\code{\link{list_sudd_titles}()}
}
\concept{sudd}
