% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d.gen.R
\name{tags}
\alias{tags}
\title{List available tags}
\usage{
tags(tiers = 1:3)
}
\arguments{
\item{tiers}{Tiers to include tags from. An integerish vector.}
}
\value{
A character vector.
}
\description{
Lists the set of available \href{https://rpkg.dev/c2d/articles/codebook.html#tags}{referendum tags} on the specified \code{tiers}.
}
\examples{
c2d::tags(tiers = 1:2)
}
\seealso{
Other tag-related functions:
\code{\link{hierarchize_tags}()},
\code{\link{infer_tags}()}
}
\concept{tags}
