% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{tbl_n_rfrnds}
\alias{tbl_n_rfrnds}
\title{Tabulate number of referendums}
\usage{
tbl_n_rfrnds(
  data,
  by = NULL,
  complete_fcts = TRUE,
  n_rows = Inf,
  order = NULL,
  incl_row_head = TRUE,
  incl_col_head = TRUE,
  add_total_row = NULL,
  add_total_col = NULL,
  lbl_total_row = gt::md("**Total**"),
  lbl_total_col = gt::md("**Total**")
)
}
\arguments{
\item{data}{RDB referendum data as returned by \code{\link[=rfrnds]{rfrnds()}}. A data frame that at minimum contains the columns specified in \code{by} (if any).}

\item{by}{Up to three additional \code{data} columns to group by before counting number of referendums. \link[dplyr:dplyr_tidy_select]{Tidy selections} are supported.}

\item{complete_fcts}{Whether or not to complete the result with implicitly missing combinations of those columns specified in \code{by} which are of type factor.}

\item{n_rows}{Maximum number of rows to be included in the resulting table. All the rows exceeding that limit are replaced by a single row of ellipses. An
integer scalar or \code{Inf} for an unlimited number of rows.}

\item{order}{How to order the rows of the resulting table. One of
\itemize{
\item \code{"ascending"} to sort in ascending order by the number of referendums,
\item \code{"descending"} to sort in descending order by the number of referendums, or
\item \code{NULL} to leave the sorting unchanged.
}}

\item{incl_row_head}{Whether or not to include a row heading with the \link[=printify_var_names]{printified} name of the first \code{by} variable.}

\item{incl_col_head}{Whether or not to include column headings (in the table's \link[gt:tab_stubhead]{stub}) with the \link[=printify_var_names]{printified} names of
the second and third \code{by} variables.}

\item{add_total_row}{Whether or not to add a summary row at the very end of the table containing column totals. If \code{NULL}, a total row is added only if
at least one column is provided in \code{by}.}

\item{add_total_col}{Whether or not to add a summary column at the very end of the table containing row totals. If \code{NULL}, a total column is added only if
multiple columns are provided in \code{by}.}

\item{lbl_total_row}{Label of the summary row containing column totals. Only relevant if \code{add_total_row = TRUE}. A character scalar. \code{\link[gt:md]{gt::md()}} or
\code{\link[gt:html]{gt::html()}} can be used to format the label text.}

\item{lbl_total_col}{Label of the summary column containing row totals. Only relevant if \code{add_total_col = TRUE}. A character scalar. \code{\link[gt:md]{gt::md()}} or
\code{\link[gt:html]{gt::html()}} can be used to format the label text.}
}
\value{
A \code{\link[gt:gt]{gt_tbl}} object.
}
\description{
Creates a ready-to-print \link[gt:gt]{gt} table with the number of referendums, optionally counted \code{by} up to three additional variables.
}
\details{
The first variable specified in \code{by} will be reflected in additional rows in the resulting table, i.e. expand it vertically. The second and third variables
will be reflected in additional columns, i.e. expand it horizontally.
}
\examples{
rdb::rfrnds(max_cache_age = Inf,
            quiet = TRUE) |>
  rdb::tbl_n_rfrnds()

# grouped by a single column
rdb::rfrnds(max_cache_age = Inf,
            quiet = TRUE) |>
  rdb::tbl_n_rfrnds(by = level)

# grouped by two columns
rdb::rfrnds(max_cache_age = Inf,
            quiet = TRUE) |>
  rdb::tbl_n_rfrnds(by = c(type, level))

# grouped by three columns
rdb::rfrnds(max_cache_age = Inf,
            quiet = TRUE) |>
  rdb::tbl_n_rfrnds(by = c(country_name, level, type),
                    n_rows = 10L,
                    order = "descending")

# count ballot dates instead of referendums
rdb::rfrnds(max_cache_age = Inf,
            quiet = TRUE) |>
  rdb::as_ballot_dates() |>
  rdb::tbl_n_rfrnds(by = c(country_name, level),
                    n_rows = 10L,
                    order = "descending")
}
\seealso{
Other tabulation functions:
\code{\link{tbl_n_rfrnds_per_period}()}
}
\concept{tabulate}
