% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d-GEN.R
\name{tidy_referendums}
\alias{tidy_referendums}
\title{Tidy "raw" MongoDB referendum data}
\usage{
tidy_referendums(data, tidy = TRUE)
}
\arguments{
\item{data}{The data from the C2D MongoDB as a list (converted from the JSON).}

\item{tidy}{Whether or not to tidy the referendum data, i.e. apply various data cleansing tasks and add additional variables. If \code{FALSE}, the raw MongoDB
referendum data will only be modified just enough to be able to return it as a tibble. Note that the latter doesn't conform to the \link{codebook}
(i.a. different variable names).}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Converts from the "raw" MongoDB to the tidied \code{referendums()} schema.
}
\keyword{internal}
