% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{tidy_rfrnds}
\alias{tidy_rfrnds}
\title{Tidy "raw" RDB API referendum data}
\usage{
tidy_rfrnds(data, tidy = TRUE)
}
\arguments{
\item{data}{The MongoDB data as a list (converted from the JSON returned by the RDB API using \code{\link[jsonlite:fromJSON]{jsonlite::fromJSON()}}).}

\item{tidy}{Whether or not to tidy the referendum data, i.e. apply various data cleansing tasks and add additional variables. If \code{FALSE}, the raw MongoDB
referendum data will only be modified just enough to be able to return it as a \link[tibble:tbl_df-class]{tibble}. Note that untidy data doesn't conform to the
\link[=data_codebook]{codebook} (i.a. different variable names).}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Converts the "raw" MongoDB data from the RDB API to the tidied \code{\link[=rfrnds]{rfrnds()}} schema.
}
\details{
You can reverse this function again using \code{\link[=untidy_rfrnds]{untidy_rfrnds()}}.
}
\keyword{internal}
