% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d.gen.R
\name{topics}
\alias{topics}
\title{List available topics}
\usage{
topics(tiers = 1:3)
}
\arguments{
\item{tiers}{Tiers to include topics from. An integerish vector.}
}
\value{
A character vector.
}
\description{
Lists the set of available \href{https://rpkg.dev/c2d/articles/codebook.html#topics}{referendum topics} on the specified \code{tiers}.
}
\examples{
c2d::topics(tiers = 1:2)
}
\seealso{
Other referendum topic functions:
\code{\link{data_topics}},
\code{\link{hierarchize_topics_fast}()},
\code{\link{hierarchize_topics}()},
\code{\link{infer_topics}()}
}
\concept{topics}
