% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d.gen.R
\name{untidy_referendums}
\alias{untidy_referendums}
\title{Untidy into "raw" C2D API referendum data}
\usage{
untidy_referendums(data, as_tibble = FALSE)
}
\arguments{
\item{data}{The data to untidy as returned by \code{\link[=referendums]{referendums()}}.}
}
\value{
If \code{as_tibble = FALSE}, a list with one element per referendum, suitable to convert \code{\link[jsonlite:fromJSON]{jsonlite::toJSON()}} and then feed to the C2D API.

Otherwise a \link[tibble:tbl_df-class]{tibble} of the same format as returned by \code{\link[=referendums]{referendums(tidy = FALSE)}}.
}
\description{
Converts from the tidied \code{\link[=referendums]{referendums()}} to the "raw" MongoDB schema used by the C2D API. Basically reverts \code{\link[=tidy_referendums]{tidy_referendums()}}.
}
\keyword{internal}
