% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d.gen.R
\name{untidy_rfrnds}
\alias{untidy_rfrnds}
\title{Untidy into "raw" C2D API referendum data}
\usage{
untidy_rfrnds(data, as_tibble = FALSE)
}
\arguments{
\item{data}{The data to untidy as returned by \code{\link[=rfrnds]{rfrnds()}}.}

\item{as_tibble}{Whether or not to return the result as a \link[tibble:tbl_df-class]{tibble}. If \code{FALSE}, a list is returned.}
}
\value{
If \code{as_tibble = FALSE}, a list with one element per referendum, suitable to be converted \code{\link[jsonlite:fromJSON]{jsonlite::toJSON()}} and then fed to the C2D API.

Otherwise a \link[tibble:tbl_df-class]{tibble} of the same format as returned by \code{\link[=rfrnds]{rfrnds(tidy = FALSE)}}.
}
\description{
Converts from the tidied \code{\link[=rfrnds]{rfrnds()}} to the "raw" MongoDB schema used by the C2D API. Basically reverts \code{\link[=tidy_rfrnds]{tidy_rfrnds()}}.
}
\keyword{internal}
