% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{update_administrative_units}
\alias{update_administrative_units}
\title{Update RDB \code{administrative_units} table}
\usage{
update_administrative_units(
  sweep = TRUE,
  connection = connect(),
  disconnect = TRUE
)
}
\arguments{
\item{sweep}{Whether or not to also sweep the table, i.e. delete rows that are not contained in \code{data}.}

\item{connection}{Database connection to a PostgreSQL DBMS as returned by \code{\link[=connect]{connect()}}. An S4 object that inherits from \code{\link[DBI:DBIConnection-class]{DBI::DBIConnection}}.}

\item{disconnect}{Whether or not to \link[DBI:dbDisconnect]{terminate} the database connection when finished.}
}
\value{
The newly written \code{administrative_units} table data as a \link[tibble:tbl_df-class]{tibble}, invisibly.
}
\description{
Updates the RDB's \code{administrative_units} table with data combined from the \code{subnational_entities}, \code{countries}, \code{subnational_entities} and \code{municipalities}
tables.
}
\seealso{
Other administration functions:
\code{\link{clean_attachment_bucket}()},
\code{\link{config_nocodb_tbls}()},
\code{\link{connect}()},
\code{\link{create_nocodb_tbl_cols}()},
\code{\link{create_nocodb_tbls}()},
\code{\link{delete_rfrnds}()},
\code{\link{dm}()},
\code{\link{notify_postgrest}()},
\code{\link{purge_nocodb}()},
\code{\link{reset_nocodb}()},
\code{\link{reset_rdb}()},
\code{\link{restart_neon_ep}()},
\code{\link{restart_postgrest}()},
\code{\link{update_countries}()},
\code{\link{update_langs}()},
\code{\link{update_municipalities}()},
\code{\link{update_rfrnds}()},
\code{\link{update_subnational_entities}()},
\code{\link{update_tbl}()},
\code{\link{update_topics}()}
}
\concept{admin}
