% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{update_municipalities}
\alias{update_municipalities}
\title{Generate RDB \code{municipalities} table}
\usage{
update_municipalities(sweep = TRUE, connection = connect(), disconnect = TRUE)
}
\arguments{
\item{sweep}{Whether or not to also sweep the table, i.e. delete rows that are not contained in \code{data}.}

\item{connection}{Database connection to \link[=connect]{RDB's PostgreSQL DBMS}. An S4 object that inherits from \code{\link[DBI:DBIConnection-class]{DBI::DBIConnection}}.}

\item{disconnect}{Whether or not to \link[DBI:dbDisconnect]{terminate} the database connection when finished.}
}
\value{
The updated \code{municipalities} table data as a tibble, invisibly.
}
\description{
Updates the RDB's \code{municipalities} table with the data from \code{\link{data_municipalities}}.
}
\seealso{
Other administration functions:
\code{\link{config_ncdb}()},
\code{\link{connect}()},
\code{\link{delete_rfrnds}()},
\code{\link{dm}()},
\code{\link{update_countries}()},
\code{\link{update_langs}()},
\code{\link{update_rfrnds}()},
\code{\link{update_subnational_entities}()}
}
\concept{admin}
