% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{update_ncdb_base}
\alias{update_ncdb_base}
\title{Update NocoDB base}
\usage{
update_ncdb_base(
  title = NULL,
  description = NULL,
  color = NULL,
  show_null_and_empty_in_filter = NULL,
  base_id = ncdb_base_id(),
  auth_token = pal::pkg_config_val(key = "ncdb_token", pkg = this_pkg)
)
}
\arguments{
\item{title}{Title of the new base. A character scalar.}

\item{description}{Description of the new base. A character scalar.}

\item{color}{Color to use for the base icon. A hexadecimal color code like \code{"#3f4eff"}.}

\item{show_null_and_empty_in_filter}{Whether or not to distinguish \code{NULL}s from empty values in column filters. If \code{FALSE}, \code{NULL} and empty fields are
treated alike.}

\item{base_id}{NocoDB base identifier as returned by \code{\link[=ncdb_base_id]{ncdb_base_id()}}. A character scalar.}

\item{auth_token}{NocoDB API authentication token. A character scalar.}
}
\value{
\code{base_id}, invisibly.
}
\description{
Updates the specified base on the RDB NocoDB server via the
\href{https://meta-apis-v2.nocodb.com/#tag/Base/operation/base-update}{PATCH \verb{/api/v2/meta/bases/\{base_id\}}} API endpoint.
}
\seealso{
Other NocoDB API functions
\code{\link{config_ncdb}()},
\code{\link{create_ncdb_base}()},
\code{\link{create_ncdb_user}()},
\code{\link{ncdb_api}()},
\code{\link{ncdb_base}()},
\code{\link{ncdb_base_id}()},
\code{\link{ncdb_bases}()},
\code{\link{ncdb_tbl}()},
\code{\link{ncdb_tbl_col}()},
\code{\link{ncdb_tbl_col_id}()},
\code{\link{ncdb_tbl_cols}()},
\code{\link{ncdb_tbl_id}()},
\code{\link{ncdb_tbls}()},
\code{\link{ncdb_users}()},
\code{\link{reorder_ncdb_tbl}()},
\code{\link{set_ncdb_display_val}()},
\code{\link{set_ncdb_display_vals}()},
\code{\link{set_ncdb_tbl_metadata}()},
\code{\link{update_ncdb_tbl}()},
\code{\link{update_ncdb_tbl_col}()},
\code{\link{upload_ncdb_attachments}()}
}
\concept{ncdb}
\keyword{internal}
