% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{update_ncdb_tbl}
\alias{update_ncdb_tbl}
\title{Update NocoDB table metadata}
\usage{
update_ncdb_tbl(
  tbl_id,
  body_json,
  auto_unbox = TRUE,
  auth_token = pal::pkg_config_val(key = "ncdb_token", pkg = this_pkg),
  quiet = FALSE
)
}
\arguments{
\item{tbl_id}{NocoDB table identifier as returned by \code{\link[=ncdb_tbl_id]{ncdb_tbl_id()}}. A character scalar.}

\item{auth_token}{NocoDB API authentication token. A character scalar.}

\item{quiet}{Whether or not to suppress printing status output from internal processing.}
}
\value{
\code{tbl_id}, invisibly.
}
\description{
Updates the metadata of the specified table on the RDB NocoDB server via the
\href{https://meta-apis-v2.nocodb.com/#tag/DB-Table/operation/db-table-update}{PATCH \verb{/api/v2/meta/tables/\{tbl_id\}}} API endpoint.
}
\seealso{
Other NocoDB API functions
\code{\link{config_ncdb}()},
\code{\link{create_ncdb_base}()},
\code{\link{create_ncdb_user}()},
\code{\link{ncdb_api}()},
\code{\link{ncdb_base}()},
\code{\link{ncdb_base_id}()},
\code{\link{ncdb_bases}()},
\code{\link{ncdb_tbl}()},
\code{\link{ncdb_tbl_col}()},
\code{\link{ncdb_tbl_col_id}()},
\code{\link{ncdb_tbl_cols}()},
\code{\link{ncdb_tbl_id}()},
\code{\link{ncdb_tbls}()},
\code{\link{ncdb_users}()},
\code{\link{reorder_ncdb_tbl}()},
\code{\link{set_ncdb_display_val}()},
\code{\link{set_ncdb_display_vals}()},
\code{\link{set_ncdb_tbl_metadata}()},
\code{\link{update_ncdb_base}()},
\code{\link{update_ncdb_tbl_col}()},
\code{\link{upload_ncdb_attachments}()}
}
\concept{ncdb}
\keyword{internal}
