% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{update_ncdb_tbl_col}
\alias{update_ncdb_tbl_col}
\title{Update NocoDB table column metadata}
\usage{
update_ncdb_tbl_col(
  col_id,
  body_json,
  auto_unbox = TRUE,
  auth_token = pal::pkg_config_val(key = "ncdb_token", pkg = this_pkg)
)
}
\arguments{
\item{col_id}{NocoDB column identifier as returned by \code{\link[=ncdb_tbl_col_id]{ncdb_tbl_col_id()}}. A character scalar.}

\item{auth_token}{NocoDB API authentication token. A character scalar.}
}
\value{
?
}
\description{
Updates the metadata of the specified table column on the RDB NocoDB server via the
\href{https://meta-apis-v2.nocodb.com/#tag/DB-Table-Column/operation/db-table-column-update}{PATCH \verb{/api/v2/meta/columns/\{col_id\}}} API endpoint.
}
\seealso{
Other NocoDB API functions
\code{\link{config_ncdb}()},
\code{\link{ncdb_api}()},
\code{\link{ncdb_base_id}()},
\code{\link{ncdb_bases}()},
\code{\link{ncdb_tbl}()},
\code{\link{ncdb_tbl_col}()},
\code{\link{ncdb_tbl_col_id}()},
\code{\link{ncdb_tbl_cols}()},
\code{\link{ncdb_tbl_id}()},
\code{\link{ncdb_tbls}()},
\code{\link{reorder_ncdb_tbl}()},
\code{\link{set_ncdb_display_val}()},
\code{\link{set_ncdb_display_vals}()},
\code{\link{set_ncdb_tbl_metadata}()},
\code{\link{update_ncdb_tbl}()}
}
\concept{ncdb}
\keyword{internal}
