% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{update_nocodb_users}
\alias{update_nocodb_users}
\title{Update RDB \code{nocodb_users} table}
\usage{
update_nocodb_users(
  hostname = pal::pkg_config_val(key = "nocodb_host", pkg = this_pkg),
  sweep = FALSE,
  connection = connect(),
  disconnect = TRUE
)
}
\arguments{
\item{hostname}{NocoDB server \href{https://en.wikipedia.org/wiki/Hostname}{hostname}. A character scalar.}

\item{sweep}{Whether or not to also sweep the table, i.e. delete rows that are not contained in \code{data}.}

\item{connection}{\code{\link[RPostgres:PqConnection-class]{PqConnection}}\cr
Database connection to a PostgreSQL DBMS as returned by \code{\link[=connect]{connect()}}.}

\item{disconnect}{\code{\link[base:logical]{logical(1)}}\cr
Whether or not to \link[DBI:dbDisconnect]{terminate} the database connection when finished.}
}
\value{
The newly written \code{nocodb_users} table data as a \link[tibble:tbl_df-class]{tibble}, invisibly.
}
\description{
Updates the RDB's \code{nocodb_users} table with the current \link[nocodb:users]{user data} from the specified NocoDB server.
}
\seealso{
Other administration functions:
\code{\link{backup_rdb}()},
\code{\link{clean_attachment_bucket}()},
\code{\link{config_nocodb_tbls}()},
\code{\link{connect}()},
\code{\link{create_nocodb_tbl_cols}()},
\code{\link{create_nocodb_tbls}()},
\code{\link{delete_rfrnds}()},
\code{\link{dm}()},
\code{\link{notify_postgrest}()},
\code{\link{purge_nocodb}()},
\code{\link{reset_nocodb}()},
\code{\link{reset_rdb}()},
\code{\link{restart_neon_ep}()},
\code{\link{restart_postgrest}()},
\code{\link{restore_rdb}()},
\code{\link{update_administrative_units}()},
\code{\link{update_countries}()},
\code{\link{update_langs}()},
\code{\link{update_municipalities}()},
\code{\link{update_rfrnds}()},
\code{\link{update_subnational_entities}()},
\code{\link{update_tbl}()},
\code{\link{update_topics}()}
}
\concept{admin}
