% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{update_tbl}
\alias{update_tbl}
\title{Update RDB table}
\usage{
update_tbl(
  data,
  tbl_name = tbl_metadata$name,
  sweep = FALSE,
  override_pk_sys_val = FALSE,
  connection = connect(),
  disconnect = TRUE
)
}
\arguments{
\item{data}{New or updated table data. A data frame that must contain at least the table's primary key column plus any additional columns with new values to
update the corresponding database fields with.}

\item{tbl_name}{\code{\link[base:character]{character(1)}}\cr
Table name. One of
\itemize{
\item \code{"actors"}
\item \code{"options"}
\item \code{"legal_instruments"}
\item \code{"legal_norms"}
\item \code{"referendum_types"}
\item \code{"referendum_clusters"}
\item \code{"referendums"}
\item \code{"referendum_types_legal_norms"}
\item \code{"referendum_types_referendums"}
\item \code{"topics_referendums"}
\item \code{"referendum_titles"}
\item \code{"referendum_questions"}
\item \code{"referendum_urls"}
\item \code{"referendum_positions"}
\item \code{"referendum_votes"}
\item \code{"referendum_sub_votes"}
\item \code{"administrative_units"}
\item \code{"supranational_entities"}
\item \code{"countries"}
\item \code{"subnational_entities"}
\item \code{"municipalities"}
\item \code{"languages"}
\item \code{"topics"}
}}

\item{sweep}{Whether or not to also sweep the table, i.e. delete rows that are not contained in \code{data}.}

\item{override_pk_sys_val}{Whether or not to add new rows that have primary key values set (i.e. which are not \code{NA}) to a \code{tbl} whose primary key is
\verb{GENERATED ALWAYS}. Setting \code{override_pk_sys_val = TRUE} is roughly equivalent to \href{https://www.postgresql.org/docs/current/sql-insert.html#id-1.9.3.152.6.2}{specifying \verb{OVERRIDING SYSTEM VALUE} in an SQL \code{INSERT} statement}.}

\item{connection}{\code{\link[RPostgres:PqConnection-class]{PqConnection}}\cr
Database connection to a PostgreSQL DBMS as returned by \code{\link[=connect]{connect()}}.}

\item{disconnect}{\code{\link[base:logical]{logical(1)}}\cr
Whether or not to \link[DBI:dbDisconnect]{terminate} the database connection when finished.}
}
\value{
\code{data}, invisibly.
}
\description{
Updates the specified RDB table with the provided data.
}
\seealso{
Other administration functions:
\code{\link{backup_rdb}()},
\code{\link{clean_attachment_bucket}()},
\code{\link{config_nocodb_tbls}()},
\code{\link{connect}()},
\code{\link{create_nocodb_tbl_cols}()},
\code{\link{create_nocodb_tbls}()},
\code{\link{delete_rfrnds}()},
\code{\link{dm}()},
\code{\link{notify_postgrest}()},
\code{\link{purge_nocodb}()},
\code{\link{reset_nocodb}()},
\code{\link{reset_rdb}()},
\code{\link{restart_neon_ep}()},
\code{\link{restart_postgrest}()},
\code{\link{restore_rdb}()},
\code{\link{update_administrative_units}()},
\code{\link{update_countries}()},
\code{\link{update_langs}()},
\code{\link{update_municipalities}()},
\code{\link{update_rfrnds}()},
\code{\link{update_subnational_entities}()},
\code{\link{update_topics}()}
}
\concept{admin}
