% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{update_topics}
\alias{update_topics}
\title{Update RDB \code{topics} table}
\usage{
update_topics(sweep = TRUE, connection = connect(), disconnect = TRUE)
}
\arguments{
\item{sweep}{Whether or not to also sweep the table, i.e. delete rows that are not contained in \code{data}.}

\item{connection}{\code{\link[RPostgres:PqConnection-class]{PqConnection}}\cr
Database connection to a PostgreSQL DBMS as returned by \code{\link[=connect]{connect()}}.}

\item{disconnect}{\code{\link[base:logical]{logical(1)}}\cr
Whether or not to \link[DBI:dbDisconnect]{terminate} the database connection when finished.}
}
\value{
\code{\link[=nest_topics]{nest_topics(data_topics)}}, invisibly.
}
\description{
Updates the RDB's \code{languages} table with the data from \code{\link[=nest_topics]{nest_topics(data_topics)}}.
}
\seealso{
Other administration functions:
\code{\link{backup_rdb}()},
\code{\link{clean_attachment_bucket}()},
\code{\link{config_nocodb_tbls}()},
\code{\link{connect}()},
\code{\link{create_nocodb_tbl_cols}()},
\code{\link{create_nocodb_tbls}()},
\code{\link{delete_rfrnds}()},
\code{\link{dm}()},
\code{\link{notify_postgrest}()},
\code{\link{purge_nocodb}()},
\code{\link{reset_nocodb}()},
\code{\link{reset_rdb}()},
\code{\link{restart_neon_ep}()},
\code{\link{restart_postgrest}()},
\code{\link{restore_rdb}()},
\code{\link{update_administrative_units}()},
\code{\link{update_countries}()},
\code{\link{update_langs}()},
\code{\link{update_municipalities}()},
\code{\link{update_rfrnds}()},
\code{\link{update_subnational_entities}()},
\code{\link{update_tbl}()}

Other referendum topic functions:
\code{\link{data_topics}},
\code{\link{hierarchize_topics}()},
\code{\link{hierarchize_topics_fast}()},
\code{\link{infer_topics}()},
\code{\link{nest_topics}()},
\code{\link{topics}()}
}
\concept{admin}
\concept{topics}
