% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{upload_ncdb_attachments}
\alias{upload_ncdb_attachments}
\title{Upload NocoDB attachments}
\usage{
upload_ncdb_attachments(
  paths,
  names = fs::path_file(paths),
  types = mime::guess_type(paths),
  auth_token = pal::pkg_config_val(key = "ncdb_token", pkg = this_pkg),
  max_tries = 5L,
  verbosity = NULL
)
}
\arguments{
\item{paths}{Path(s) to the file(s) to be uploaded. A character vector.}

\item{names}{File name(s) to assign to the uploaded file(s). A character vector. Note that a random string like \verb{_NnW3C} is always appended to the resulting
file name before its file type suffix, so \code{name = "some-pic.jpg"} will result in something like \code{some-pic_NnW3C.jpg}, for example.}

\item{types}{\href{https://developer.mozilla.org/en-US/docs/Web/HTTP/Basics_of_HTTP/MIME_types}{Media type}(s) of the uploaded files. A character vector.}

\item{auth_token}{NocoDB API authentication token. A character scalar.}

\item{max_tries}{Maximum number of attempts to retry in case of an HTTP error. An integerish scalar.}

\item{verbosity}{How much information to print? This is a wrapper
around \code{req_verbose()} that uses an integer to control verbosity:
\itemize{
\item 0: no output
\item 1: show headers
\item 2: show headers and bodies
\item 3: show headers, bodies, and curl status messages.
}

Use \code{\link[httr2:with_verbosity]{with_verbosity()}} to control the verbosity of requests that
you can't affect directly.}
}
\value{
A \link[tibble:tbl_df-class]{tibble} containing metadata about the uploaded attachments.
}
\description{
Uploads one or more files to the RDB NocoDB server as \href{https://docs.nocodb.com/fields/field-types/custom-types/attachment/}{attachments} via the
\href{https://meta-apis-v2.nocodb.com/#tag/Storage}{\code{POST /api/v2/storage/upload}} API endpoint. Uploaded files are stored under \verb{/nc/uploads/} in the Backblaze
B2 bucket \code{rdb-attachments}.
}
\seealso{
Other NocoDB API functions
\code{\link{config_ncdb}()},
\code{\link{create_ncdb_base}()},
\code{\link{create_ncdb_user}()},
\code{\link{ncdb_api}()},
\code{\link{ncdb_base}()},
\code{\link{ncdb_base_id}()},
\code{\link{ncdb_bases}()},
\code{\link{ncdb_tbl}()},
\code{\link{ncdb_tbl_col}()},
\code{\link{ncdb_tbl_col_id}()},
\code{\link{ncdb_tbl_cols}()},
\code{\link{ncdb_tbl_id}()},
\code{\link{ncdb_tbls}()},
\code{\link{ncdb_users}()},
\code{\link{reorder_ncdb_tbl}()},
\code{\link{set_ncdb_display_val}()},
\code{\link{set_ncdb_display_vals}()},
\code{\link{set_ncdb_tbl_metadata}()},
\code{\link{update_ncdb_base}()},
\code{\link{update_ncdb_tbl}()},
\code{\link{update_ncdb_tbl_col}()}
}
\concept{ncdb}
\keyword{internal}
