% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{url_admin_portal}
\alias{url_admin_portal}
\title{Assemble RDB admin portal URL}
\usage{
url_admin_portal(
  ...,
  .use_testing_server = pal::pkg_config_val(key = "use_testing_server", pkg = this_pkg)
)
}
\arguments{
\item{...}{Optional path components added to the base URL.}

\item{.use_testing_server}{Whether or not to use the testing servers instead of the production servers for RDB Services API calls etc.}
}
\value{
A character scalar.
}
\description{
Assemble RDB admin portal URL
}
\examples{
rdb:::url_admin_portal("referendum/5bbbfd7b92a21351232e46b5")
}
\seealso{
Other URL assembly functions:
\code{\link{url_api}()},
\code{\link{url_codebook}()},
\code{\link{url_website}()}
}
\concept{url_assembly}
\keyword{internal}
