% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{url_api_old}
\alias{url_api_old}
\title{Assemble C2D Services API URL}
\usage{
url_api_old(...)
}
\arguments{
\item{...}{Optional path components added to the base URL.}
}
\value{
A character scalar.
}
\description{
Assemble C2D Services API URL
}
\examples{
rdb:::url_api_old("health")
}
\seealso{
Other URL assembly functions:
\code{\link{url_admin_portal}()},
\code{\link{url_codebook}()},
\code{\link{url_website}()}
}
\concept{url_assembly}
\keyword{internal}
