% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{url_codebook}
\alias{url_codebook}
\title{Assemble codebook URL}
\usage{
url_codebook(var_name = NULL)
}
\arguments{
\item{var_name}{Optional variable name to add as the \href{https://en.wikipedia.org/wiki/URI_fragment}{fragment identifier} of the returned URL, which leads to a
direct link to the relevant codebook section. Must be either one of the column names of \code{\link{data_codebook}}, or a valid fragment identifier of a codebook
section above the individual variables (institution-level-variables, referendum-level-variables or topics).}
}
\value{
A character scalar.
}
\description{
Assemble codebook URL
}
\examples{
rdb:::url_codebook("level")
rdb:::url_codebook("topics")
}
\seealso{
Other URL assembly functions:
\code{\link{url_admin_portal}()},
\code{\link{url_api_old}()},
\code{\link{url_ncdb}()},
\code{\link{url_website}()}
}
\concept{url_assembly}
\keyword{internal}
