% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{url_ncdb}
\alias{url_ncdb}
\title{Assemble NocoDB URL}
\usage{
url_ncdb(...)
}
\arguments{
\item{...}{Optional path components added to the base URL.}
}
\value{
A character scalar.
}
\description{
Assemble NocoDB URL
}
\examples{
rdb:::url_ncdb("api/v2/meta/bases/")
}
\seealso{
Other URL assembly functions:
\code{\link{url_admin_portal}()},
\code{\link{url_api_old}()},
\code{\link{url_codebook}()},
\code{\link{url_website}()}
}
\concept{url_assembly}
\keyword{internal}
