% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d-GEN.R
\name{v_vals}
\alias{v_vals}
\title{Get the set of possible \emph{values} of a referendum data variable}
\usage{
v_vals(v_name)
}
\arguments{
\item{v_name}{A variable name. Must be one of the column names of \code{\link{data_codebook}}.}
}
\value{
If \code{v_name}'s values are restricted to a predefined set and
\itemize{
\item \code{v_name} is \emph{not} of type list, a vector of the same type as \code{v_name}.
\item \code{v_name} is of type list, a vector of the same type as the elements of \code{v_name}.
}

Else \code{NULL}.
}
\description{
Returns a vector of the possible predefined values a specific column in \code{\link[=referendums]{referendums()}} can hold. If the variable values aren't restricted to a predefined
set, \code{NULL} is returned.
}
\examples{
v_vals("result")
v_vals("id")
}
\seealso{
\code{\link{data_codebook}}

Other referendum metadata functions:
\code{\link{data_codebook}},
\code{\link{val_lbls}()}
}
\concept{metadata}
