% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{val_lbls}
\alias{val_lbls}
\title{Get set of possible \emph{value labels} of referendum data variable}
\usage{
val_lbls(var_name, incl_affixes = TRUE)
}
\arguments{
\item{var_name}{Variable name present in \code{\link{data_codebook}} for which the labels are to be returned. A character scalar.}

\item{incl_affixes}{Whether or not to add the corresponding \code{value_label_prefix} and \code{value_label_suffix} to the returned labels.}
}
\value{
A character vector. Of length \code{0} if \code{var_name}'s values are not restricted to a predefined set or no value labels are defined in the
\link[=data_codebook]{codebook}.
}
\description{
Returns a character vector of value labels of a specific \code{\link[=rfrnds]{rfrnds()}} column, in the same order as \code{\link[=var_vals]{var_vals()}}, or of length \code{0} if \code{var_name}'s values are
not restricted to a predefined set or no value labels are defined in the \link[=data_codebook]{codebook}.
}
\examples{
rdb::val_lbls("result",
              incl_affixes = FALSE)
rdb::val_lbls("result")

# Convert the labels to sentence case with trailing dot
rdb::val_lbls("result") |> pal::sentenceify()
}
\seealso{
Other referendum metadata functions:
\code{\link{data_codebook}},
\code{\link{prettify_var_names}()},
\code{\link{val_scale}()},
\code{\link{var_name_unnested}()},
\code{\link{var_vals}()}
}
\concept{metadata}
