% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{val_scale}
\alias{val_scale}
\title{Get \emph{value scale} of referendum data variables}
\usage{
val_scale(var_names)
}
\arguments{
\item{var_names}{Variable name(s) present in \code{\link{data_codebook}} for which the value scale is to be returned. A character vector.}
}
\value{
A character scalar.
}
\description{
Returns the value scale of the specified \code{\link[=rfrnds]{rfrnds()}} columns.
}
\examples{
rdb::val_scale("level")
paste0("topics_tier_", 1:3) |> rdb::val_scale()
}
\seealso{
Other referendum metadata functions:
\code{\link{data_codebook}},
\code{\link{printify_var_names}()},
\code{\link{val_lbls}()},
\code{\link{var_name_unnested}()},
\code{\link{var_vals}()}
}
\concept{metadata}
