% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d-GEN.R
\name{validate_referendums}
\alias{validate_referendums}
\title{Validate referendums}
\usage{
validate_referendums(
  data,
  check_applicability_constraint = TRUE,
  check_id_sudd_prefix = TRUE
)
}
\arguments{
\item{data}{The referendum data to validate, as returned by \code{\link[=referendums]{referendums()}}.}

\item{check_applicability_constraint}{Whether or not to check that no applicability constraints as defined in the \link[=data_codebook]{codebook} are violated.}

\item{check_id_sudd_prefix}{Whether or not to check that all \href{https://rpkg.dev/c2d/articles/codebook.html#id-sudd}{\code{id_sudd}} prefixes are valid.}
}
\value{
\code{data}, invisibly.
}
\description{
Validate referendums
}
