% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{var_vals}
\alias{var_vals}
\title{Get the set of possible \emph{values} of a referendum data variable}
\usage{
var_vals(var_name)
}
\arguments{
\item{var_name}{Variable name present in \code{\link{data_codebook}}. A character scalar.}
}
\value{
If \code{var_name}'s values are restricted to a predefined set and
\itemize{
\item \code{var_name} is \emph{not} of type list, a vector of the same type as \code{var_name}.
\item \code{var_name} is of type list, a vector of the same type as the elements of \code{var_name}.
}

Else \code{NULL}.
}
\description{
Returns a vector of the possible predefined values a specific column in \code{\link[=rfrnds]{rfrnds()}} can hold. If the variable values aren't restricted to a predefined
set, \code{NULL} is returned.
}
\examples{
var_vals("result")
var_vals("id")
}
\seealso{
Other referendum metadata functions:
\code{\link{data_codebook}},
\code{\link{printify_var_names}()},
\code{\link{val_lbls}()}
}
\concept{metadata}
