% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d.gen.R
\name{add_country_code_long}
\alias{add_country_code_long}
\title{Add long country code to referendum data}
\usage{
add_country_code_long(data)
}
\arguments{
\item{data}{C2D referendum data as returned by \code{\link[=rfrnds]{rfrnds()}}. A data frame that at minimum contains the column \code{country_code} (with \href{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2}{ISO 3166-1 alpha-2} or \href{https://en.wikipedia.org/wiki/ISO_3166-3}{ISO 3166-3 alpha-4} codes).}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Augments \code{data} with an additional column holding the current or former three-letter \href{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-3}{ISO 3166-1 alpha-3}
code of the country in which the referendum took place (see \href{https://en.wikipedia.org/wiki/ISO_3166-3_alpha-3}{ISO 3166-3} for former country codes).
}
\examples{
c2d::rfrnds(date_max = clock::date_today(zone = "UTC"),
            date_min = clock::date_today(zone = "UTC") |> clock::add_years(-1L)) |>
  c2d:::add_country_code_long() |>
  dplyr::select(any_of("id"),
                starts_with("country_code"))
}
\seealso{
Other referendum data augmentation functions:
\code{\link{add_country_code_continual}()},
\code{\link{add_country_name_long}()},
\code{\link{add_former_country_flag}()},
\code{\link{add_period}()},
\code{\link{add_turnout}()},
\code{\link{add_world_regions}()}
}
\concept{augment}
