% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d.gen.R
\name{add_period}
\alias{add_period}
\title{Add period}
\usage{
add_period(
  data,
  period = c("week", "month", "quarter", "year", "decade", "century")
)
}
\arguments{
\item{data}{C2D referendum data as returned by \code{\link[=rfrnds]{rfrnds()}}. A data frame that at minimum contains the column \code{date}.}

\item{period}{Type of period to add. One of \code{"week"},\code{"month"},\code{"quarter"},\code{"year"},\code{"decade"} or \code{"century"}.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Adds an additional column holding the specified period in which the referendum took place. The new column is named after \code{period} and its values are always
of type integer.
}
\examples{
c2d::rfrnds(country_code = "AT") |>
  c2d::add_period() |>
  dplyr::select(id, date, week)

c2d::rfrnds(country_code = "AT") |>
  c2d::add_period("year") |>
  dplyr::select(id, date, year)
}
\seealso{
Other referendum data augmentation functions:
\code{\link{add_world_regions}()}
}
\concept{augment}
