% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d-GEN.R
\name{add_referendums}
\alias{add_referendums}
\title{Add new referendums to the C2D database}
\usage{
add_referendums(
  data,
  email = names(getOption("c2d.credentials")[1L]),
  password = getOption("c2d.credentials")[1L]
)
}
\arguments{
\item{data}{The new referendum data. A \link[tibble:tbl_df-class]{tibble} that in any case must contain the columns
\itemize{
\item \href{https://rpkg.dev/c2d/articles/codebook.html#country_code}{\code{country_code}}
\item \href{https://rpkg.dev/c2d/articles/codebook.html#level}{\code{level}}
\item \href{https://rpkg.dev/c2d/articles/codebook.html#date}{\code{date}}
\item \href{https://rpkg.dev/c2d/articles/codebook.html#title_en}{\code{title_en}}
\item \href{https://rpkg.dev/c2d/articles/codebook.html#result}{\code{result}}
\item \href{https://rpkg.dev/c2d/articles/codebook.html#electorate_total}{\code{electorate_total}}
\item \href{https://rpkg.dev/c2d/articles/codebook.html#electorate_abroad}{\code{electorate_abroad}}
\item \href{https://rpkg.dev/c2d/articles/codebook.html#votes_yes}{\code{votes_yes}}
\item \href{https://rpkg.dev/c2d/articles/codebook.html#votes_no}{\code{votes_no}}
\item \href{https://rpkg.dev/c2d/articles/codebook.html#votes_empty}{\code{votes_empty}}
\item \href{https://rpkg.dev/c2d/articles/codebook.html#votes_invalid}{\code{votes_invalid}}
\item \href{https://rpkg.dev/c2d/articles/codebook.html#is_draft}{\code{is_draft}}
\item \href{https://rpkg.dev/c2d/articles/codebook.html#type}{\code{type}}
}

plus the column \href{https://rpkg.dev/c2d/articles/codebook.html#subnational_entity_name}{\code{subnational_entity_name}} for referendums of
\href{https://rpkg.dev/c2d/articles/codebook.html#subnational_entity_name}{\code{level}} below \code{"national"}, and the column
\href{https://rpkg.dev/c2d/articles/codebook.html#municipality}{\code{municipality}} for referendums of \code{level = "local"},

plus any additional \link[=codebook]{valid} columns containing the values for the corresponding database fields.}

\item{email}{The e-mail address of the C2D API user account to be used for authentication. A character scalar.}

\item{password}{The password of the C2D API user account to be used for authentication. A character scalar.}
}
\value{
\code{data}, invisibly.
}
\description{
Adds new referendum entries to the C2D database via \href{https://github.com/ccmdesign/c2d-app/blob/master/docs/services.md#3-referendum-routes}{its API}.
}
\details{
Note that the following variables are not supported, i.e. simply dropped from \code{data}:
\itemize{
\item \code{id_official}
\item \code{id_sudd}
\item \code{files}
}
}
\seealso{
Other referendum data functions:
\code{\link{count_referendums}()},
\code{\link{download_file_attachment}()},
\code{\link{edit_referendums}()},
\code{\link{referendums}()},
\code{\link{referendum}()},
\code{\link{search_referendums}()},
\code{\link{validate_referendums}()}
}
\concept{referendum}
