% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{add_rfrnds}
\alias{add_rfrnds}
\title{Add new referendums to the RDB}
\usage{
add_rfrnds(
  data,
  email = pal::pkg_config_val(key = "api_username", pkg = this_pkg),
  password = pal::pkg_config_val(key = "api_password", pkg = this_pkg),
  use_testing_server = pal::pkg_config_val(key = "use_testing_server", pkg = this_pkg),
  quiet = FALSE
)
}
\arguments{
\item{data}{The new referendum data. A \link[tibble:tbl_df-class]{tibble} that in any case must contain the columns
\itemize{
\item \href{https://rdb.rpkg.dev/articles/codebook.html#country-code}{\code{country_code}}
\item \href{https://rdb.rpkg.dev/articles/codebook.html#level}{\code{level}}
\item \href{https://rdb.rpkg.dev/articles/codebook.html#date}{\code{date}}
\item \href{https://rdb.rpkg.dev/articles/codebook.html#title-en}{\code{title_en}}
\item \href{https://rdb.rpkg.dev/articles/codebook.html#result}{\code{result}}
\item \href{https://rdb.rpkg.dev/articles/codebook.html#electorate-total}{\code{electorate_total}}
\item \href{https://rdb.rpkg.dev/articles/codebook.html#electorate-abroad}{\code{electorate_abroad}}
\item \href{https://rdb.rpkg.dev/articles/codebook.html#votes-yes}{\code{votes_yes}}
\item \href{https://rdb.rpkg.dev/articles/codebook.html#votes-no}{\code{votes_no}}
\item \href{https://rdb.rpkg.dev/articles/codebook.html#votes-empty}{\code{votes_empty}}
\item \href{https://rdb.rpkg.dev/articles/codebook.html#votes-invalid}{\code{votes_invalid}}
\item \href{https://rdb.rpkg.dev/articles/codebook.html#is-draft}{\code{is_draft}}
\item \href{https://rdb.rpkg.dev/articles/codebook.html#type}{\code{type}}
}

plus the column \href{https://rdb.rpkg.dev/articles/codebook.html#subnational-entity-name}{\code{subnational_entity_name}} for referendums of
\href{https://rdb.rpkg.dev/articles/codebook.html#subnational-entity-name}{\code{level}} below \code{"national"}, and the column \href{https://rdb.rpkg.dev/articles/codebook.html#municipality}{\code{municipality}} for referendums
of \code{level = "local"},

plus any additional \link[=codebook]{valid} columns containing the values for the corresponding database fields.}

\item{email}{The e-mail address of the RDB API user account to be used for authentication. A character scalar.}

\item{password}{The password of the RDB API user account to be used for authentication. A character scalar.}

\item{use_testing_server}{Whether or not to use the testing servers instead of the production servers for RDB Services API calls etc.}

\item{quiet}{Whether or not to print the newly created referendum IDs to console.}
}
\value{
A character vector of newly created referendum IDs.
}
\description{
Adds new referendum entries to the Referendum Database (RDB) via \href{https://github.com/zdaarau/c2d-app/blob/master/docs/services.md#3-referendum-routes}{its API}.
}
\details{
Note that adding/editing the column \code{files} is not supported, i.e. it is simply dropped from \code{data}.
}
\seealso{
Other referendum data functions:
\code{\link{assert_vars}()},
\code{\link{count_rfrnds}()},
\code{\link{delete_rfrnds}()},
\code{\link{download_file_attachment}()},
\code{\link{edit_rfrnds}()},
\code{\link{rfrnd_exists}()},
\code{\link{rfrnds_old}()},
\code{\link{rfrnds}()},
\code{\link{rfrnd}()},
\code{\link{search_rfrnds}()},
\code{\link{validate_rfrnds}()}
}
\concept{rfrnd}
