% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{add_turnout}
\alias{add_turnout}
\title{Add turnout to referendum data}
\usage{
add_turnout(data, rough = TRUE, excl_dubious = TRUE)
}
\arguments{
\item{data}{RDB referendum data as returned by \code{\link[=rfrnds]{rfrnds()}}. A data frame that at minimum contains the columns \code{electorate_total}, \code{votes_yes}, \code{votes_no},
\code{votes_empty} and \code{votes_invalid}.}

\item{rough}{Whether to fall back on a "rough" calculation of the turnout in case any of the variables \code{votes_empty} or \code{votes_invalid} is unknown (\code{NA}), or
to be strict and return \code{NA} in such a case.}

\item{excl_dubious}{Whether or not to exclude obviously dubious turnout numbers (those > 1.0) by setting them to \code{NA}. Such numbers stem either from
data errors or (officially) tampered numbers.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Augments \code{data} with an additional column \code{turnout} containing the voter turnout calculated as:

\Sexpr[results=rd, stage=build]{
katex::math_to_rd(tex = "\\\\frac{votes\\\\_yes+votes\\\\_no+votes\\\\_empty+votes\\\\_invalid}{electorate\\\\_total}",
ascii = "(votes_yes + votes_no + votes_empty + votes_invalid) / electorate_total",
displayMode = TRUE)
}
}
\examples{
# rough turnout numbers
rdb::rfrnds(quiet = TRUE) |>
  rdb::add_turnout() |>
  dplyr::select(id,
                electorate_total,
                starts_with("votes_"),
                turnout)

# strict turnout numbers
rdb::rfrnds(quiet = TRUE) |>
  rdb::add_turnout(rough = FALSE) |>
  dplyr::select(id,
                electorate_total,
                starts_with("votes_"),
                turnout)
}
\seealso{
Other referendum data augmentation functions:
\code{\link{add_country_code_continual}()},
\code{\link{add_country_code_long}()},
\code{\link{add_country_name_long}()},
\code{\link{add_country_name}()},
\code{\link{add_former_country_flag}()},
\code{\link{add_period}()},
\code{\link{add_urls}()},
\code{\link{add_world_regions}()}
}
\concept{augment}
