% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d-GEN.R
\name{assemble_query_filter}
\alias{assemble_query_filter}
\title{Assemble MongoDB query filter document}
\usage{
assemble_query_filter(
  country_code = NULL,
  subnational_entity_name = NULL,
  municipality = NULL,
  level = NULL,
  type = NULL,
  date = NULL,
  is_draft = NULL,
  date_time_created_min = NULL,
  date_time_created_max = NULL,
  query_filter = NULL,
  base64_encode = TRUE
)
}
\arguments{
\item{country_code}{The \code{country_code}(s) to be included. A character vector.}

\item{subnational_entity_name}{The \code{subnational_entity_name}(s) to be included. A character vector.}

\item{municipality}{The \code{municipality}(s) to be included. A character vector.}

\item{level}{The \code{level}(s) to be included. A character vector.}

\item{type}{The \code{type}(s) to be included. A character vector.}

\item{date}{The \code{date}(s) to be included. A character vector.}

\item{is_draft}{\code{TRUE} means to include only referendum entries with \emph{draft} status, \code{FALSE} to include only normal entries. Set to \code{NULL} in order to
include both draft and normal entries.}

\item{date_time_created_min}{The minimum \code{date_time_created} to be included. A \link[base:Dates]{date}, a \link[base:DateTimeClasses]{datetime}, or something
coercible to (like \code{"2006-01-02"} or \code{"2006-01-02T15:04:05"}).}

\item{date_time_created_max}{The maximum \code{date_time_created} to be included. A \link[base:Dates]{date}, a \link[base:DateTimeClasses]{datetime}, or something
coercible to (like \code{"2006-01-02"} or \code{"2006-01-02T15:04:05"}).}

\item{query_filter}{A valid \href{https://docs.mongodb.com/manual/core/document/#query-filter-documents}{MongoDB JSON query filter document} which allows for
maximum control over what data is included. This takes precedence over all the other parameters, i.e. if \code{query_filter} is provided, the parameters
\code{country_code}, \code{subnational_entity_name}, \code{municipality}, \code{level}, \code{type}, \code{date}, \code{is_draft}, \code{date_time_created_min} and \code{date_time_created_max})`` are ignored.}

\item{base64_encode}{Whether or not to \href{https://en.wikipedia.org/wiki/Base64}{Base64}-encode the resulting query filter document. Note that the
\code{query_filter} argument provided to other functions of this package must be Base64-encoded.}
}
\value{
A character scalar containing a valid \href{https://docs.mongodb.com/manual/core/document/#query-filter-documents}{MongoDB JSON query filter document},
\href{https://en.wikipedia.org/wiki/Base64}{Base64}-encoded if \code{base64_encode = TRUE}.
}
\description{
Assemble MongoDB query filter document
}
