% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{assert_vars}
\alias{assert_vars}
\title{Assert referendum variables are present}
\usage{
assert_vars(data, vars)
}
\arguments{
\item{data}{RDB referendum data as returned by \code{\link[=rfrnds]{rfrnds()}}.}

\item{vars}{Names of the variables to check. A character vector.}
}
\value{
\code{data}, invisibly.
}
\description{
Asserts the specified \code{vars} are present in the supplied referendum \code{data}. Depending on \code{vars}, additional integrity checks are performed.
}
\examples{
rdb::rfrnd(id = "5bbbe26a92a21351232dd73f") |> rdb::assert_vars(vars = "country_code")

try(
  tibble::tibble(country_code = "AN") |> rdb::assert_vars(vars = "country_code")
)
}
\seealso{
Other referendum data functions:
\code{\link{add_rfrnds}()},
\code{\link{count_rfrnds}()},
\code{\link{delete_rfrnds}()},
\code{\link{download_file_attachment}()},
\code{\link{edit_rfrnds}()},
\code{\link{rfrnd_exists}()},
\code{\link{rfrnds_old}()},
\code{\link{rfrnds}()},
\code{\link{rfrnd}()},
\code{\link{search_rfrnds}()},
\code{\link{validate_rfrnds}()}
}
\concept{rfrnd}
