% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d.gen.R
\name{auth_session}
\alias{auth_session}
\title{Authenticate a user session for the \href{https://github.com/ccmdesign/c2d-app/blob/master/docs/services.md#1-reflexive-routes}{C2D API}}
\usage{
auth_session(
  email = Sys.getenv("C2D_API_USER"),
  password = Sys.getenv("C2D_API_PW"),
  quiet = FALSE
)
}
\arguments{
\item{email}{The e-mail address of the user for which a session should be created. A character scalar.}

\item{password}{The password of the user for which a session should be created. A character scalar.}

\item{quiet}{Whether or not to suppress printing status output from internal processing.}
}
\value{
The user session token as a character scalar, invisibly.
}
\description{
Creates a new user session token if necessary. The token is stored in the R option \code{c2d.user_session_tokens}, a \link[tibble:tbl_df-class]{tibble} with the columns
\code{email}, \code{token} and \code{date_time_last_active}.
}
\details{
By default, \code{email} and \code{password} are read from the \href{https://en.wikipedia.org/wiki/Environment_variable}{environment variables} \code{C2D_API_EMAIL} and
\code{C2D_API_PW} respectively.

User session tokens expire automatically after 15 days of inactivity.
}
\keyword{internal}
