% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d-GEN.R
\name{auth_session}
\alias{auth_session}
\title{Authenticate a user session for the \href{https://github.com/ccmdesign/c2d-app/blob/master/docs/services.md#1-reflexive-routes}{C2D API}}
\usage{
auth_session(
  email = names(getOption("c2d.credentials")[1L]),
  password = getOption("c2d.credentials")[1L]
)
}
\arguments{
\item{email}{The e-mail address of the user for which a session should be created. A character scalar.}

\item{password}{The password of the user for which a session should be created. A character scalar.}
}
\value{
The user session token as a character scalar, invisibly.
}
\description{
Creates a new user session token if necessary. The token is stored in the R option \code{c2d.user_session_tokens}, named by \code{email}. By default, the \code{email} and
\code{password} are read from the first element of the R option \code{c2d.credentials} (a named character vector where names are e-mail addresses and values are
passwords).`
}
\keyword{internal}
