% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{auth_session}
\alias{auth_session}
\title{Authenticate a user session for the \href{https://github.com/ccmdesign/c2d-app/blob/master/docs/services.md#1-reflexive-routes}{RDB API}}
\usage{
auth_session(
  email = pal::pkg_config_val(key = "api_username", pkg = this_pkg),
  password = pal::pkg_config_val(key = "api_password", pkg = this_pkg),
  use_testing_server = pal::pkg_config_val(key = "use_testing_server", pkg = this_pkg),
  quiet = FALSE
)
}
\arguments{
\item{email}{The e-mail address of the user for which a session should be created. A character scalar.}

\item{password}{The password of the user for which a session should be created. A character scalar.}

\item{use_testing_server}{Whether or not to use the testing servers instead of the production servers for RDB Services API calls etc.}

\item{quiet}{Whether or not to suppress printing status output from internal processing.}
}
\value{
The user session token as a character scalar, invisibly.
}
\description{
Creates a new user session token if necessary. The token is stored in the R option \code{rdb.user_session_tokens}, a \link[tibble:tbl_df-class]{tibble} with the columns
\code{email}, \code{token} and \code{date_time_last_active}.
}
\details{
\code{email} and \code{password} default to the \link[=pkg_config]{package configuration options} \code{api_username} and \code{api_password} respectively.

User session tokens expire automatically after 15 days of inactivity.
}
\keyword{internal}
