% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d-package.R
\docType{package}
\name{c2d-package}
\alias{c2d}
\alias{c2d-package}
\title{c2d: Download Data from the C2D Database, Which Covers Direct Democratic
    Votes Worldwide}
\description{
c2d provides access to the C2D Database (https://c2d.ch) from
    R. This database aims to record all direct democratic votes worldwide
    and is operated by the Centre for Democracy Studies Aarau (ZDA) at the
    University of Zurich, Switzerland.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rpkg.dev/c2d}
  \item \url{https://gitlab.com/zdaarau/c2d}
  \item Report bugs at \url{https://gitlab.com/zdaarau/c2d/issues}
}

}
\author{
\strong{Maintainer}: Salim Brüggemann \email{salim-b@pm.me} (\href{https://orcid.org/0000-0002-5329-5987}{ORCID})

Authors:
\itemize{
  \item Uwe Serdült \email{uwe.serduelt@zda.uzh.ch} (\href{https://orcid.org/0000-0002-2383-3158}{ORCID})
}

Other contributors:
\itemize{
  \item  Centre for Democracy Studies Aarau (ZDA) \email{info@zdaarau.ch} [copyright holder]
}

}
\keyword{internal}
