% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{config_nocodb}
\alias{config_nocodb}
\title{Configure NocoDB}
\usage{
config_nocodb(hostname = nocodb_hostname, quiet = FALSE)
}
\arguments{
\item{hostname}{NocoDB server \href{https://en.wikipedia.org/wiki/Hostname}{hostname}. A character scalar.}

\item{quiet}{Whether or not to suppress printing status output from internal processing.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Configures the RDB NocoDB server according to our needs. Currently, this includes:
\itemize{
\item Setting the correct display value columm for all tables via \link[nocodb:set_display_vals]{nocodb::set_display_vals}.
\item Setting the proper metadata for all tables via \link[nocodb:set_tbl_metadata]{nocodb::set_tbl_metadata}.
}
}
\seealso{
Other administration functions:
\code{\link{connect}()},
\code{\link{delete_rfrnds}()},
\code{\link{dm}()},
\code{\link{update_countries}()},
\code{\link{update_langs}()},
\code{\link{update_municipalities}()},
\code{\link{update_rfrnds}()},
\code{\link{update_subnational_entities}()},
\code{\link{update_tbl}()}
}
\concept{admin}
\concept{nocodb}
