% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{connect}
\alias{connect}
\title{Connect to RDB DBMS}
\usage{
connect(
  dbname = pg_db,
  host = pal::pkg_config_val(key = "pg_host", pkg = this_pkg),
  user = pal::pkg_config_val(key = "pg_user", pkg = this_pkg),
  password = pal::pkg_config_val(key = "pg_password", pkg = this_pkg),
  sslmode = "verify-full"
)
}
\arguments{
\item{dbname}{Name of the RDB PostgreSQL database. Defaults to \code{"rdb"}.}

\item{host}{RDB PostgreSQL (neon.tech) host address. Defaults to a \href{https://neon.tech/docs/introduction/read-replicas}{read replica} which has no write access to the database.}

\item{user}{RDB PostgreSQL username for database operations performed via this package. To be able to edit the database, the user must have sufficient privileges. Defaults to the read-only user \code{r_anon}.}

\item{password}{Password for \code{pg_user}. Defaults to the password for the read-only user \code{r_anon}.}

\item{sslmode}{PostgreSQL connection parameter that determines whether or with what priority a secure TLS connection is negotiated with the server. See the
\href{https://www.postgresql.org/docs/current/libpq-connect.html#LIBPQ-CONNECT-SSLMODE}{official documentation} for more information about the supported
keywords.}
}
\value{
An S4 object that inherits from \code{\link[DBI:DBIConnection-class]{DBI::DBIConnection}}. This object is used to communicate with the database engine.
}
\description{
Connects to RDB's database management system, a PostgreSQL server managed by \href{https://neon.tech/}{neon.tech}.
}
\examples{
con <- rdb::connect()

con |> DBI::dbListTables()
}
\seealso{
Other administration functions:
\code{\link{backup_rdb}()},
\code{\link{clean_attachment_bucket}()},
\code{\link{config_nocodb_tbls}()},
\code{\link{create_nocodb_tbl_cols}()},
\code{\link{create_nocodb_tbls}()},
\code{\link{delete_rfrnds}()},
\code{\link{dm}()},
\code{\link{notify_postgrest}()},
\code{\link{purge_nocodb}()},
\code{\link{reset_nocodb}()},
\code{\link{reset_rdb}()},
\code{\link{restart_neon_ep}()},
\code{\link{restart_postgrest}()},
\code{\link{restore_rdb}()},
\code{\link{update_administrative_units}()},
\code{\link{update_countries}()},
\code{\link{update_langs}()},
\code{\link{update_municipalities}()},
\code{\link{update_nocodb_users}()},
\code{\link{update_rfrnds}()},
\code{\link{update_subnational_entities}()},
\code{\link{update_tbl}()},
\code{\link{update_topics}()}
}
\concept{admin}
