% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\docType{data}
\name{data_codebook}
\alias{data_codebook}
\alias{codebook}
\title{RDB Codebook}
\format{
A \link[tibble:tbl_df-class]{tibble}.
}
\usage{
data_codebook
}
\description{
A tibble containing the complete metadata of all \code{\link[=rfrnds]{rfrnds()}} variables. The codebook below is also available \href{https://rdb.rpkg.dev/articles/codebook.html}{online}.
}
\section{Codebook}{
\subsection{Referendum-level variables}{

For the following variables, the \href{https://en.wikipedia.org/wiki/Unit_of_observation}{unit of observation} is the individual referendum.
\subsection{\code{id}}{

A 24-character value that uniquely identifies the referendum and corresponds to the \href{https://docs.mongodb.com/manual/reference/bson-types/#std-label-objectid}{MongoDB \code{ObjectId}} in the RDB.
\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal} level.
}

}

\subsection{\code{id_official}}{

The referendum's official identifier assigned by the authorities; this is only guaranteed to be unique per respective jurisdiction.
\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal} level.
}

}

\subsection{\code{id_sudd}}{

The referendum's identifier assigned by \href{https://sudd.ch/}{sudd.ch}; multiple referendum's might share the same identifier.
\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal} level.
}

}

\subsection{\code{country_code}}{

The two- or four-letter code identifying the country in which the referendum took place; an \href{https://en.wikipedia.org/wiki/ISO_3166-3}{ISO 3166-3 alpha-4 code} for historical countries which ceased to exist, otherwise an \href{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2}{ISO 3166-1 alpha-2 code}.
\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal} level.
}

}

\subsection{\code{country_code_long}}{

The three-letter ISO 3166-1 alpha-3 code identifying the country in which the referendum took place.
\subsection{Origin}{

This variable is derived from \href{https://rdb.rpkg.dev/articles/codebook.html#country-code}{\code{country_code}} and can be generated on demand via \code{\link[=add_country_code_long]{add_country_code_long()}}.
}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal} level.
}

}

\subsection{\code{country_code_continual}}{

The current or future \href{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2}{ISO 3166-1 alpha-2} code of the country where the referendum took place; identical to \code{country_code} if the country still exists, otherwise the \code{country_code} of the successor country (the largest one in terms of population if it was succeeded by multiple countries).
\subsection{Origin}{

This variable is derived from \href{https://rdb.rpkg.dev/articles/codebook.html#country-code}{\code{country_code}} and can be generated on demand via \code{\link[=add_country_code_continual]{add_country_code_continual()}}.
}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal} level.
}

}

\subsection{\code{country_name}}{

The English name of the country in which the referendum took place.
\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal} level.
}

}

\subsection{\code{country_name_long}}{

The official full English name(s) of the country in which the referendum took place.
\subsection{Origin}{

This variable is derived from \href{https://rdb.rpkg.dev/articles/codebook.html#country-code}{\code{country_code}} and can be generated on demand via \code{\link[=add_country_name_long]{add_country_name_long()}}.
}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal} level.
}

}

\subsection{\code{subnational_entity_name}}{

The English name of the \href{https://en.wikipedia.org/wiki/Administrative_division}{subnational political entity} in which the referendum took place.
\subsection{Applicability constraint}{

This variable is only applicable under the following condition:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{level \%in\% c('local', 'subnational')
}\if{html}{\out{</div>}}
}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal} level.
}

}

\subsection{\code{municipality_name}}{

The municipality in which the referendum took place.
\subsection{Applicability constraint}{

This variable is only applicable under the following condition:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{level == 'local'
}\if{html}{\out{</div>}}
}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal} level.
}

}

\subsection{\code{un_country_code}}{

The UN M49 code of the country in which the referendum took place.
\subsection{Origin}{

This variable is derived from \href{https://rdb.rpkg.dev/articles/codebook.html#country-code}{\code{country_code}} and can be generated on demand via \code{\link[=add_world_regions]{add_world_regions()}}.
}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal} level.
}

}

\subsection{\code{un_region_tier_1_code}}{

The area code of the UN tier-1 region in which the referendum took place.
\subsection{Origin}{

This variable is derived from \href{https://rdb.rpkg.dev/articles/codebook.html#country-code}{\code{country_code}} and can be generated on demand via \code{\link[=add_world_regions]{add_world_regions()}}.
}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal} level.
}

}

\subsection{\code{un_region_tier_1_name}}{

The English name of the UN tier-1 region in which the referendum took place.
\subsection{Origin}{

This variable is derived from \href{https://rdb.rpkg.dev/articles/codebook.html#country-code}{\code{country_code}} and can be generated on demand via \code{\link[=add_world_regions]{add_world_regions()}}.
}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal} level.
}

}

\subsection{\code{un_region_tier_2_code}}{

The area code of the UN tier-2 region in which the referendum took place.
\subsection{Origin}{

This variable is derived from \href{https://rdb.rpkg.dev/articles/codebook.html#country-code}{\code{country_code}} and can be generated on demand via \code{\link[=add_world_regions]{add_world_regions()}}.
}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal} level.
}

}

\subsection{\code{un_region_tier_2_name}}{

The English name of the UN tier-2 region in which the referendum took place.
\subsection{Origin}{

This variable is derived from \href{https://rdb.rpkg.dev/articles/codebook.html#country-code}{\code{country_code}} and can be generated on demand via \code{\link[=add_world_regions]{add_world_regions()}}.
}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal} level.
}

}

\subsection{\code{un_region_tier_3_code}}{

The area code of the UN tier-3 region in which the referendum took place.
\subsection{Origin}{

This variable is derived from \href{https://rdb.rpkg.dev/articles/codebook.html#country-code}{\code{country_code}} and can be generated on demand via \code{\link[=add_world_regions]{add_world_regions()}}.
}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal} level.
}

}

\subsection{\code{un_region_tier_3_name}}{

The English name of the UN tier-3 region in which the referendum took place.
\subsection{Origin}{

This variable is derived from \href{https://rdb.rpkg.dev/articles/codebook.html#country-code}{\code{country_code}} and can be generated on demand via \code{\link[=add_world_regions]{add_world_regions()}}.
}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal} level.
}

}

\subsection{\code{un_subregion}}{

The UN subregion in which the referendum took place; corresponds to the English UN region name of the lowest available tier (except for Northern Europe).
\subsection{Origin}{

This variable is derived from \href{https://rdb.rpkg.dev/articles/codebook.html#country-code}{\code{country_code}} and can be generated on demand via \code{\link[=add_world_regions]{add_world_regions()}}.
}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal} level.
}

}

\subsection{\code{level}}{

The political level at which the referendum took place.
\subsection{Possible values}{\tabular{ll}{
   Value \tab Meaning \cr
   \code{"local"} \tab The referendum took place at the local political level. \cr
   \code{"subnational"} \tab The referendum took place at the subnational political level. \cr
   \code{"national"} \tab The referendum took place at the national political level. \cr
}

}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{ordinal (ascending)} level.
}

}

\subsection{\code{date}}{

The date on which the referendum took place.
\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{interval} level.
}

}

\subsection{\code{week}}{

The week number (1–53) in which the referendum was held.
\subsection{Origin}{

This variable is derived from \href{https://rdb.rpkg.dev/articles/codebook.html#date}{\code{date}} and can be generated on demand via \code{\link[=add_period]{add_period()}}.
}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{interval} level.
}

}

\subsection{\code{month}}{

The month number (1–12) in which the referendum was held.
\subsection{Origin}{

This variable is derived from \href{https://rdb.rpkg.dev/articles/codebook.html#date}{\code{date}} and can be generated on demand via \code{\link[=add_period]{add_period()}}.
}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{interval} level.
}

}

\subsection{\code{quarter}}{

The quarter number (1–4) in which the referendum was held.
\subsection{Origin}{

This variable is derived from \href{https://rdb.rpkg.dev/articles/codebook.html#date}{\code{date}} and can be generated on demand via \code{\link[=add_period]{add_period()}}.
}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{interval} level.
}

}

\subsection{\code{year}}{

The year in which the referendum was held.
\subsection{Origin}{

This variable is derived from \href{https://rdb.rpkg.dev/articles/codebook.html#date}{\code{date}} and can be generated on demand via \code{\link[=add_period]{add_period()}}.
}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{interval} level.
}

}

\subsection{\code{decade}}{

The decade in which the referendum was held.
\subsection{Origin}{

This variable is derived from \href{https://rdb.rpkg.dev/articles/codebook.html#date}{\code{date}} and can be generated on demand via \code{\link[=add_period]{add_period()}}.
}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{interval} level.
}

}

\subsection{\code{century}}{

The century in which the referendum was held.
\subsection{Origin}{

This variable is derived from \href{https://rdb.rpkg.dev/articles/codebook.html#date}{\code{date}} and can be generated on demand via \code{\link[=add_period]{add_period()}}.
}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{interval} level.
}

}

\subsection{\code{is_former_country}}{

Whether or not the country where the referendum took place ceased to exist and thus has an \href{https://en.wikipedia.org/wiki/ISO_3166-3}{ISO 3166-3 alpha-4} instead of an \href{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2}{ISO 3166-1 alpha-2} \code{country_code}.
\subsection{Possible values}{\tabular{l}{
   Value \cr
   \code{TRUE} \cr
   \code{FALSE} \cr
}

}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal (binary)} level.
}

}

\subsection{\code{title_en}}{

The referendum proposal's English title.
\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal} level.
}

}

\subsection{\code{title_de}}{

The referendum proposal's German title.
\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal} level.
}

}

\subsection{\code{title_fr}}{

The referendum proposal's French title.
\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal} level.
}

}

\subsection{\code{question}}{

The proposal question exactly as it was asked in the referendum.
\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal} level.
}

}

\subsection{\code{question_en}}{

An English translation of \code{question} if it wasn't officially asked in English, otherwise the same as \code{question}.
\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal} level.
}

}

\subsection{\code{committee_name}}{

The name of the committee that initiated the referendum proposal.
\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal} level.
}

}

\subsection{\code{result}}{

The result of the referendum.
\subsection{Possible values}{\tabular{ll}{
   Value \tab Meaning \cr
   \code{"yes"} \tab The referendum proposal was approved. \cr
   \code{"no"} \tab The referendum proposal was rejected. \cr
}

}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal (binary)} level.
}

}

\subsection{\code{subterritories_yes}}{

The number of \href{https://en.wikipedia.org/wiki/Administrative_division}{administrative divisions} one level below \code{level} that approved the referendum proposal.
\subsection{Applicability constraint}{

This variable is only applicable under the following condition:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{level != 'local'
}\if{html}{\out{</div>}}
}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{ratio} level.
}

}

\subsection{\code{subterritories_no}}{

The number of \href{https://en.wikipedia.org/wiki/Administrative_division}{administrative divisions} one level below \code{level} that rejected the referendum proposal.
\subsection{Applicability constraint}{

This variable is only applicable under the following condition:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{level != 'local'
}\if{html}{\out{</div>}}
}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{ratio} level.
}

}

\subsection{\code{electorate_total}}{

The total number of people entitled to vote in the referendum.
\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{ratio} level.
}

}

\subsection{\code{electorate_abroad}}{

The number of people entitled to vote in the referendum \emph{who are living abroad}; by definition a subset of \code{electorate_total}.
\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{ratio} level.
}

}

\subsection{\code{votes_yes}}{

The number of votes approving the referendum proposal.
\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{ratio} level.
}

}

\subsection{\code{votes_no}}{

The number of votes rejecting the referendum proposal.
\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{ratio} level.
}

}

\subsection{\code{votes_empty}}{

The number of votes that did not specify any preference on the referendum proposal.
\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{ratio} level.
}

}

\subsection{\code{votes_invalid}}{

The number of votes that were invalid.
\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{ratio} level.
}

}

\subsection{\code{votes_per_subterritory}}{

The number of \code{votes_yes}, \code{votes_no}, \code{votes_empty} and \code{votes_invalid} per \href{https://en.wikipedia.org/wiki/Administrative_division}{administrative division} one level below \code{level}; currently only filled for Swiss referendums on the national level.
\subsection{Applicability constraint}{

This variable is only applicable under the following condition:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{level != 'local'
}\if{html}{\out{</div>}}
}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{ratio} level.
}

}

\subsection{\code{turnout}}{

The voter turnout at the referendum.
\subsection{Origin}{

This variable is derived from \href{https://rdb.rpkg.dev/articles/codebook.html#votes-yes}{\code{votes_yes}}, \href{https://rdb.rpkg.dev/articles/codebook.html#votes-no}{\code{votes_no}}, \href{https://rdb.rpkg.dev/articles/codebook.html#votes-empty}{\code{votes_empty}}, \href{https://rdb.rpkg.dev/articles/codebook.html#votes-invalid}{\code{votes_invalid}} and \href{https://rdb.rpkg.dev/articles/codebook.html#electorate-total}{\code{electorate_total}} and can be generated on demand via \code{\link[=add_turnout]{add_turnout()}}.
}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{ratio} level.
}

}

\subsection{\code{lower_house_yes}}{

The number of parliamentarians that approved the referendum proposal in the corresponding parliamentary voting; in \href{https://en.wikipedia.org/wiki/Bicameralism}{bicameral legislatures}, this number includes only members of the \emph{lower house}, in \href{https://en.wikipedia.org/wiki/Unicameralism}{unicameral legislatures} all parliamentarians; currently only filled for Swiss referendums on the national level (\href{https://en.wikipedia.org/wiki/National_Council_(Switzerland)}{National Council}).
\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{ratio} level.
}

}

\subsection{\code{lower_house_no}}{

The number of parliamentarians that rejected the referendum proposal in the corresponding parliamentary voting; in \href{https://en.wikipedia.org/wiki/Bicameralism}{bicameral legislatures}, this number includes only members of the \emph{lower house}, in \href{https://en.wikipedia.org/wiki/Unicameralism}{unicameral legislatures} all parliamentarians; currently only filled for Swiss referendums on the national level (\href{https://en.wikipedia.org/wiki/National_Council_(Switzerland)}{National Council}).
\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{ratio} level.
}

}

\subsection{\code{lower_house_abstentions}}{

The number of parliamentarians that abstained from the corresponding parliamentary voting on the referendum proposal; in \href{https://en.wikipedia.org/wiki/Bicameralism}{bicameral legislatures}, this number includes only members of the \emph{lower house}, in \href{https://en.wikipedia.org/wiki/Unicameralism}{unicameral legislatures} all parliamentarians; currently only filled for Swiss referendums on the national level (\href{https://en.wikipedia.org/wiki/National_Council_(Switzerland)}{National Council}).
\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{ratio} level.
}

}

\subsection{\code{upper_house_yes}}{

The number of parliamentarians that approved the referendum proposal in the corresponding parliamentary voting; applies only to \href{https://en.wikipedia.org/wiki/Bicameralism}{bicameral legislatures} and includes only members of the \emph{upper house}; currently only filled for Swiss referendums on the national level (\href{https://en.wikipedia.org/wiki/Council_of_States_(Switzerland)}{Council of States}).
\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{ratio} level.
}

}

\subsection{\code{upper_house_no}}{

The number of parliamentarians that rejected the referendum proposal in the corresponding parliamentary voting; applies only to \href{https://en.wikipedia.org/wiki/Bicameralism}{bicameral legislatures} and includes only members of the \emph{upper house}; currently only filled for Swiss referendums on the national level (\href{https://en.wikipedia.org/wiki/Council_of_States_(Switzerland)}{Council of States}).
\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{ratio} level.
}

}

\subsection{\code{upper_house_abstentions}}{

The number of parliamentarians that abstained from the corresponding parliamentary voting on the referendum proposal; applies only to \href{https://en.wikipedia.org/wiki/Bicameralism}{bicameral legislatures} and includes only members of the \href{https://en.wikipedia.org/wiki/Upper_house}{upper house}; currently only filled for Swiss referendums on the national level (\href{https://en.wikipedia.org/wiki/Council_of_States_(Switzerland)}{Council of States}).
\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{ratio} level.
}

}

\subsection{\code{position_government}}{

The government's recommendation on how to vote in the referendum.
\subsection{Possible values}{\tabular{ll}{
   Value \tab Meaning \cr
   \code{"yes"} \tab The government recommends to approve the referendum proposal. \cr
   \code{"no"} \tab The government recommends to reject the referendum proposal. \cr
}

}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal (binary)} level.
}

}

\subsection{Topics}{

The political topics to classify referendums are organized on a three-tier hierarchy as follows:
\itemize{
\item state organisation
\itemize{
\item national identity
\item political system
\itemize{
\item constitution
\item procedure for constitutional reform
\item legislative procedure
\item electoral system
}
\item institutions
\itemize{
\item government, administration
\item parliament
\item courts
\item national bank
}
\item citizens' rights
\itemize{
\item citizens' initiatives
\item referendum
\item voting rights
}
\item federalism
\itemize{
\item territorial questions
\item intergovernmental relations
\item division of tasks
}
\item legal system
\itemize{
\item international law
\item fundamental rights
\item citizenship
\item private law
\item criminal law
\item data protection
}
}
\item foreign policy
\itemize{
\item position on foreign policy
\itemize{
\item neutrality
\item independence
\item good offices
}
\item European policy
\itemize{
\item EFTA
\item EU
\item EEA
\item other European organisations
}
\item international organisations
\itemize{
\item UN
\item other international organisations
}
\item development cooperation
\item bilateral treaties
\item foreign trade policy
\itemize{
\item export promotion
\item customs
}
\item diplomacy
\item Swiss abroad
}
\item security policy
\itemize{
\item public security
\itemize{
\item civil protection
\item state security
\item police
}
\item army
\itemize{
\item armed forces in general
\item military organisation
\item arms
\item military facilities
\item conscientious objection, civilian service
\item abolition of the armed forces
\item military training
\item international deployment
}
\item national economic supply
}
\item economy
\itemize{
\item economic policy
\itemize{
\item short-term economic policy
\item competition policy
\item structural policy
\item pricing policy
\item consumer protection
}
\item employment
\itemize{
\item employment conditions
\item working hours
\item industrial relations
\item employment policy
}
\item finance
\itemize{
\item monetary policy
\item banks, stock exchange, insurance
}
\item tourism and leisure
\itemize{
\item tourism
\item hospitality
\item lottery and gambling
}
}
\item agriculture
\itemize{
\item agricultural policy
\item livestock
\item crop production
\item forestry
\item fishing, hunting and pets
}
\item public finance
\itemize{
\item taxation
\itemize{
\item tax policy
\item tax system
\item direct taxation
\item indirect taxation
}
\item financial system
\item public expenditure
\item budget cuts and remediation measures
}
\item energy
\itemize{
\item energy policy
\item nuclear energy
\item hydro-electric power
\item alternative energy
\item oil and gas
}
\item transport and infrastructure
\itemize{
\item transport policy
\itemize{
\item urban transport
\item transit traffic
}
\item road transport
\itemize{
\item road construction
\item heavy traffic
}
\item rail transport
\itemize{
\item goods traffic
\item passenger traffic
}
\item air transport
\item shipping
\item post
\item telecommunications
}
\item environment and living space
\itemize{
\item soil
\itemize{
\item spatial planning
\item land law
}
\item housing
\itemize{
\item tenancy issues
\item building of residential housing, property ownership
}
\item environment
\itemize{
\item environmental policy
\item noise protection
\item air quality control
\item water pollution control
\item soil protection
\item waste
\item protection of nature and cultural heritage
\item animal protection
}
}
\item social policy
\itemize{
\item health
\itemize{
\item health policy
\item medical research and technology
\item medicines
\item addictive substances
\item reproductive medicine
}
\item social security
\itemize{
\item pension insurance
\item disability insurance
\item occupational benefits provision
\item health and accident insurance
\item maternity insurance
\item unemployment insurance
\item compensation for loss of earnings for persons on military service or civil protection duty
\item welfare
}
\item social groups
\itemize{
\item immigration policy
\item refugees
\item status of women
\item family policy
\item children and young people
\item senior citizens
\item persons with disabilities
\item homosexuals
}
}
\item education and research
\itemize{
\item education policy
\item schools
\item universities
\item research
\itemize{
\item genetic engineering
\item animal testing
}
\item professional and vocational education
}
\item culture, religion and media
\itemize{
\item cultural policy
\item language policy
\item religion, churches
\item sport
\item media and communication
\itemize{
\item media policy
\item press
\item radio, television and electronic media
\item freedom of the media
}
}
}

The hierarchical relations between the three topic variables \code{topics_tier_1}, \code{topics_tier_2} and \code{topics_tier_3} can be reconstructed at any time using \code{\link[=hierarchize_topics]{hierarchize_topics()}}. This function can also be used to simply determine the parent topic(s) of any topic.
\subsection{\code{topics_tier_1}}{

A maximum of 3 predefined top-tier political topics that best classify the referendum proposal.
}

\subsection{Possible values}{\tabular{l}{
   Value \cr
   \code{"state organisation"} \cr
   \code{"foreign policy"} \cr
   \code{"security policy"} \cr
   \code{"economy"} \cr
   \code{"agriculture"} \cr
   \code{"public finance"} \cr
   \code{"energy"} \cr
   \code{"transport and infrastructure"} \cr
   \code{"environment and living space"} \cr
   \code{"social policy"} \cr
   \code{"education and research"} \cr
   \code{"culture, religion and media"} \cr
}


\emph{This is a multi-value variable and thus of type \code{list}. When unnested via \href{https://rdb.rpkg.dev/reference/unnest_var.html}{\code{unnest_var()}}, its name becomes \code{{variable_name_unnested}}.}
}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal} level.
}

\subsection{\code{topics_tier_2}}{

A maximum of 3 predefined second-tier political topics that best classify the referendum proposal; the presence of a second-tier topic always implies presence of its parent top-tier topic.
}

\subsection{Possible values}{\tabular{l}{
   Value \cr
   \code{"national identity"} \cr
   \code{"political system"} \cr
   \code{"institutions"} \cr
   \code{"citizens' rights"} \cr
   \code{"federalism"} \cr
   \code{"legal system"} \cr
   \code{"position on foreign policy"} \cr
   \code{"European policy"} \cr
   \code{"international organisations"} \cr
   \code{"development cooperation"} \cr
   \code{"bilateral treaties"} \cr
   \code{"foreign trade policy"} \cr
   \code{"diplomacy"} \cr
   \code{"Swiss abroad"} \cr
   \code{"public security"} \cr
   \code{"army"} \cr
   \code{"national economic supply"} \cr
   \code{"economic policy"} \cr
   \code{"employment"} \cr
   \code{"finance"} \cr
   \code{"tourism and leisure"} \cr
   \code{"agricultural policy"} \cr
   \code{"livestock"} \cr
   \code{"crop production"} \cr
   \code{"forestry"} \cr
   \code{"fishing, hunting and pets"} \cr
   \code{"taxation"} \cr
   \code{"financial system"} \cr
   \code{"public expenditure"} \cr
   \code{"budget cuts and remediation measures"} \cr
   \code{"energy policy"} \cr
   \code{"nuclear energy"} \cr
   \code{"hydro-electric power"} \cr
   \code{"alternative energy"} \cr
   \code{"oil and gas"} \cr
   \code{"transport policy"} \cr
   \code{"road transport"} \cr
   \code{"rail transport"} \cr
   \code{"air transport"} \cr
   \code{"shipping"} \cr
   \code{"post"} \cr
   \code{"telecommunications"} \cr
   \code{"soil"} \cr
   \code{"housing"} \cr
   \code{"environment"} \cr
   \code{"health"} \cr
   \code{"social security"} \cr
   \code{"social groups"} \cr
   \code{"education policy"} \cr
   \code{"schools"} \cr
   \code{"universities"} \cr
   \code{"research"} \cr
   \code{"professional and vocational education"} \cr
   \code{"cultural policy"} \cr
   \code{"language policy"} \cr
   \code{"religion, churches"} \cr
   \code{"sport"} \cr
   \code{"media and communication"} \cr
}


\emph{This is a multi-value variable and thus of type \code{list}. When unnested via \href{https://rdb.rpkg.dev/reference/unnest_var.html}{\code{unnest_var()}}, its name becomes \code{{variable_name_unnested}}.}
}

\subsection{Applicability constraint}{

This variable is only applicable under the following condition:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{purrr::map2_lgl(.x = topics_tier_1, .y = topics_tier_2, .f = \\(x, y) all(rdb::infer_topics(topics = y, tier = 1L) \%in\% x))
}\if{html}{\out{</div>}}
}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal} level.
}

\subsection{\code{topics_tier_3}}{

A maximum of 3 predefined third-tier political topics that best classify the referendum proposal; the presence of a third-tier topic always implies presence of its parent second-tier topic and grandparent top-tier topic.
}

\subsection{Possible values}{\tabular{l}{
   Value \cr
   \code{"constitution"} \cr
   \code{"procedure for constitutional reform"} \cr
   \code{"legislative procedure"} \cr
   \code{"electoral system"} \cr
   \code{"government, administration"} \cr
   \code{"parliament"} \cr
   \code{"courts"} \cr
   \code{"national bank"} \cr
   \code{"citizens' initiatives"} \cr
   \code{"referendum"} \cr
   \code{"voting rights"} \cr
   \code{"territorial questions"} \cr
   \code{"intergovernmental relations"} \cr
   \code{"division of tasks"} \cr
   \code{"international law"} \cr
   \code{"fundamental rights"} \cr
   \code{"citizenship"} \cr
   \code{"private law"} \cr
   \code{"criminal law"} \cr
   \code{"data protection"} \cr
   \code{"neutrality"} \cr
   \code{"independence"} \cr
   \code{"good offices"} \cr
   \code{"EFTA"} \cr
   \code{"EU"} \cr
   \code{"EEA"} \cr
   \code{"other European organisations"} \cr
   \code{"UN"} \cr
   \code{"other international organisations"} \cr
   \code{"export promotion"} \cr
   \code{"customs"} \cr
   \code{"civil protection"} \cr
   \code{"state security"} \cr
   \code{"police"} \cr
   \code{"armed forces in general"} \cr
   \code{"military organisation"} \cr
   \code{"arms"} \cr
   \code{"military facilities"} \cr
   \code{"conscientious objection, civilian service"} \cr
   \code{"abolition of the armed forces"} \cr
   \code{"military training"} \cr
   \code{"international deployment"} \cr
   \code{"short-term economic policy"} \cr
   \code{"competition policy"} \cr
   \code{"structural policy"} \cr
   \code{"pricing policy"} \cr
   \code{"consumer protection"} \cr
   \code{"employment conditions"} \cr
   \code{"working hours"} \cr
   \code{"industrial relations"} \cr
   \code{"employment policy"} \cr
   \code{"monetary policy"} \cr
   \code{"banks, stock exchange, insurance"} \cr
   \code{"tourism"} \cr
   \code{"hospitality"} \cr
   \code{"lottery and gambling"} \cr
   \code{"tax policy"} \cr
   \code{"tax system"} \cr
   \code{"direct taxation"} \cr
   \code{"indirect taxation"} \cr
   \code{"urban transport"} \cr
   \code{"transit traffic"} \cr
   \code{"road construction"} \cr
   \code{"heavy traffic"} \cr
   \code{"goods traffic"} \cr
   \code{"passenger traffic"} \cr
   \code{"spatial planning"} \cr
   \code{"land law"} \cr
   \code{"tenancy issues"} \cr
   \code{"building of residential housing, property ownership"} \cr
   \code{"environmental policy"} \cr
   \code{"noise protection"} \cr
   \code{"air quality control"} \cr
   \code{"water pollution control"} \cr
   \code{"soil protection"} \cr
   \code{"waste"} \cr
   \code{"protection of nature and cultural heritage"} \cr
   \code{"animal protection"} \cr
   \code{"health policy"} \cr
   \code{"medical research and technology"} \cr
   \code{"medicines"} \cr
   \code{"addictive substances"} \cr
   \code{"reproductive medicine"} \cr
   \code{"pension insurance"} \cr
   \code{"disability insurance"} \cr
   \code{"occupational benefits provision"} \cr
   \code{"health and accident insurance"} \cr
   \code{"maternity insurance"} \cr
   \code{"unemployment insurance"} \cr
   \code{"compensation for loss of earnings for persons on military service or civil protection duty"} \cr
   \code{"welfare"} \cr
   \code{"immigration policy"} \cr
   \code{"refugees"} \cr
   \code{"status of women"} \cr
   \code{"family policy"} \cr
   \code{"children and young people"} \cr
   \code{"senior citizens"} \cr
   \code{"persons with disabilities"} \cr
   \code{"homosexuals"} \cr
   \code{"genetic engineering"} \cr
   \code{"animal testing"} \cr
   \code{"media policy"} \cr
   \code{"press"} \cr
   \code{"radio, television and electronic media"} \cr
   \code{"freedom of the media"} \cr
}


\emph{This is a multi-value variable and thus of type \code{list}. When unnested via \href{https://rdb.rpkg.dev/reference/unnest_var.html}{\code{unnest_var()}}, its name becomes \code{{variable_name_unnested}}.}
}

\subsection{Applicability constraint}{

This variable is only applicable under the following condition:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{purrr::map2_lgl(.x = topics_tier_2, .y = topics_tier_3, .f = \\(x, y) all(rdb::infer_topics(topics = y, tier = 2L) \%in\% x)) & purrr::map2_lgl(.x = topics_tier_1, .y = topics_tier_3, .f = \\(x, y) all(rdb::infer_topics(topics = y, tier = 1L) \%in\% x))
}\if{html}{\out{</div>}}
}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal} level.
}

}

\subsection{\code{remarks}}{

A free text field with additional information about the referendum.
\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal} level.
}

}

\subsection{\code{files}}{

Metadata about files attached to the referendum, including
\itemize{
\item \code{date_time_attached}: the date and time on which the file was attached to the referendum's data entry
\item \code{is_deleted}: whether or not the file has been deleted since it was attached
\item \code{file_name}: the filename as it appears when the file is downloaded
\item \code{file_size}: the file's size in bytes
\item \code{s3_object_key}: the key uniquely identifying the file in the RDB \href{https://en.wikipedia.org/wiki/Amazon_S3#Design}{Amazon S3 bucket}; the file can be downloaded by feeding this key to \href{https://rdb.rpkg.dev/reference/download_file_attachment.html}{\code{rdb::download_file_attachment()}}.
}
}

\subsection{\code{url_sudd}}{

The URL of the corresponding entry on \href{https://sudd.ch/}{sudd.ch}.
\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal} level.
}

}

\subsection{\code{url_swissvotes}}{

The URL of the corresponding entry on \href{https://swissvotes.ch/}{swissvotes.ch}.
\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal} level.
}

}

\subsection{\code{sources}}{

Information about the source(s) of the referendum data like additional URLs or other references.
\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal} level.
}

}

\subsection{\code{is_draft}}{

Whether or not the entry is marked as a draft in the database.
\subsection{Possible values}{\tabular{ll}{
   Value \tab Meaning \cr
   \code{TRUE} \tab The entry is marked as a draft. \cr
   \code{FALSE} \tab The entry is not marked as a draft. \cr
}

}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal (binary)} level.
}

}

\subsection{\code{created_at}}{

The date and time on which the referendum's data entry was created.
\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{interval} level.
}

}

\subsection{\code{updated_at}}{

The date and time on which the referendum's data entry was last modified.
\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{interval} level.
}

}

\subsection{\code{archive}}{

The referendum's data from an earlier, obsolete RDB state.
}

\subsection{\code{type}}{

The institutional type of direct democracy which the referendum is an instance of.
\subsection{Possible values}{\tabular{ll}{
   Value \tab Meaning \cr
   \code{"mandatory referendum"} \tab A referendum that was automatically triggered by certain legal conditions, usually found in the constitution. \cr
   \code{"optional referendum"} \tab A referendum against a law passed by the parliament that came about because the quorum of citizen opposition was met (usually a certain number of signatures) within a specified period after the resolution was passed. \cr
   \code{"counter proposal"} \tab A counter proposal by the government or parliament to another referendum. \cr
   \code{"citizens' initiative"} \tab A referendum launched by citizens, usually via a signature collection. \cr
   \code{"governmental referendum"} \tab A referendum launched by the government. \cr
   \code{"citizens' assembly"} \tab A citizens' assembly, usually in person where people vote by raising their hands. \cr
   \code{"recall"} \tab A recall of all members of a political institution/body. \cr
}

}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal} level.
}

}

}

\subsection{Institution-level variables}{

For the following variables, the \href{https://en.wikipedia.org/wiki/Unit_of_observation}{unit of observation} is the referendum's \emph{institutional instrument}, a
combination of the referendum's \code{type} and the jurisdiction it took place.

The set of variables that constitute the \emph{geographical} component of the jurisdiction is dependent on the referendum's \code{level}:\tabular{ll}{
   \code{level} \tab Geographical jurisdiction is given by \cr
   \code{national} \tab \code{country_code} \cr
   \code{subnational} \tab \code{country_code} + \code{subnational_entity_name} \cr
   \code{municipal} \tab \code{country_code} + \code{subnational_entity_name} + \code{municipality} \cr
}


The actual jurisdiction is also dependent on a \emph{temporal} component since the legal setting in a geographical jurisdiction can vary over time. Thus, the actual
jurisdicion is dependent on the geographical jurisdiction plus the \code{date} a referendum took place (more specifically the date \emph{range} in between a
specific legal setting is in force).
\subsection{\code{inst_legal_basis_type}}{

The type of the legal basis of the referendum \code{type}.
\subsection{Possible values}{\tabular{ll}{
   Value \tab Meaning \cr
   \code{"non-official"} \tab The referendum \code{type} has no legal basis. \cr
   \code{"ad-hoc"} \tab The referendum \code{type} has a legal basis which was specifically created for it. \cr
   \code{"official"} \tab The referendum \code{type} has a legal basis that wasn’t specifically created for it. \cr
}

}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{ordinal (ascending)} level.
}

}

\subsection{\code{inst_has_urgent_legal_basis}}{

Whether or not the legal basis of the referendum \code{type} was adopted urgently, i.e. in an abridged procedure.
\subsection{Possible values}{\tabular{ll}{
   Value \tab Meaning \cr
   \code{TRUE} \tab The legal basis of the referendum \code{type} was adopted in an abridged procedure (urgent). \cr
   \code{FALSE} \tab The legal basis of the referendum \code{type} was adopted in a normal procedure. \cr
}

}

\subsection{Applicability constraint}{

This variable is only applicable under the following condition:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{inst_trigger_actor \%in\% c('government', 'president', 'monarch', 'parliament')
}\if{html}{\out{</div>}}
}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal (binary)} level.
}

}

\subsection{\code{inst_is_binding}}{

Whether the result of the referendum \code{type} has binding force or the decision of the voters may be disregarded.
\subsection{Possible values}{\tabular{ll}{
   Value \tab Meaning \cr
   \code{TRUE} \tab The result of the referendum \code{type} is binding for the legislator. \cr
   \code{FALSE} \tab The result of the referendum \code{type} is only consultative, i.e. the legislator does not have to follow the decision of the voters. \cr
}

}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal (binary)} level.
}

}

\subsection{\code{inst_is_counter_proposal}}{

Whether or not the referendum \code{type} is a counterproposal to a citizen’s (or other) initiative.
\subsection{Possible values}{\tabular{ll}{
   Value \tab Meaning \cr
   \code{TRUE} \tab The referendum \code{type} is a counterproposal to a citizen’s (or other) initiative. \cr
   \code{FALSE} \tab The referendum \code{type} is not a counterproposal to a citizen’s (or other) initiative. \cr
}

}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal (binary)} level.
}

}

\subsection{\code{inst_is_assembly}}{

Whether or not the referendum \code{type} is held as an assembly (all voters at the same time and place; physical presence at the assembly is required to be able to vote; usually voting means raising one’s hand).
\subsection{Possible values}{\tabular{ll}{
   Value \tab Meaning \cr
   \code{TRUE} \tab The referendum \code{type} is a counterproposal to a citizen’s (or other) initiative. \cr
   \code{FALSE} \tab The referendum \code{type} is not a counterproposal to a citizen’s (or other) initiative. \cr
}

}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal (binary)} level.
}

}

\subsection{\code{inst_trigger_type}}{

The way the referendum \code{type} is triggered; completely dependent on \code{inst_trigger_actor}.
\subsection{Possible values}{\tabular{ll}{
   Value \tab Meaning \cr
   \code{"automatic"} \tab The referendum \code{type} is triggered by a constitutional/legal requirement. \cr
   \code{"top down"} \tab The referendum \code{type} is triggered by an institution of the political elite like the monarch/president/government, the parliament, a territorial unit, the UN or another institution. \cr
   \code{"bottom up"} \tab The referendum \code{type} is triggered by citizen demand (e.g. a signature collection). \cr
}

}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal} level.
}

}

\subsection{\code{inst_trigger_actor}}{

The actor that triggers the referendum \code{type}.
\subsection{Possible values}{\tabular{ll}{
   Value \tab Meaning \cr
   \code{"citizens"} \tab The referendum \code{type} is triggered by a direct citizen demand (usually a signature collection). \cr
   \code{"constitution"} \tab The referendum \code{type} is triggered by a constitutional requirement. \cr
   \code{"law"} \tab The referendum \code{type} is triggered by a legal requirement. \cr
   \code{"parliament"} \tab The referendum \code{type} is triggered by the parliament. \cr
   \code{"government"} \tab The referendum \code{type} is triggered by the government or prime minister. \cr
   \code{"president"} \tab The referendum \code{type} is triggered by the president. \cr
   \code{"monarch"} \tab The referendum \code{type} is triggered by the monarch. \cr
   \code{"territorial unit"} \tab The referendum \code{type} is triggered by a province or municipality. \cr
   \code{"UN"} \tab The referendum \code{type} is triggered by the United Nations (UN) or one of its organs. \cr
   \code{"other institution"} \tab The referendum \code{type} is triggered by another institution such as the church or a state agency. \cr
}

}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal} level.
}

}

\subsection{\code{inst_trigger_actor_level}}{

The political \code{level} of \code{inst_trigger_actor} in comparison to \code{level}.
\subsection{Possible values}{\tabular{ll}{
   Value \tab Meaning \cr
   \code{"lower"} \tab A lower political level triggers a referendum on one of the higher political levels. \cr
   \code{"equal"} \tab The trigger actor and the referendum are on the same political level. \cr
   \code{"higher"} \tab A higher political level triggers a referendum on one of the lower political levels. \cr
}

}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{ordinal (ascending)} level.
}

}

\subsection{\code{inst_trigger_threshold}}{

The threshold of support among \code{inst_trigger_actor} that has to be reached in order to trigger the referendum \code{type}.
\subsection{Applicability constraint}{

This variable is only applicable under the following condition:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{inst_trigger_actor \%in\% c('citizens', 'parliament', 'territorial unit')
}\if{html}{\out{</div>}}
}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal} level.
}

}

\subsection{\code{inst_trigger_time_limit}}{

The time limit in which \code{inst_trigger_threshold} (usually a minimum number of signatures) has to be reached in order to trigger the referendum \code{type}.
\subsection{Applicability constraint}{

This variable is only applicable under the following condition:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{inst_trigger_actor \%in\% c('citizens', 'territorial unit')
}\if{html}{\out{</div>}}
}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal} level.
}

}

\subsection{\code{inst_object_type}}{

The type of the object the referendum \code{type} refers to.
\subsection{Possible values}{\tabular{ll}{
   Value \tab Meaning \cr
   \code{"question"} \tab A question which can be answered with yes or no (used as fallback option if nothing else applies). \cr
   \code{"legal text (general proposal)"} \tab A general proposal of a legal text, meaning the proposal is not formulated in a definite way, and therefore, if approved, needs to be implemented and concretized in a further legislative process. \cr
   \code{"legal text (formulated proposal)"} \tab A precisely formulated legal text which, if approved, will directly enter into force. \cr
   \code{"institution"} \tab A recall of all members of a particular institution/body. \cr
   \code{"person"} \tab An appointment of a single person. \cr
   \code{"principle"} \tab A question about what direction to take on a particular issue/topic, but without a proposed legal text; e.g., "Should there be a total revision of our constitution?". \cr
}

}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal} level.
}

}

\subsection{\code{inst_object_author}}{

The actor that defines the text/content of \code{inst_object_type}.
\subsection{Possible values}{\tabular{ll}{
   Value \tab Meaning \cr
   \code{"citizens"} \tab Citizens (usually a small group initiating a citizen demand). \cr
   \code{"parliament"} \tab The parliament. \cr
   \code{"government"} \tab The government or prime minister. \cr
   \code{"president"} \tab The president. \cr
   \code{"monarch"} \tab The monarch. \cr
   \code{"territorial unit"} \tab A province or municipality. \cr
   \code{"UN"} \tab The United Nations (UN) or one of its organs. \cr
   \code{"other institution"} \tab Another institution such as the church or a state agency. \cr
}

}

\subsection{Applicability constraint}{

This variable is only applicable under the following condition:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{inst_object_type \%in\% c('question', 'legal text (general proposal)', 'legal text (formulated proposal)', 'principle')
}\if{html}{\out{</div>}}
}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal} level.
}

}

\subsection{\code{inst_object_legal_level}}{

The legal level of \code{inst_object_type}.
\subsection{Possible values}{\tabular{ll}{
   Value \tab Meaning \cr
   \code{"ordinance"} \tab An ordinance, i.e. a legal norm based on a law or the constitution. \cr
   \code{"treaty"} \tab A domestic treaty, e.g. an agreement between the constituent states of a republic. \cr
   \code{"law"} \tab A law, i.e. a legal norm based on the constitution and created by the legislative. \cr
   \code{"constitution"} \tab The constitution, i.e. the aggregate of fundamental principles or established precedents that constitute the legal basis of a state and commonly determine how that state is to be governed. \cr
   \code{"international treaty"} \tab An international treaty, i.e. a bilateral or multilateral agreement or an accession to and international organization. \cr
}

}

\subsection{Applicability constraint}{

This variable is only applicable under the following condition:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{inst_object_type \%in\% c('legal text (general proposal)', 'legal text (formulated proposal)')
}\if{html}{\out{</div>}}
}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal} level.
}

}

\subsection{\code{inst_object_revision_extent}}{

The extent of revision of \code{inst_object_type} which can be achieved within one referendum.
\subsection{Possible values}{\tabular{ll}{
   Value \tab Meaning \cr
   \code{"partial"} \tab Only parts of the concerned legal text can be revised at the same time. \cr
   \code{"total"} \tab Only the entirety of the legal text can be changed at the same time. \cr
   \code{"both"} \tab Both, parts and the entirety of the legal text, can be changed at the same time. \cr
}

}

\subsection{Applicability constraint}{

This variable is only applicable under the following condition:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{inst_object_type \%in\% c('legal text (general proposal)', 'legal text (formulated proposal)')
}\if{html}{\out{</div>}}
}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal} level.
}

}

\subsection{\code{inst_object_revision_modes}}{

The mode(s) of revision that \code{inst_object_type} allows.
\subsection{Possible values}{\tabular{ll}{
   Value \tab Meaning \cr
   \code{"introduce"} \tab With this instrument, it is possible to introduce an entirely new text/concept. \cr
   \code{"revise"} \tab With this instrument, it is possible to change some parts of the legal text/... \cr
   \code{"abrogate"} \tab With this instrument, it is possible to strike parts of the legal text / to recall a person/institution. \cr
}


\emph{This is a multi-value variable and thus of type \code{list}. When unnested via \href{https://rdb.rpkg.dev/reference/unnest_var.html}{\code{unnest_var()}}, its name becomes \code{{variable_name_unnested}}.}
}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal} level.
}

}

\subsection{\code{inst_is_divisible}}{

Whether or not \code{inst_object_type} (the proposed legal text) may be split up so voters can decide upon these parts separately.
\subsection{Possible values}{\tabular{ll}{
   Value \tab Meaning \cr
   \code{TRUE} \tab The proposed legal text may be split up. \cr
   \code{FALSE} \tab The proposed legal text may not be split up. \cr
}

}

\subsection{Applicability constraint}{

This variable is only applicable under the following condition:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{inst_object_type \%in\% c('question', 'legal text (general proposal)', 'legal text (formulated proposal)')
}\if{html}{\out{</div>}}
}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal (binary)} level.
}

}

\subsection{\code{inst_is_variable}}{

Whether or not the proposed legal text may be presented in two or more (alternative) variants for voters to choose amongst.
\subsection{Possible values}{\tabular{ll}{
   Value \tab Meaning \cr
   \code{TRUE} \tab The proposed legal text may be presented in variants. \cr
   \code{FALSE} \tab The proposed legal text may not be presented in variants. \cr
}

}

\subsection{Applicability constraint}{

This variable is only applicable under the following condition:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{inst_object_type \%in\% c('question', 'legal text (general proposal)', 'legal text (formulated proposal)')
}\if{html}{\out{</div>}}
}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal (binary)} level.
}

}

\subsection{\code{inst_topics_only}}{

The legal topic(s) to which the referendum \code{type} is limited.
\subsection{Possible values}{\tabular{ll}{
   Value \tab Meaning \cr
   \code{"infrastructural acts"} \tab The referendum \code{type} is limited to TODO. \cr
   \code{"territorial questions"} \tab The referendum \code{type} is limited to TODO. \cr
   \code{"competence shifts"} \tab The referendum \code{type} is limited to TODO. \cr
   \code{"financial acts"} \tab The referendum \code{type} is limited to TODO. \cr
   \code{"financial acts (expenses)"} \tab The referendum \code{type} is limited to TODO. \cr
   \code{"financial acts (taxes)"} \tab The referendum \code{type} is limited to TODO. \cr
   \code{"financial acts (obligations)"} \tab The referendum \code{type} is limited to TODO. \cr
   \code{"treaties"} \tab The referendum \code{type} is limited to TODO (automatically limited to if \code{inst_object_legal_level == "law"}). \cr
   \code{"total revisions of the constitution"} \tab The referendum \code{type} is limited to TODO. \cr
}


\emph{This is a multi-value variable and thus of type \code{list}. When unnested via \href{https://rdb.rpkg.dev/reference/unnest_var.html}{\code{unnest_var()}}, its name becomes \code{{variable_name_unnested}}.}
}

\subsection{Applicability constraint}{

This variable is only applicable under the following condition:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{inst_topics_excluded \%>\% purrr::map_lgl(rlang::is_empty)
}\if{html}{\out{</div>}}
}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal} level.
}

}

\subsection{\code{inst_topics_excluded}}{

The legal topic(s) that are excluded from the referendum \code{type}.
\subsection{Possible values}{\tabular{ll}{
   Value \tab Meaning \cr
   \code{"jus cogens"} \tab TODO is exluded from the referendum \code{type}. \cr
   \code{"international treaties"} \tab TODO is exluded from the referendum \code{type}. \cr
   \code{"pardons"} \tab TODO is exluded from the referendum \code{type}. \cr
   \code{"personnel appointments"} \tab TODO is exluded from the referendum \code{type}. \cr
   \code{"budgets"} \tab TODO is exluded from the referendum \code{type}. \cr
   \code{"financial matters"} \tab TODO is exluded from the referendum \code{type}. \cr
   \code{"everything within parliamentary competence"} \tab TODO is exluded from the referendum \code{type}. \cr
}


\emph{This is a multi-value variable and thus of type \code{list}. When unnested via \href{https://rdb.rpkg.dev/reference/unnest_var.html}{\code{unnest_var()}}, its name becomes \code{{variable_name_unnested}}.}
}

\subsection{Applicability constraint}{

This variable is only applicable under the following condition:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{inst_topics_only \%>\% purrr::map_lgl(rlang::is_empty)
}\if{html}{\out{</div>}}
}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal} level.
}

}

\subsection{\code{inst_quorum_turnout}}{

The minimum percentage of the \code{electorate_total} that have to participate in the referendum \code{type} in order for it to be passed.
\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal} level.
}

}

\subsection{\code{inst_quorum_approval}}{

The minimum number of voters that have to approve the referendum \code{type} in order for it to be passed.
\subsection{Possible values}{\tabular{ll}{
   Value \tab Meaning \cr
   \code{"simple majority"} \tab > 50 \% of the voters have to approve the referendum proposal. \cr
   \code{"qualified majority"} \tab A percentage higher than > 50 \% of the voters has to approve the referendum proposal. \cr
   \code{"double majority"} \tab Two simultaneous quorums of > 50 \% or higher of the voters apply. \cr
}

}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal} level.
}

}

\subsection{\code{inst_has_precondition}}{

Whether or not an institutional actor must take a preconditional decision in order for the referendum \code{type} to be held.
\subsection{Possible values}{\tabular{ll}{
   Value \tab Meaning \cr
   \code{TRUE} \tab An institutional actor must take a preconditional decision in order for the referendum \code{type} to be held. \cr
   \code{FALSE} \tab No institutional actor must take a preconditional decision in order for the referendum \code{type} to be held. \cr
}

}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal (binary)} level.
}

}

\subsection{\code{inst_precondition_actor}}{

The actor that must take a preconditional decision in order for the referendum \code{type} to be held.
\subsection{Possible values}{\tabular{ll}{
   Value \tab Meaning \cr
   \code{"parliament"} \tab The parliament. \cr
   \code{"government"} \tab The government or prime minister. \cr
   \code{"president"} \tab The president. \cr
   \code{"UN"} \tab The United Nations (UN) or one of its organs. \cr
   \code{"parliament and government"} \tab The parliament and the government or prime minister. \cr
   \code{"parliament and president"} \tab The parliament and the president. \cr
   \code{"central government and SPLM"} \tab The central government and the \href{https://en.wikipedia.org/wiki/Sudan_People's_Liberation_Movement}{Sudan People's Liberation Movement}. \cr
}

}

\subsection{Applicability constraint}{

This variable is only applicable under the following condition:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{isTRUE(inst_has_precondition)
}\if{html}{\out{</div>}}
}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal} level.
}

}

\subsection{\code{inst_precondition_decision}}{

The decision that \code{inst_precondition_actor} has taken.
\subsection{Possible values}{\tabular{ll}{
   Value \tab Meaning \cr
   \code{"yes"} \tab Affirmative answer (e.g. "Yes, we will hold a vote that has been requested by citizens."). \cr
   \code{"no"} \tab Negative answer (e.g. "No, we will not implement the demands of a citizens’ initiative voluntarily."). \cr
}

}

\subsection{Level of measurement}{

This variable's \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the \strong{nominal (binary)} level.
}

}

}
}

\examples{
rdb::data_codebook
}
\seealso{
Other referendum metadata functions:
\code{\link{prettify_var_names}()},
\code{\link{val_lbls}()},
\code{\link{val_scale}()},
\code{\link{var_name_unnested}()},
\code{\link{var_vals}()}
}
\concept{metadata}
\keyword{datasets}
