% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d.gen.R
\docType{data}
\name{data_codebook}
\alias{data_codebook}
\alias{codebook}
\title{C2D Codebook}
\format{
A \link[tibble:tbl_df-class]{tibble}.
}
\usage{
data_codebook
}
\description{
A tibble containing the complete metadata of all \code{\link[=referendums]{referendums()}} variables. The Codebook below is also available
\href{https://rpkg.dev/c2d/articles/codebook.html}{online}.
}
\details{
\subsection{Referendum-level variables}{

For the following variables, the \href{https://en.wikipedia.org/wiki/Unit_of_observation}{unit of observation} is the
individual referendum.
\subsection{\code{id}}{

A 24-character value that uniquely identifies the referendum and
corresponds to the \href{https://docs.mongodb.com/manual/reference/bson-types/#std-label-objectid}{MongoDB \code{ObjectId}}
in the C2D database.
\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{nominal} level.
}

}

\subsection{\code{id_official}}{

The referendum’s official identifier assigned by the authorities; this
is only guaranteed to be unique per respective jurisdiction.
\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{nominal} level.
}

}

\subsection{\code{id_sudd}}{

The referendum’s identifier assigned by \href{https://sudd.ch/}{sudd.ch};
multiple referendum’s might share the same identifier.
\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{nominal} level.
}

}

\subsection{\code{country_code}}{

The \href{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2}{ISO 3166-1 alpha-2} code
identifying the country in which the referendum took place.
\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{nominal} level.
}

}

\subsection{\code{country_name}}{

The English name of the country in which the referendum took place.
\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{nominal} level.
}

}

\subsection{\code{subnational_entity_name}}{

The English name of the \href{https://en.wikipedia.org/wiki/Administrative_division}{subnational political entity} in which
the referendum took place.
\subsection{Applicability constraint}{

This variable is only applicable under the following condition:\if{html}{\out{<div class="sourceCode r">}}\preformatted{level \%in\% c('local', 'subnational')
}\if{html}{\out{</div>}}
}

\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{nominal} level.
}

}

\subsection{\code{municipality}}{

The municipality in which the referendum took place.
\subsection{Applicability constraint}{

This variable is only applicable under the following condition:\if{html}{\out{<div class="sourceCode r">}}\preformatted{level == 'local'
}\if{html}{\out{</div>}}
}

\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{nominal} level.
}

}

\subsection{\code{level}}{

The political level at which the referendum took place.
\subsection{Possible values}{\tabular{ll}{
   Value \tab Meaning \cr
   \code{"local"} \tab The referendum took place at the local political level. \cr
   \code{"subnational"} \tab The referendum took place at the subnational political level. \cr
   \code{"national"} \tab The referendum took place at the national political level. \cr
}

}

\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{ordinal (ascending)} level.
}

}

\subsection{\code{date}}{

The date on which the referendum took place.
\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{interval} level.
}

}

\subsection{\code{title_de}}{

The referendum’s German title.
\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{nominal} level.
}

}

\subsection{\code{title_fr}}{

The referendum’s French title.
\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{nominal} level.
}

}

\subsection{\code{title_en}}{

The referendum’s English title.
\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{nominal} level.
}

}

\subsection{\code{committee_name}}{

The name of the committee that initiated the referendum.
\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{nominal} level.
}

}

\subsection{\code{result}}{

The result of the referendum.
\subsection{Possible values}{\tabular{ll}{
   Value \tab Meaning \cr
   \code{"yes"} \tab The referendum proposal was approved. \cr
   \code{"no"} \tab The referendum proposal was rejected. \cr
}

}

\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{nominal (binary)} level.
}

}

\subsection{\code{subterritories_yes}}{

The number of \href{https://en.wikipedia.org/wiki/Administrative_division}{administrative divisions} one
level below \code{level} that approved the referendum.
\subsection{Applicability constraint}{

This variable is only applicable under the following condition:\if{html}{\out{<div class="sourceCode r">}}\preformatted{level != 'local'
}\if{html}{\out{</div>}}
}

\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{ratio} level.
}

}

\subsection{\code{subterritories_no}}{

The number of \href{https://en.wikipedia.org/wiki/Administrative_division}{administrative divisions} one
level below \code{level} that rejected the referendum.
\subsection{Applicability constraint}{

This variable is only applicable under the following condition:\if{html}{\out{<div class="sourceCode r">}}\preformatted{level != 'local'
}\if{html}{\out{</div>}}
}

\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{ratio} level.
}

}

\subsection{\code{electorate_total}}{

The total number of people entitled to vote in the referendum.
\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{ratio} level.
}

}

\subsection{\code{electorate_abroad}}{

The number of people entitled to vote in the referendum \emph{who are living
abroad}; by definition a subset of \code{electorate_total}.
\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{ratio} level.
}

}

\subsection{\code{votes_yes}}{

The number of votes approving the referendum proposal.
\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{ratio} level.
}

}

\subsection{\code{votes_no}}{

The number of votes rejecting the referendum proposal.
\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{ratio} level.
}

}

\subsection{\code{votes_empty}}{

The number of votes that did not specify any preference on the
referendum proposal.
\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{ratio} level.
}

}

\subsection{\code{votes_invalid}}{

The number of votes that were invalid.
\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{ratio} level.
}

}

\subsection{\code{votes_per_subterritory}}{

The number of \code{votes_yes}, \code{votes_no}, \code{votes_empty} and \code{votes_invalid}
per \href{https://en.wikipedia.org/wiki/Administrative_division}{administrative division} one
level below \code{level}; currently only filled for Swiss referendums on the
national level.
\subsection{Applicability constraint}{

This variable is only applicable under the following condition:\if{html}{\out{<div class="sourceCode r">}}\preformatted{level != 'local'
}\if{html}{\out{</div>}}
}

}

\subsection{\code{lower_house_yes}}{

The number of parliamentarians that approved the referendum proposal in
the corresponding parliamentary voting; in \href{https://en.wikipedia.org/wiki/Bicameralism}{bicameral legislatures}, this number
includes only members of the \emph{lower house}, in \href{https://en.wikipedia.org/wiki/Unicameralism}{unicameral legislatures} all
parliamentarians; currently only filled for Swiss referendums on the
national level (\href{https://en.wikipedia.org/wiki/National_Council_(Switzerland)}{National Council}).
\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{ratio} level.
}

}

\subsection{\code{lower_house_no}}{

The number of parliamentarians that rejected the referendum proposal in
the corresponding parliamentary voting; in \href{https://en.wikipedia.org/wiki/Bicameralism}{bicameral legislatures}, this number
includes only members of the \emph{lower house}, in \href{https://en.wikipedia.org/wiki/Unicameralism}{unicameral legislatures} all
parliamentarians; currently only filled for Swiss referendums on the
national level (\href{https://en.wikipedia.org/wiki/National_Council_(Switzerland)}{National Council}).
\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{ratio} level.
}

}

\subsection{\code{lower_house_abstentions}}{

The number of parliamentarians that abstained from the corresponding
parliamentary voting on the referendum proposal; in \href{https://en.wikipedia.org/wiki/Bicameralism}{bicameral legislatures}, this number
includes only members of the \emph{lower house}, in \href{https://en.wikipedia.org/wiki/Unicameralism}{unicameral legislatures} all
parliamentarians; currently only filled for Swiss referendums on the
national level (\href{https://en.wikipedia.org/wiki/National_Council_(Switzerland)}{National Council}).
\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{ratio} level.
}

}

\subsection{\code{upper_house_yes}}{

The number of parliamentarians that approved the referendum proposal in
the corresponding parliamentary voting; applies only to \href{https://en.wikipedia.org/wiki/Bicameralism}{bicameral legislatures} and includes
only members of the \emph{upper house}; currently only filled for Swiss
referendums on the national level (\href{https://en.wikipedia.org/wiki/Council_of_States_(Switzerland)}{Council of States}).
\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{ratio} level.
}

}

\subsection{\code{upper_house_no}}{

The number of parliamentarians that rejected the referendum proposal in
the corresponding parliamentary voting; applies only to \href{https://en.wikipedia.org/wiki/Bicameralism}{bicameral legislatures} and includes
only members of the \emph{upper house}; currently only filled for Swiss
referendums on the national level (\href{https://en.wikipedia.org/wiki/Council_of_States_(Switzerland)}{Council of States}).
\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{ratio} level.
}

}

\subsection{\code{upper_house_abstentions}}{

The number of parliamentarians that abstained from the corresponding
parliamentary voting on the referendum proposal; applies only to
\href{https://en.wikipedia.org/wiki/Bicameralism}{bicameral legislatures} and
includes only members of the \href{https://en.wikipedia.org/wiki/Upper_house}{upper house}; currently only filled
for Swiss referendums on the national level (\href{https://en.wikipedia.org/wiki/Council_of_States_(Switzerland)}{Council of States}).
\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{ratio} level.
}

}

\subsection{\code{position_government}}{

The government’s recommendation on how to vote in the referendum.
\subsection{Possible values}{\tabular{ll}{
   Value \tab Meaning \cr
   \code{"yes"} \tab The government recommends to approve the referendum. \cr
   \code{"no"} \tab The government recommends to reject the referendum. \cr
}

}

\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{nominal (binary)} level.
}

}

\subsection{Tags}{

The tags (political topics) are organized on a three-tier hierarchy as
follows:
\itemize{
\item \if{html}{\out{<span class="color-azure">}}state organisation\if{html}{\out{</span>}}
\itemize{
\item \if{html}{\out{<span class="color-violet">}}national identity\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-violet">}}political system\if{html}{\out{</span>}}
\itemize{
\item \if{html}{\out{<span class="color-rose">}}constitution\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}procedure for constitutional
reform\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}legislative procedure\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}electoral system\if{html}{\out{</span>}}
}
\item \if{html}{\out{<span class="color-violet">}}institutions\if{html}{\out{</span>}}
\itemize{
\item \if{html}{\out{<span class="color-rose">}}government, administration\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}parliament\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}courts\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}national bank\if{html}{\out{</span>}}
}
\item \if{html}{\out{<span class="color-violet">}}citizens’ rights\if{html}{\out{</span>}}
\itemize{
\item \if{html}{\out{<span class="color-rose">}}citizens’ initiatives\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}referendum\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}voting rights\if{html}{\out{</span>}}
}
\item \if{html}{\out{<span class="color-violet">}}federalism\if{html}{\out{</span>}}
\itemize{
\item \if{html}{\out{<span class="color-rose">}}territorial questions\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}intergovernmental relations\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}division of tasks\if{html}{\out{</span>}}
}
\item \if{html}{\out{<span class="color-violet">}}legal system\if{html}{\out{</span>}}
\itemize{
\item \if{html}{\out{<span class="color-rose">}}international law\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}fundamental rights\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}citizenship\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}private law\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}criminal law\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}data protection\if{html}{\out{</span>}}
}
}
\item \if{html}{\out{<span class="color-azure">}}foreign policy\if{html}{\out{</span>}}
\itemize{
\item \if{html}{\out{<span class="color-violet">}}position on foreign policy\if{html}{\out{</span>}}
\itemize{
\item \if{html}{\out{<span class="color-rose">}}neutrality\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}independence\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}good offices\if{html}{\out{</span>}}
}
\item \if{html}{\out{<span class="color-violet">}}European policy\if{html}{\out{</span>}}
\itemize{
\item \if{html}{\out{<span class="color-rose">}}EFTA\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}EU\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}EEA\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}other European organisations\if{html}{\out{</span>}}
}
\item \if{html}{\out{<span class="color-violet">}}international organisations\if{html}{\out{</span>}}
\itemize{
\item \if{html}{\out{<span class="color-rose">}}UN\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}other international
organisations\if{html}{\out{</span>}}
}
\item \if{html}{\out{<span class="color-violet">}}development cooperation\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-violet">}}bilateral treaties\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-violet">}}foreign trade policy\if{html}{\out{</span>}}
\itemize{
\item \if{html}{\out{<span class="color-rose">}}export promotion\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}customs\if{html}{\out{</span>}}
}
\item \if{html}{\out{<span class="color-violet">}}diplomacy\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-violet">}}Swiss abroad\if{html}{\out{</span>}}
}
\item \if{html}{\out{<span class="color-azure">}}security policy\if{html}{\out{</span>}}
\itemize{
\item \if{html}{\out{<span class="color-violet">}}public security\if{html}{\out{</span>}}
\itemize{
\item \if{html}{\out{<span class="color-rose">}}civil protection\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}state security\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}police\if{html}{\out{</span>}}
}
\item \if{html}{\out{<span class="color-violet">}}army\if{html}{\out{</span>}}
\itemize{
\item \if{html}{\out{<span class="color-rose">}}armed forces in general\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}military organisation\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}arms\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}military facilities\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}conscientious objection, civilian
service\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}abolition of the armed
forces\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}military training\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}international deployment\if{html}{\out{</span>}}
}
\item \if{html}{\out{<span class="color-violet">}}national economic supply\if{html}{\out{</span>}}
}
\item \if{html}{\out{<span class="color-azure">}}economy\if{html}{\out{</span>}}
\itemize{
\item \if{html}{\out{<span class="color-violet">}}economic policy\if{html}{\out{</span>}}
\itemize{
\item \if{html}{\out{<span class="color-rose">}}short-term economic policy\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}competition policy\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}structural policy\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}pricing policy\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}consumer protection\if{html}{\out{</span>}}
}
\item \if{html}{\out{<span class="color-violet">}}employment\if{html}{\out{</span>}}
\itemize{
\item \if{html}{\out{<span class="color-rose">}}employment conditions\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}working hours\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}industrial relations\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}employment policy\if{html}{\out{</span>}}
}
\item \if{html}{\out{<span class="color-violet">}}finance\if{html}{\out{</span>}}
\itemize{
\item \if{html}{\out{<span class="color-rose">}}monetary policy\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}banks, stock exchange,
insurance\if{html}{\out{</span>}}
}
\item \if{html}{\out{<span class="color-violet">}}tourism and leisure\if{html}{\out{</span>}}
\itemize{
\item \if{html}{\out{<span class="color-rose">}}tourism\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}hospitality\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}lottery and gambling\if{html}{\out{</span>}}
}
}
\item \if{html}{\out{<span class="color-azure">}}agriculture\if{html}{\out{</span>}}
\itemize{
\item \if{html}{\out{<span class="color-violet">}}agricultural policy\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-violet">}}livestock\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-violet">}}crop production\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-violet">}}forestry\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-violet">}}fishing, hunting and pets\if{html}{\out{</span>}}
}
\item \if{html}{\out{<span class="color-azure">}}public finance\if{html}{\out{</span>}}
\itemize{
\item \if{html}{\out{<span class="color-violet">}}taxation\if{html}{\out{</span>}}
\itemize{
\item \if{html}{\out{<span class="color-rose">}}tax policy\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}tax system\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}direct taxation\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}indirect taxation\if{html}{\out{</span>}}
}
\item \if{html}{\out{<span class="color-violet">}}financial system\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-violet">}}public expenditure\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-violet">}}budget cuts and remediation
measures\if{html}{\out{</span>}}
}
\item \if{html}{\out{<span class="color-azure">}}energy\if{html}{\out{</span>}}
\itemize{
\item \if{html}{\out{<span class="color-violet">}}energy policy\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-violet">}}nuclear energy\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-violet">}}hydro-electric power\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-violet">}}alternative energy\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-violet">}}oil and gas\if{html}{\out{</span>}}
}
\item \if{html}{\out{<span class="color-azure">}}transport and infrastructure\if{html}{\out{</span>}}
\itemize{
\item \if{html}{\out{<span class="color-violet">}}transport policy\if{html}{\out{</span>}}
\itemize{
\item \if{html}{\out{<span class="color-rose">}}urban transport\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}transit traffic\if{html}{\out{</span>}}
}
\item \if{html}{\out{<span class="color-violet">}}road transport\if{html}{\out{</span>}}
\itemize{
\item \if{html}{\out{<span class="color-rose">}}road construction\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}heavy traffic\if{html}{\out{</span>}}
}
\item \if{html}{\out{<span class="color-violet">}}rail transport\if{html}{\out{</span>}}
\itemize{
\item \if{html}{\out{<span class="color-rose">}}goods traffic\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}passenger traffic\if{html}{\out{</span>}}
}
\item \if{html}{\out{<span class="color-violet">}}air transport\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-violet">}}shipping\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-violet">}}post\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-violet">}}telecommunications\if{html}{\out{</span>}}
}
\item \if{html}{\out{<span class="color-azure">}}environment and living space\if{html}{\out{</span>}}
\itemize{
\item \if{html}{\out{<span class="color-violet">}}soil\if{html}{\out{</span>}}
\itemize{
\item \if{html}{\out{<span class="color-rose">}}spatial planning\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}land law\if{html}{\out{</span>}}
}
\item \if{html}{\out{<span class="color-violet">}}housing\if{html}{\out{</span>}}
\itemize{
\item \if{html}{\out{<span class="color-rose">}}tenancy issues\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}building of residential housing,
property ownership\if{html}{\out{</span>}}
}
\item \if{html}{\out{<span class="color-violet">}}environment\if{html}{\out{</span>}}
\itemize{
\item \if{html}{\out{<span class="color-rose">}}environmental policy\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}noise protection\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}air quality control\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}water pollution control\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}soil protection\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}waste\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}protection of nature and cultural
heritage\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}animal protection\if{html}{\out{</span>}}
}
}
\item \if{html}{\out{<span class="color-azure">}}social policy\if{html}{\out{</span>}}
\itemize{
\item \if{html}{\out{<span class="color-violet">}}health\if{html}{\out{</span>}}
\itemize{
\item \if{html}{\out{<span class="color-rose">}}health policy\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}medical research and
technology\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}medicines\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}addictive substances\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}reproductive medicine\if{html}{\out{</span>}}
}
\item \if{html}{\out{<span class="color-violet">}}social security\if{html}{\out{</span>}}
\itemize{
\item \if{html}{\out{<span class="color-rose">}}pension insurance\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}disability insurance\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}occupational benefits
provision\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}health and accident
insurance\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}maternity insurance\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}unemployment insurance\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}compensation for loss of earnings
for persons on military service or civil protection
duty\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}welfare\if{html}{\out{</span>}}
}
\item \if{html}{\out{<span class="color-violet">}}social groups\if{html}{\out{</span>}}
\itemize{
\item \if{html}{\out{<span class="color-rose">}}immigration policy\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}refugees\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}status of women\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}family policy\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}children and young people\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}senior citizens\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}persons with disabilities\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}homosexuals\if{html}{\out{</span>}}
}
}
\item \if{html}{\out{<span class="color-azure">}}education and research\if{html}{\out{</span>}}
\itemize{
\item \if{html}{\out{<span class="color-violet">}}education policy\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-violet">}}schools\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-violet">}}universities\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-violet">}}research\if{html}{\out{</span>}}
\itemize{
\item \if{html}{\out{<span class="color-rose">}}genetic engineering\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}animal testing\if{html}{\out{</span>}}
}
\item \if{html}{\out{<span class="color-violet">}}professional and vocational
education\if{html}{\out{</span>}}
}
\item \if{html}{\out{<span class="color-azure">}}culture, religion and media\if{html}{\out{</span>}}
\itemize{
\item \if{html}{\out{<span class="color-violet">}}cultural policy\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-violet">}}language policy\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-violet">}}religion, churches\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-violet">}}sport\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-violet">}}media and communication\if{html}{\out{</span>}}
\itemize{
\item \if{html}{\out{<span class="color-rose">}}media policy\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}press\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}radio, television and electronic
media\if{html}{\out{</span>}}
\item \if{html}{\out{<span class="color-rose">}}freedom of the media\if{html}{\out{</span>}}
}
}
}

The hierarchical relations between the three tag variables
\code{tags_tier_1}, \code{tags_tier_2} and \code{tags_tier_3} can be reconstructed at
any time using
\href{https://rpkg.dev/c2d/reference/hierarchize_tags.html}{\code{hierarchize_tags()}}.
This function can also be used to simply determine the parent tag(s) of
any tag.
\subsection{\code{tags_tier_1}}{

A maximum of 3 predefined top-tier political topics that best classify
the referendum.
}

\subsection{Possible values}{\tabular{l}{
   Value \cr
   \code{"state organisation"} \cr
   \code{"foreign policy"} \cr
   \code{"security policy"} \cr
   \code{"economy"} \cr
   \code{"agriculture"} \cr
   \code{"public finance"} \cr
   \code{"energy"} \cr
   \code{"transport and infrastructure"} \cr
   \code{"environment and living space"} \cr
   \code{"social policy"} \cr
   \code{"education and research"} \cr
   \code{"culture, religion and media"} \cr
}


\emph{This is a multi-value variable and thus of type \code{list}.}
}

\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{nominal} level.
}

\subsection{\code{tags_tier_2}}{

A maximum of 3 predefined second-tier political topics that best
classify the referendum; the presence of a second-tier topic always
implies presence of its parent top-tier topic.
}

\subsection{Possible values}{\tabular{l}{
   Value \cr
   \code{"national identity"} \cr
   \code{"political system"} \cr
   \code{"institutions"} \cr
   \code{"citizens' rights"} \cr
   \code{"federalism"} \cr
   \code{"legal system"} \cr
   \code{"position on foreign policy"} \cr
   \code{"European policy"} \cr
   \code{"international organisations"} \cr
   \code{"development cooperation"} \cr
   \code{"bilateral treaties"} \cr
   \code{"foreign trade policy"} \cr
   \code{"diplomacy"} \cr
   \code{"Swiss abroad"} \cr
   \code{"public security"} \cr
   \code{"army"} \cr
   \code{"national economic supply"} \cr
   \code{"economic policy"} \cr
   \code{"employment"} \cr
   \code{"finance"} \cr
   \code{"tourism and leisure"} \cr
   \code{"agricultural policy"} \cr
   \code{"livestock"} \cr
   \code{"crop production"} \cr
   \code{"forestry"} \cr
   \code{"fishing, hunting and pets"} \cr
   \code{"taxation"} \cr
   \code{"financial system"} \cr
   \code{"public expenditure"} \cr
   \code{"budget cuts and remediation measures"} \cr
   \code{"energy policy"} \cr
   \code{"nuclear energy"} \cr
   \code{"hydro-electric power"} \cr
   \code{"alternative energy"} \cr
   \code{"oil and gas"} \cr
   \code{"transport policy"} \cr
   \code{"road transport"} \cr
   \code{"rail transport"} \cr
   \code{"air transport"} \cr
   \code{"shipping"} \cr
   \code{"post"} \cr
   \code{"telecommunications"} \cr
   \code{"soil"} \cr
   \code{"housing"} \cr
   \code{"environment"} \cr
   \code{"health"} \cr
   \code{"social security"} \cr
   \code{"social groups"} \cr
   \code{"education policy"} \cr
   \code{"schools"} \cr
   \code{"universities"} \cr
   \code{"research"} \cr
   \code{"professional and vocational education"} \cr
   \code{"cultural policy"} \cr
   \code{"language policy"} \cr
   \code{"religion, churches"} \cr
   \code{"sport"} \cr
   \code{"media and communication"} \cr
}


\emph{This is a multi-value variable and thus of type \code{list}.}
}

\subsection{Applicability constraint}{

This variable is only applicable under the following condition:\if{html}{\out{<div class="sourceCode r">}}\preformatted{purrr::map2_lgl(.x = tags_tier_1, .y = tags_tier_2, .f = ~ all(c2d::infer_tags(tags = .y, tier = 1L) \%in\% .x))
}\if{html}{\out{</div>}}
}

\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{nominal} level.
}

\subsection{\code{tags_tier_3}}{

A maximum of 3 predefined third-tier political topics that best classify
the referendum; the presence of a third-tier topic always implies
presence of its parent second-tier topic and grandparent top-tier topic.
}

\subsection{Possible values}{\tabular{l}{
   Value \cr
   \code{"constitution"} \cr
   \code{"procedure for constitutional reform"} \cr
   \code{"legislative procedure"} \cr
   \code{"electoral system"} \cr
   \code{"government, administration"} \cr
   \code{"parliament"} \cr
   \code{"courts"} \cr
   \code{"national bank"} \cr
   \code{"citizens' initiatives"} \cr
   \code{"referendum"} \cr
   \code{"voting rights"} \cr
   \code{"territorial questions"} \cr
   \code{"intergovernmental relations"} \cr
   \code{"division of tasks"} \cr
   \code{"international law"} \cr
   \code{"fundamental rights"} \cr
   \code{"citizenship"} \cr
   \code{"private law"} \cr
   \code{"criminal law"} \cr
   \code{"data protection"} \cr
   \code{"neutrality"} \cr
   \code{"independence"} \cr
   \code{"good offices"} \cr
   \code{"EFTA"} \cr
   \code{"EU"} \cr
   \code{"EEA"} \cr
   \code{"other European organisations"} \cr
   \code{"UN"} \cr
   \code{"other international organisations"} \cr
   \code{"export promotion"} \cr
   \code{"customs"} \cr
   \code{"civil protection"} \cr
   \code{"state security"} \cr
   \code{"police"} \cr
   \code{"armed forces in general"} \cr
   \code{"military organisation"} \cr
   \code{"arms"} \cr
   \code{"military facilities"} \cr
   \code{"conscientious objection, civilian service"} \cr
   \code{"abolition of the armed forces"} \cr
   \code{"military training"} \cr
   \code{"international deployment"} \cr
   \code{"short-term economic policy"} \cr
   \code{"competition policy"} \cr
   \code{"structural policy"} \cr
   \code{"pricing policy"} \cr
   \code{"consumer protection"} \cr
   \code{"employment conditions"} \cr
   \code{"working hours"} \cr
   \code{"industrial relations"} \cr
   \code{"employment policy"} \cr
   \code{"monetary policy"} \cr
   \code{"banks, stock exchange, insurance"} \cr
   \code{"tourism"} \cr
   \code{"hospitality"} \cr
   \code{"lottery and gambling"} \cr
   \code{"tax policy"} \cr
   \code{"tax system"} \cr
   \code{"direct taxation"} \cr
   \code{"indirect taxation"} \cr
   \code{"urban transport"} \cr
   \code{"transit traffic"} \cr
   \code{"road construction"} \cr
   \code{"heavy traffic"} \cr
   \code{"goods traffic"} \cr
   \code{"passenger traffic"} \cr
   \code{"spatial planning"} \cr
   \code{"land law"} \cr
   \code{"tenancy issues"} \cr
   \code{"building of residential housing, property ownership"} \cr
   \code{"environmental policy"} \cr
   \code{"noise protection"} \cr
   \code{"air quality control"} \cr
   \code{"water pollution control"} \cr
   \code{"soil protection"} \cr
   \code{"waste"} \cr
   \code{"protection of nature and cultural heritage"} \cr
   \code{"animal protection"} \cr
   \code{"health policy"} \cr
   \code{"medical research and technology"} \cr
   \code{"medicines"} \cr
   \code{"addictive substances"} \cr
   \code{"reproductive medicine"} \cr
   \code{"pension insurance"} \cr
   \code{"disability insurance"} \cr
   \code{"occupational benefits provision"} \cr
   \code{"health and accident insurance"} \cr
   \code{"maternity insurance"} \cr
   \code{"unemployment insurance"} \cr
   \code{"compensation for loss of earnings for persons on military service or civil protection duty"} \cr
   \code{"welfare"} \cr
   \code{"immigration policy"} \cr
   \code{"refugees"} \cr
   \code{"status of women"} \cr
   \code{"family policy"} \cr
   \code{"children and young people"} \cr
   \code{"senior citizens"} \cr
   \code{"persons with disabilities"} \cr
   \code{"homosexuals"} \cr
   \code{"genetic engineering"} \cr
   \code{"animal testing"} \cr
   \code{"media policy"} \cr
   \code{"press"} \cr
   \code{"radio, television and electronic media"} \cr
   \code{"freedom of the media"} \cr
}


\emph{This is a multi-value variable and thus of type \code{list}.}
}

\subsection{Applicability constraint}{

This variable is only applicable under the following condition:\if{html}{\out{<div class="sourceCode r">}}\preformatted{purrr::map2_lgl(.x = tags_tier_2, .y = tags_tier_3, .f = ~ all(c2d::infer_tags(tags = .y, tier = 2L) \%in\% .x)) & purrr::map2_lgl(.x = tags_tier_1, .y = tags_tier_3, .f = ~ all(c2d::infer_tags(tags = .y, tier = 1L) \%in\% .x))
}\if{html}{\out{</div>}}
}

\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{nominal} level.
}

}

\subsection{\code{remarks}}{

A free text field with additional information about the referendum.
}

\subsection{\code{files}}{

Metadata about files attached to the referendum, including
\itemize{
\item \code{date_time_attached}: the date and time on which the file was
attached to the referendum’s data entry
\item \code{is_deleted}: whether or not the file has been deleted since it was
attached
\item \code{file_name}: the filename as it appears when the file is downloaded
\item \code{file_size}: the file’s size in bytes
\item \code{s3_object_key}: the key uniquely identifying the file in the C2D
\href{https://en.wikipedia.org/wiki/Amazon_S3#Design}{Amazon S3 bucket};
the file can be downloaded by feeding this key to
\href{https://rpkg.dev/c2d/reference/download_file_attachment.html}{\code{c2d::download_file_attachment()}}.
}
}

\subsection{\code{url_sudd}}{

The URL of the corresponding entry on \href{https://sudd.ch/}{sudd.ch}.
\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{nominal} level.
}

}

\subsection{\code{url_swissvotes}}{

The URL of the corresponding entry on
\href{https://swissvotes.ch/}{swissvotes.ch}.
\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{nominal} level.
}

}

\subsection{\code{is_draft}}{

Whether or not the entry is marked as a draft in the database.
\subsection{Possible values}{\tabular{ll}{
   Value \tab Meaning \cr
   \code{TRUE} \tab The entry is marked as a draft. \cr
   \code{FALSE} \tab The entry is not marked as a draft. \cr
}

}

\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{nominal (binary)} level.
}

}

\subsection{\code{date_time_created}}{

The date and time on which the referendum’s data entry was created.
\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{interval} level.
}

}

\subsection{\code{archive}}{

The referendum’s data from an earlier, obsolete state of the C2D
database.
}

\subsection{\code{type}}{

The institutional type of direct democracy which the referendum is an
instance of.
\subsection{Possible values}{\tabular{ll}{
   Value \tab Meaning \cr
   \code{"mandatory referendum"} \tab A referendum that was automatically triggered by certain legal conditions, usually found in the constitution. \cr
   \code{"optional referendum"} \tab A referendum against a law passed by the parliament that came about because the quorum of citizen opposition was met (usually a certain number of signatures) within a specified period after the resolution was passed. \cr
   \code{"counter proposal"} \tab A counter proposal by the government or parliament to another referendum. \cr
   \code{"citizens' initiative"} \tab A referendum launched by citizens, usually via a signature collection. \cr
   \code{"governmental referendum"} \tab A referendum launched by the government. \cr
   \code{"citizens' assembly"} \tab A citizens’ assembly, usually in person where people vote by raising their hands. \cr
   \code{"recall"} \tab A recall of all members of a political institution/body. \cr
}

}

\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{nominal} level.
}

}

}

\subsection{Institution-level variables}{

For the following variables, the \href{https://en.wikipedia.org/wiki/Unit_of_observation}{unit of observation} is the
referendum’s \emph{institutional instrument}, a combination of the
referendum’s \code{type} and the jurisdiction it took place.

The set of variables that constitute the \emph{geographical} component of the
jurisdiction is dependent on the referendum’s \code{level}:\tabular{ll}{
   \code{level} \tab Geographical jurisdiction is given by \cr
   \code{national} \tab \code{country_code} \cr
   \code{subnational} \tab \code{country_code} + \code{subnational_entity_name} \cr
   \code{local} \tab \code{country_code} + \code{subnational_entity_name} + \code{municipality} \cr
}


The actual jurisdiction is also dependent on a \emph{temporal} component
since the legal setting in a geographical jurisdiction can vary over
time. Thus, the actual jurisdicion is dependent on the geographical
jurisdiction plus the \code{date} a referendum took place (more specifically
the date \emph{range} in between a specific legal setting is in force).
\subsection{\code{inst_legal_basis_type}}{

The type of the legal basis of the referendum \code{type}.
\subsection{Possible values}{\tabular{ll}{
   Value \tab Meaning \cr
   \code{"non-official"} \tab The referendum \code{type} has no legal basis. \cr
   \code{"ad-hoc"} \tab The referendum \code{type} has a legal basis which was specifically created for it. \cr
   \code{"official"} \tab The referendum \code{type} has a legal basis that wasn’t specifically created for it. \cr
}

}

\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{ordinal (ascending)} level.
}

}

\subsection{\code{inst_has_urgent_legal_basis}}{

Whether or not the legal basis of the referendum \code{type} was adopted
urgently, i.e. in an abridged procedure.
\subsection{Possible values}{\tabular{ll}{
   Value \tab Meaning \cr
   \code{TRUE} \tab The legal basis of the referendum \code{type} was adopted in an abridged procedure (urgent). \cr
   \code{FALSE} \tab The legal basis of the referendum \code{type} was adopted in a normal procedure. \cr
}

}

\subsection{Applicability constraint}{

This variable is only applicable under the following condition:\if{html}{\out{<div class="sourceCode r">}}\preformatted{inst_trigger_actor \%in\% c('government', 'president', 'monarch', 'parliament')
}\if{html}{\out{</div>}}
}

\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{nominal (binary)} level.
}

}

\subsection{\code{inst_is_binding}}{

Whether the result of the referendum \code{type} has binding force or the
decision of the voters may be disregarded.
\subsection{Possible values}{\tabular{ll}{
   Value \tab Meaning \cr
   \code{TRUE} \tab The result of the referendum \code{type} is binding for the legislator. \cr
   \code{FALSE} \tab The result of the referendum \code{type} is only consultative, i.e. the legislator does not have to follow the decision of the voters. \cr
}

}

\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{nominal (binary)} level.
}

}

\subsection{\code{inst_is_counter_proposal}}{

Whether or not the referendum \code{type} is a counterproposal to a citizen’s
(or other) initiative.
\subsection{Possible values}{\tabular{ll}{
   Value \tab Meaning \cr
   \code{TRUE} \tab The referendum \code{type} is a counterproposal to a citizen’s (or other) initiative. \cr
   \code{FALSE} \tab The referendum \code{type} is not a counterproposal to a citizen’s (or other) initiative. \cr
}

}

\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{nominal (binary)} level.
}

}

\subsection{\code{inst_is_assembly}}{

Whether or not the referendum \code{type} is held as an assembly (all voters
at the same time and place; physical presence at the assembly is
required to be able to vote; usually voting means raising one’s hand).
\subsection{Possible values}{\tabular{ll}{
   Value \tab Meaning \cr
   \code{TRUE} \tab The referendum \code{type} is a counterproposal to a citizen’s (or other) initiative. \cr
   \code{FALSE} \tab The referendum \code{type} is not a counterproposal to a citizen’s (or other) initiative. \cr
}

}

\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{nominal (binary)} level.
}

}

\subsection{\code{inst_trigger_type}}{

The way the referendum \code{type} is triggered; completely dependent on
\code{inst_trigger_actor}.
\subsection{Possible values}{\tabular{ll}{
   Value \tab Meaning \cr
   \code{"automatic"} \tab The referendum \code{type} is triggered by a constitutional/legal requirement. \cr
   \code{"top down"} \tab The referendum \code{type} is triggered by an institution of the political elite like the monarch/president/government, the parliament, a territorial unit, the UN or another institution. \cr
   \code{"bottom up"} \tab The referendum \code{type} is triggered by citizen demand (e.g. a signature collection). \cr
}

}

\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{nominal} level.
}

}

\subsection{\code{inst_trigger_actor}}{

The actor that triggers the referendum \code{type}.
\subsection{Possible values}{\tabular{ll}{
   Value \tab Meaning \cr
   \code{"citizens"} \tab The referendum \code{type} is triggered by a direct citizen demand (usually a signature collection). \cr
   \code{"constitution"} \tab The referendum \code{type} is triggered by a constitutional requirement. \cr
   \code{"law"} \tab The referendum \code{type} is triggered by a legal requirement. \cr
   \code{"parliament"} \tab The referendum \code{type} is triggered by the parliament. \cr
   \code{"government"} \tab The referendum \code{type} is triggered by the government or prime minister. \cr
   \code{"president"} \tab The referendum \code{type} is triggered by the president. \cr
   \code{"monarch"} \tab The referendum \code{type} is triggered by the monarch. \cr
   \code{"territorial unit"} \tab The referendum \code{type} is triggered by a province or municipality. \cr
   \code{"UN"} \tab The referendum \code{type} is triggered by the United Nations (UN) or one of its organs. \cr
   \code{"other institution"} \tab The referendum \code{type} is triggered by another institution such as the church or a state agency. \cr
}

}

\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{nominal} level.
}

}

\subsection{\code{inst_trigger_actor_level}}{

The political \code{level} of \code{inst_trigger_actor} in comparison to \code{level}.
\subsection{Possible values}{\tabular{ll}{
   Value \tab Meaning \cr
   \code{"lower"} \tab A lower political level triggers a referendum on one of the higher political levels. \cr
   \code{"equal"} \tab The trigger actor and the referendum are on the same political level. \cr
   \code{"higher"} \tab A higher political level triggers a referendum on one of the lower political levels. \cr
}

}

\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{ordinal (ascending)} level.
}

}

\subsection{\code{inst_trigger_threshold}}{

The threshold of support among \code{inst_trigger_actor} that has to be
reached in order to trigger the referendum \code{type}.
\subsection{Applicability constraint}{

This variable is only applicable under the following condition:\if{html}{\out{<div class="sourceCode r">}}\preformatted{inst_trigger_actor \%in\% c('citizens', 'parliament', 'territorial unit')
}\if{html}{\out{</div>}}
}

\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{nominal} level.
}

}

\subsection{\code{inst_trigger_time_limit}}{

The time limit in which \code{inst_trigger_threshold} (usually a minimum
number of signatures) has to be reached in order to trigger the
referendum \code{type}.
\subsection{Applicability constraint}{

This variable is only applicable under the following condition:\if{html}{\out{<div class="sourceCode r">}}\preformatted{inst_trigger_actor \%in\% c('citizens', 'territorial unit')
}\if{html}{\out{</div>}}
}

\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{nominal} level.
}

}

\subsection{\code{inst_object_type}}{

The type of the object the referendum \code{type} refers to.
\subsection{Possible values}{\tabular{ll}{
   Value \tab Meaning \cr
   \code{"question"} \tab A question which can be answered with yes or no (used as fallback option if nothing else applies). \cr
   \code{"legal text (general proposal)"} \tab A general proposal of a legal text, meaning the proposal is not formulated in a definite way, and therefore, if approved, needs to be implemented and concretized in a further legislative process. \cr
   \code{"legal text (formulated proposal)"} \tab A precisely formulated legal text which, if approved, will directly enter into force. \cr
   \code{"institution"} \tab A recall of all members of a particular institution/body. \cr
   \code{"person"} \tab An appointment of a single person. \cr
   \code{"principle"} \tab A question about what direction to take on a particular issue/topic, but without a proposed legal text; e.g., “Should there be a total revision of our constitution?”. \cr
}

}

\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{nominal} level.
}

}

\subsection{\code{inst_object_author}}{

The actor that defines the text/content of \code{inst_object_type}.
\subsection{Possible values}{\tabular{ll}{
   Value \tab Meaning \cr
   \code{"citizens"} \tab Citizens (usually a small group initiating a citizen demand). \cr
   \code{"parliament"} \tab The parliament. \cr
   \code{"government"} \tab The government or prime minister. \cr
   \code{"president"} \tab The president. \cr
   \code{"monarch"} \tab The monarch. \cr
   \code{"territorial unit"} \tab A province or municipality. \cr
   \code{"UN"} \tab The United Nations (UN) or one of its organs. \cr
   \code{"other institution"} \tab Another institution such as the church or a state agency. \cr
}

}

\subsection{Applicability constraint}{

This variable is only applicable under the following condition:\if{html}{\out{<div class="sourceCode r">}}\preformatted{inst_object_type \%in\% c('question', 'legal text (general proposal)', 'legal text (formulated proposal)', 'principle')
}\if{html}{\out{</div>}}
}

\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{nominal} level.
}

}

\subsection{\code{inst_object_legal_level}}{

The legal level of \code{inst_object_type}.
\subsection{Possible values}{\tabular{ll}{
   Value \tab Meaning \cr
   \code{"ordinance"} \tab An ordinance, i.e. a legal norm based on a law or the constitution. \cr
   \code{"treaty"} \tab A domestic treaty, e.g. an agreement between the constituent states of a republic. \cr
   \code{"law"} \tab A law, i.e. a legal norm based on the constitution and created by the legislative. \cr
   \code{"constitution"} \tab The constitution, i.e. the aggregate of fundamental principles or established precedents that constitute the legal basis of a state and commonly determine how that state is to be governed. \cr
   \code{"international treaty"} \tab An international treaty, i.e. a bilateral or multilateral agreement or an accession to and international organization. \cr
}

}

\subsection{Applicability constraint}{

This variable is only applicable under the following condition:\if{html}{\out{<div class="sourceCode r">}}\preformatted{inst_object_type \%in\% c('legal text (general proposal)', 'legal text (formulated proposal)')
}\if{html}{\out{</div>}}
}

\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{nominal} level.
}

}

\subsection{\code{inst_object_revision_extent}}{

The extent of revision of \code{inst_object_type} which can be achieved
within one referendum.
\subsection{Possible values}{\tabular{ll}{
   Value \tab Meaning \cr
   \code{"partial"} \tab Only parts of the concerned legal text can be revised at the same time. \cr
   \code{"total"} \tab Only the entirety of the legal text can be changed at the same time. \cr
   \code{"both"} \tab Both, parts and the entirety of the legal text, can be changed at the same time. \cr
}

}

\subsection{Applicability constraint}{

This variable is only applicable under the following condition:\if{html}{\out{<div class="sourceCode r">}}\preformatted{inst_object_type \%in\% c('legal text (general proposal)', 'legal text (formulated proposal)')
}\if{html}{\out{</div>}}
}

\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{nominal} level.
}

}

\subsection{\code{inst_object_revision_modes}}{

The mode(s) of revision that \code{inst_object_type} allows.
\subsection{Possible values}{\tabular{ll}{
   Value \tab Meaning \cr
   \code{"introduce"} \tab With this instrument, it is possible to introduce an entirely new text/concept. \cr
   \code{"revise"} \tab With this instrument, it is possible to change some parts of the legal text/… \cr
   \code{"abrogate"} \tab With this instrument, it is possible to strike parts of the legal text / to recall a person/institution. \cr
}


\emph{This is a multi-value variable and thus of type \code{list}.}
}

\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{nominal} level.
}

}

\subsection{\code{inst_object_revision_divisibility}}{

The divisibility of \code{inst_object_type}.
\subsection{Possible values}{\tabular{ll}{
   Value \tab Meaning \cr
   \code{"as a whole"} \tab The proposed legal text can only be accepted/rejected as a whole. \cr
   \code{"split up"} \tab The proposed legal text can be split up, the voters may vote upon these parts separately. \cr
   \code{"as variants"} \tab The proposed legal text can be presented as two or more (alternative) variants. \cr
   \code{"both split up and as variants"} \tab The proposed legal text can be both split up as well as presented as two or more (alternative) variants. \cr
}

}

\subsection{Applicability constraint}{

This variable is only applicable under the following condition:\if{html}{\out{<div class="sourceCode r">}}\preformatted{inst_object_type \%in\% c('question', 'legal text (general proposal)', 'legal text (formulated proposal)')
}\if{html}{\out{</div>}}
}

\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{nominal} level.
}

}

\subsection{\code{inst_topics_only}}{

The legal topic(s) to which the referendum \code{type} is limited.
\subsection{Possible values}{\tabular{ll}{
   Value \tab Meaning \cr
   \code{"infrastructural acts"} \tab The referendum \code{type} is limited to TODO. \cr
   \code{"territorial questions"} \tab The referendum \code{type} is limited to TODO. \cr
   \code{"competence shifts"} \tab The referendum \code{type} is limited to TODO. \cr
   \code{"financial acts"} \tab The referendum \code{type} is limited to TODO. \cr
   \code{"financial acts (expenses)"} \tab The referendum \code{type} is limited to TODO. \cr
   \code{"financial acts (taxes)"} \tab The referendum \code{type} is limited to TODO. \cr
   \code{"financial acts (obligations)"} \tab The referendum \code{type} is limited to TODO. \cr
   \code{"treaties"} \tab The referendum \code{type} is limited to TODO (automatically limited to if \code{inst_object_legal_level == "law"}). \cr
   \code{"total revisions of the constitution"} \tab The referendum \code{type} is limited to TODO. \cr
}


\emph{This is a multi-value variable and thus of type \code{list}.}
}

\subsection{Applicability constraint}{

This variable is only applicable under the following condition:\if{html}{\out{<div class="sourceCode r">}}\preformatted{inst_topics_excluded \%>\% purrr::map_lgl(rlang::is_empty)
}\if{html}{\out{</div>}}
}

\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{nominal} level.
}

}

\subsection{\code{inst_topics_excluded}}{

The legal topic(s) that are excluded from the referendum \code{type}.
\subsection{Possible values}{\tabular{ll}{
   Value \tab Meaning \cr
   \code{"jus cogens"} \tab TODO is exluded from the referendum \code{type}. \cr
   \code{"international treaties"} \tab TODO is exluded from the referendum \code{type}. \cr
   \code{"pardons"} \tab TODO is exluded from the referendum \code{type}. \cr
   \code{"personnel appointments"} \tab TODO is exluded from the referendum \code{type}. \cr
   \code{"budgets"} \tab TODO is exluded from the referendum \code{type}. \cr
   \code{"financial matters"} \tab TODO is exluded from the referendum \code{type}. \cr
   \code{"everything within parliamentary competence"} \tab TODO is exluded from the referendum \code{type}. \cr
}


\emph{This is a multi-value variable and thus of type \code{list}.}
}

\subsection{Applicability constraint}{

This variable is only applicable under the following condition:\if{html}{\out{<div class="sourceCode r">}}\preformatted{inst_topics_only \%>\% purrr::map_lgl(rlang::is_empty)
}\if{html}{\out{</div>}}
}

\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{nominal} level.
}

}

\subsection{\code{inst_quorum_turnout}}{

The minimum percentage of the \code{electorate_total} that have to
participate in the referendum \code{type} in order for it to be passed.
\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{nominal} level.
}

}

\subsection{\code{inst_quorum_approval}}{

The minimum number of voters that have to approve the referendum \code{type}
in order for it to be passed.
\subsection{Possible values}{\tabular{ll}{
   Value \tab Meaning \cr
   \code{"simple majority"} \tab > 50 \% of the voters have to approve the referendum proposal. \cr
   \code{"qualified majority"} \tab A percentage higher than > 50 \% of the voters has to approve the referendum proposal. \cr
   \code{"double majority"} \tab Two simultaneous quorums of > 50 \% or higher of the voters apply. \cr
}

}

\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{nominal} level.
}

}

\subsection{\code{inst_has_precondition}}{

Whether or not an institutional actor must take a preconditional
decision in order for the referendum \code{type} to be held.
\subsection{Possible values}{\tabular{ll}{
   Value \tab Meaning \cr
   \code{TRUE} \tab An institutional actor must take a preconditional decision in order for the referendum \code{type} to be held. \cr
   \code{FALSE} \tab No institutional actor must take a preconditional decision in order for the referendum \code{type} to be held. \cr
}

}

\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{nominal (binary)} level.
}

}

\subsection{\code{inst_precondition_actor}}{

The actor that must take a preconditional decision in order for the
referendum \code{type} to be held.
\subsection{Possible values}{\tabular{ll}{
   Value \tab Meaning \cr
   \code{"parliament"} \tab The parliament. \cr
   \code{"government"} \tab The government or prime minister. \cr
   \code{"president"} \tab The president. \cr
   \code{"UN"} \tab The United Nations (UN) or one of its organs. \cr
   \code{"parliament and government"} \tab The parliament and the government or prime minister. \cr
   \code{"parliament and president"} \tab The parliament and the president. \cr
   \code{"central government and SPLM"} \tab The central government and the \href{https://en.wikipedia.org/wiki/Sudan_People's_Liberation_Movement}{Sudan People’s Liberation Movement}. \cr
}

}

\subsection{Applicability constraint}{

This variable is only applicable under the following condition:\if{html}{\out{<div class="sourceCode r">}}\preformatted{isTRUE(inst_has_precondition)
}\if{html}{\out{</div>}}
}

\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{nominal} level.
}

}

\subsection{\code{inst_precondition_decision}}{

The decision that \code{inst_precondition_actor} has taken.
\subsection{Possible values}{\tabular{ll}{
   Value \tab Meaning \cr
   \code{"yes"} \tab Affirmative answer (e.g. “Yes, we will hold a vote that has been requested by citizens.”). \cr
   \code{"no"} \tab Negative answer (e.g. “No, we will not implement the demands of a citizens’ initiative voluntarily.”). \cr
}

}

\subsection{Level of measurement}{

This variable’s \href{https://en.wikipedia.org/wiki/Level_of_measurement}{scale of measure} is the
\strong{nominal (binary)} level.
}

}

}
}
\seealso{
Other referendum metadata functions:
\code{\link{v_vals}()},
\code{\link{val_lbls}()}
}
\concept{metadata}
\keyword{datasets}
