% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d.gen.R
\docType{data}
\name{data_tags}
\alias{data_tags}
\title{Tag hierarchy}
\format{
A \link[tibble:tbl_df-class]{tibble}.
}
\usage{
data_tags
}
\description{
A tibble reflecting the complete \href{https://rpkg.dev/c2d/articles/codebook.html#tags}{referendum tags hierarchy}.
}
\examples{
c2d::data_tags
}
\seealso{
Other tag-related functions:
\code{\link{hierarchize_tags_fast}()},
\code{\link{hierarchize_tags}()},
\code{\link{infer_tags}()},
\code{\link{tags}()}
}
\concept{tags}
\keyword{datasets}
