% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\docType{data}
\name{data_topics}
\alias{data_topics}
\title{Topic hierarchy}
\format{
A \link[tibble:tbl_df-class]{tibble}.
}
\usage{
data_topics
}
\description{
A tibble reflecting the complete \href{https://rdb.rpkg.dev/articles/codebook.html#topics}{referendum topics hierarchy}.
}
\examples{
rdb::data_topics
}
\seealso{
Other auxiliary data:
\code{\link{data_iso_3166_1}},
\code{\link{data_iso_3166_2}},
\code{\link{data_iso_3166_3}},
\code{\link{data_iso_639_1}},
\code{\link{data_municipalities}}

Other referendum topic functions:
\code{\link{hierarchize_topics}()},
\code{\link{hierarchize_topics_fast}()},
\code{\link{infer_topics}()},
\code{\link{nest_topics}()},
\code{\link{topics}()},
\code{\link{update_topics}()}
}
\concept{aux_data}
\concept{topics}
\keyword{datasets}
