% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d.gen.R
\name{delete_rfrnds}
\alias{delete_rfrnds}
\title{Delete referendums in the C2D database}
\usage{
delete_rfrnds(
  ids,
  email = pal::pkg_config_val(key = "api_username", pkg = this_pkg),
  password = pal::pkg_config_val(key = "api_password", pkg = this_pkg),
  use_testing_server = pal::pkg_config_val(key = "use_testing_server", pkg = this_pkg)
)
}
\arguments{
\item{ids}{IDs of the referendums to be deleted. A character vector.}

\item{email}{The e-mail address of the C2D API user account to be used for authentication. A character scalar.}

\item{password}{The password of the C2D API user account to be used for authentication. A character scalar.}

\item{use_testing_server}{Whether or not to use the testing servers instead of the production servers for C2D Services API calls etc.}
}
\value{
\code{ids}, invisibly.
}
\description{
Deletes existing referendum entries in the C2D database via \href{https://github.com/ccmdesign/c2d-app/blob/master/docs/services.md#3-referendum-routes}{its API}.
}
\seealso{
Other referendum data functions:
\code{\link{add_rfrnds}()},
\code{\link{count_rfrnds}()},
\code{\link{download_file_attachment}()},
\code{\link{edit_rfrnds}()},
\code{\link{rfrnd_exists}()},
\code{\link{rfrnds}()},
\code{\link{rfrnd}()},
\code{\link{search_rfrnds}()},
\code{\link{validate_rfrnds}()}
}
\concept{rfrnd}
