% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{delete_rfrnds}
\alias{delete_rfrnds}
\title{Delete referendums from the RDB}
\usage{
delete_rfrnds(data, connection = connect(), disconnect = TRUE)
}
\arguments{
\item{data}{Referendum data to be deleted. A \link[tibble:tbl_df-class]{tibble} that at least must contain the column
\code{id}.}

\item{connection}{\code{\link[RPostgres:PqConnection-class]{PqConnection}}\cr
Database connection to a PostgreSQL DBMS as returned by \code{\link[=connect]{connect()}}.}

\item{disconnect}{\code{\link[base:logical]{logical(1)}}\cr
Whether or not to \link[DBI:dbDisconnect]{terminate} the database connection when finished.}
}
\value{
The updated \code{referendums} table data as a \link[tibble:tbl_df-class]{tibble}, invisibly.
}
\description{
Deletes \code{data} from the RDB's \code{referendums} table.
}
\seealso{
Other administration functions:
\code{\link{backup_rdb}()},
\code{\link{clean_attachment_bucket}()},
\code{\link{config_nocodb_tbls}()},
\code{\link{connect}()},
\code{\link{create_nocodb_tbl_cols}()},
\code{\link{create_nocodb_tbls}()},
\code{\link{dm}()},
\code{\link{notify_postgrest}()},
\code{\link{purge_nocodb}()},
\code{\link{reset_nocodb}()},
\code{\link{reset_rdb}()},
\code{\link{restart_neon_ep}()},
\code{\link{restart_postgrest}()},
\code{\link{restore_rdb}()},
\code{\link{update_administrative_units}()},
\code{\link{update_countries}()},
\code{\link{update_langs}()},
\code{\link{update_municipalities}()},
\code{\link{update_rfrnds}()},
\code{\link{update_subnational_entities}()},
\code{\link{update_tbl}()},
\code{\link{update_topics}()}

Other data functions:
\code{\link{assert_vars}()},
\code{\link{read_tbl}()},
\code{\link{rfrnd_types}()},
\code{\link{rfrnds}()},
\code{\link{rfrnds_old}()},
\code{\link{validate_rfrnds}()}
}
\concept{admin}
\concept{data}
