% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{dm}
\alias{dm}
\title{Get RDB data model}
\usage{
dm(
  connection = connect(user = "rdb_admin", password = pg_role_pw("rdb_admin"))
)
}
\arguments{
\item{connection}{Database connection to a PostgreSQL DBMS as returned by \code{\link[=connect]{connect()}}. An S4 object that inherits from \code{\link[DBI:DBIConnection-class]{DBI::DBIConnection}}.}
}
\value{
A \link[dm:dm]{data model} object.
}
\description{
Get RDB data model
}
\examples{
rdb::dm() |> dm::dm_draw(view_type = "all")
}
\seealso{
Other administration functions:
\code{\link{config_nocodb_tbls}()},
\code{\link{connect}()},
\code{\link{create_nocodb_tbls}()},
\code{\link{delete_rfrnds}()},
\code{\link{notify_postgrest}()},
\code{\link{purge_nocodb}()},
\code{\link{reset_nocodb}()},
\code{\link{reset_rdb}()},
\code{\link{restart_postgrest}()},
\code{\link{update_countries}()},
\code{\link{update_langs}()},
\code{\link{update_municipalities}()},
\code{\link{update_rfrnds}()},
\code{\link{update_subnational_entities}()},
\code{\link{update_tbl}()}
}
\concept{admin}
