% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{dm}
\alias{dm}
\title{Get RDB data model}
\usage{
dm(
  connection = connect(user = "rdb_admin", password = pg_role_pw("rdb_admin")),
  incl_admin_tbls = TRUE,
  incl_aux_admin_tbls = FALSE
)
}
\arguments{
\item{connection}{\code{\link[RPostgres:PqConnection-class]{PqConnection}}\cr
Database connection to a PostgreSQL DBMS as returned by \code{\link[=connect]{connect()}}.}

\item{incl_admin_tbls}{Whether or not to include tables that are administered fully programmatically
(\code{administrative_units}, \code{supranational_entities}, \code{countries}, \code{subnational_entities}, \code{municipalities}, \code{languages}, \code{topics} and \code{nocodb_users}).}

\item{incl_aux_admin_tbls}{Whether or not to include the auxiliary administrative unit tables
\code{supranational_entities}, \code{countries}, \code{subnational_entities}, \code{municipalities} and \code{nocodb_users}.}
}
\value{
A \link[dm:dm]{data model} object.
}
\description{
Get RDB data model
}
\examples{
# visualize the whole RDB model
rdb::dm() |> dm::dm_draw(view_type = "all")

# download all RDB tables into local object (e.g. for later processing)
rdb_dm <- rdb::dm() |> dplyr::collect()
}
\seealso{
Other administration functions:
\code{\link{backup_rdb}()},
\code{\link{clean_attachment_bucket}()},
\code{\link{config_nocodb_tbls}()},
\code{\link{connect}()},
\code{\link{create_nocodb_tbl_cols}()},
\code{\link{create_nocodb_tbls}()},
\code{\link{delete_rfrnds}()},
\code{\link{notify_postgrest}()},
\code{\link{purge_nocodb}()},
\code{\link{reset_nocodb}()},
\code{\link{reset_rdb}()},
\code{\link{restart_neon_ep}()},
\code{\link{restart_postgrest}()},
\code{\link{restore_rdb}()},
\code{\link{update_administrative_units}()},
\code{\link{update_countries}()},
\code{\link{update_langs}()},
\code{\link{update_municipalities}()},
\code{\link{update_nocodb_users}()},
\code{\link{update_rfrnds}()},
\code{\link{update_subnational_entities}()},
\code{\link{update_tbl}()},
\code{\link{update_topics}()}
}
\concept{admin}
