% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d-GEN.R
\name{download_file_attachment}
\alias{download_file_attachment}
\title{Download file attachment}
\usage{
download_file_attachment(s3_object_key, path = ".")
}
\arguments{
\item{s3_object_key}{The key uniquely identifying the file in the C2D \href{https://en.wikipedia.org/wiki/Amazon_S3#Design}{Amazon S3 bucket}. A character scalar.}

\item{path}{The path where the downloaded file is written to. If a directory, the original filename returned by the C2D services API will be used.}
}
\value{
A \link[httr:response]{response object}, invisibly.
}
\description{
Download file attachment
}
\examples{
library(magrittr)

\dontrun{
# get all file object keys...
c2d::referendums() \%$\%
  files \%>\%
  purrr::map_depth(2L, purrr::pluck, "object_key") \%>\%
  purrr::flatten() \%>\%
  purrr::flatten_chr() \%>\%
  # ...select first one...
  dplyr::first() \%>\%
  # ...and download the corresponding file to the current working dir
  c2d::download_file_attachment()}
}
