% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d.gen.R
\name{download_file_attachment}
\alias{download_file_attachment}
\title{Download file attachment}
\usage{
download_file_attachment(
  s3_object_key,
  path = ".",
  use_testing_server = pal::pkg_config_val(key = "use_testing_server", pkg = this_pkg)
)
}
\arguments{
\item{s3_object_key}{The key uniquely identifying the file in the C2D \href{https://en.wikipedia.org/wiki/Amazon_S3#Design}{Amazon S3 bucket}. A character scalar.}

\item{path}{The path where the downloaded file is written to. If a directory, the original filename returned by the C2D services API will be used.}

\item{use_testing_server}{Whether or not to use the testing servers instead of the production servers for C2D Services API calls etc.}
}
\value{
A \link[httr:response]{response object}, invisibly.
}
\description{
Downloads a file attachment from the C2D database. The necessary \code{s3_object_key}s identifying individual files are found in the \code{files} list column returned
by \code{\link[=rfrnds]{rfrnds()}}.
}
\examples{
\dontrun{
# get all file object keys...
c2d::rfrnds()$files |>
  purrr::map_depth(2L, purrr::pluck, "s3_object_key") |>
  purrr::flatten() |>
  purrr::flatten_chr() |>
  # ...select first three keys...
  magrittr::extract(1:3) |>
  # ...and download the corresponding files to the current working dir
  purrr::walk(c2d::download_file_attachment)}
}
\seealso{
Other referendum data functions:
\code{\link{add_rfrnds}()},
\code{\link{count_rfrnds}()},
\code{\link{delete_rfrnds}()},
\code{\link{edit_rfrnds}()},
\code{\link{rfrnds}()},
\code{\link{rfrnd}()},
\code{\link{search_rfrnds}()},
\code{\link{validate_rfrnds}()}
}
\concept{rfrnd}
