% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{download_file_attachment}
\alias{download_file_attachment}
\title{Download file attachment}
\usage{
download_file_attachment(
  s3_object_key,
  path = ".",
  use_original_filename = FALSE,
  use_testing_server = pal::pkg_config_val(key = "use_testing_server", pkg = this_pkg)
)
}
\arguments{
\item{s3_object_key}{Key uniquely identifying the file in the RDB \href{https://en.wikipedia.org/wiki/Amazon_S3#Design}{Amazon S3 bucket}. A character scalar.}

\item{path}{Path where the downloaded file is written to.}

\item{use_original_filename}{Whether to save the file attachment using its original filename as uploaded. Note that original filenames are \strong{not unique},
i.e. there are multiple file attachments with the same original filename (but differing content, of course). If \code{FALSE}, \code{s3_object_key} is used as
filename. Only relevant if \code{path} is a directory.}

\item{use_testing_server}{Whether or not to use the testing servers instead of the production servers for RDB Services API calls etc.}
}
\value{
A \link[httr:response]{response object}, invisibly.
}
\description{
Downloads a file attachment from the Referendum Database (RDB). The necessary \code{s3_object_key}s identifying individual files are found in the \code{files} list
column returned by \code{\link[=rfrnds]{rfrnds()}}.
}
\examples{
# get object keys
obj_keys <-
  rdb::rfrnds()$files |>
  purrr::list_rbind() |>
  dplyr::filter(!is_deleted) |>
  _$s3_object_key[1:3]

# download them to the current working dir
purrr::walk(obj_keys,
            rdb::download_file_attachment)

# and delete them again
fs::file_delete(obj_keys)
}
\seealso{
Other referendum data functions:
\code{\link{add_rfrnds}()},
\code{\link{assert_vars}()},
\code{\link{count_rfrnds}()},
\code{\link{delete_rfrnds}()},
\code{\link{edit_rfrnds}()},
\code{\link{rfrnd_exists}()},
\code{\link{rfrnds_old}()},
\code{\link{rfrnds}()},
\code{\link{rfrnd}()},
\code{\link{search_rfrnds}()},
\code{\link{validate_rfrnds}()}
}
\concept{rfrnd}
