% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d-GEN.R
\name{download_file_attachment}
\alias{download_file_attachment}
\title{Download file attachment}
\usage{
download_file_attachment(s3_object_key, path = ".")
}
\arguments{
\item{s3_object_key}{The key uniquely identifying the file in the C2D \href{https://en.wikipedia.org/wiki/Amazon_S3#Design}{Amazon S3 bucket}. A character scalar.}

\item{path}{The path where the downloaded file is written to. If a directory, the original filename returned by the C2D services API will be used.}
}
\value{
A \link[httr:response]{response object}, invisibly.
}
\description{
Downloads a file attachment from the C2D database. The necessary \code{s3_object_key}s identifying individual files are found in the \code{files} list column returned
by \code{\link[=referendums]{referendums()}}.
}
\examples{
\dontrun{
# get all file object keys...
c2d::referendums()$files |>
  purrr::map_depth(2L, purrr::pluck, "s3_object_key") |>
  purrr::flatten() |>
  purrr::flatten_chr() |>
  # ...select first three keys...
  magrittr::extract(1:3) |>
  # ...and download the corresponding files to the current working dir
  purrr::walk(c2d::download_file_attachment)}
}
\seealso{
Other referendum data functions:
\code{\link{add_referendums}()},
\code{\link{count_referendums}()},
\code{\link{edit_referendums}()},
\code{\link{referendums}()},
\code{\link{referendum}()},
\code{\link{search_referendums}()},
\code{\link{validate_referendums}()}
}
\concept{referendum}
