% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d.gen.R
\name{edit_rfrnds}
\alias{edit_rfrnds}
\title{Edit existing referendums in the C2D database}
\usage{
edit_rfrnds(
  data,
  email = Sys.getenv("C2D_API_USER"),
  password = Sys.getenv("C2D_API_PW")
)
}
\arguments{
\item{data}{The updated referendum data. A \link[tibble:tbl_df-class]{tibble} that must contain an \href{https://rpkg.dev/c2d/articles/codebook.html#id}{\code{id}} column
identifying the referendums to be edited, an \href{https://rpkg.dev/c2d/articles/codebook.html#is_draft}{\code{is_draft}} column setting their updated draft status,
plus a \link[=codebook]{valid} column containing the new values for each additional database field that should be updated.}

\item{email}{The e-mail address of the C2D API user account to be used for authentication. A character scalar.}

\item{password}{The password of the C2D API user account to be used for authentication. A character scalar.}
}
\value{
\code{data}, invisibly.
}
\description{
Edits existing referendum entries in the C2D database via \href{https://github.com/ccmdesign/c2d-app/blob/master/docs/services.md#3-referendum-routes}{its API}.
}
\details{
Note that the following variables are not supported, i.e. simply dropped from \code{data}:
\itemize{
\item \code{id_official}
\item \code{id_sudd}
\item \code{files}
}
}
\seealso{
Other referendum data functions:
\code{\link{add_rfrnds}()},
\code{\link{count_rfrnds}()},
\code{\link{download_file_attachment}()},
\code{\link{rfrnds}()},
\code{\link{rfrnd}()},
\code{\link{search_rfrnds}()},
\code{\link{validate_rfrnds}()}
}
\concept{rfrnd}
