% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{edit_rfrnds}
\alias{edit_rfrnds}
\title{Edit existing referendums in the RDB}
\usage{
edit_rfrnds(
  data,
  email = pal::pkg_config_val(key = "api_username", pkg = this_pkg),
  password = pal::pkg_config_val(key = "api_password", pkg = this_pkg),
  use_testing_server = pal::pkg_config_val(key = "use_testing_server", pkg = this_pkg)
)
}
\arguments{
\item{data}{Updated referendum data. A \link[tibble:tbl_df-class]{tibble} that must contain an \href{https://rdb.rpkg.dev/articles/codebook.html#id}{\code{id}} column
identifying the referendums to be edited plus any additional columns containing the new values to update the corresponding database fields with. Note that
due to \href{https://github.com/zdaarau/c2d-app/issues/50#issuecomment-1222660683}{current API requirements}, the following columns must always be supplied:
\itemize{
\item \href{https://rdb.rpkg.dev/articles/codebook.html#is-draft}{\code{is_draft}}
\item \href{https://rdb.rpkg.dev/articles/codebook.html#electorate-total}{\code{electorate_total}}
\item \href{https://rdb.rpkg.dev/articles/codebook.html#electorate-abroad}{\code{electorate_abroad}}
\item \href{https://rdb.rpkg.dev/articles/codebook.html#votes-yes}{\code{votes_yes}}
\item \href{https://rdb.rpkg.dev/articles/codebook.html#votes-no}{\code{votes_no}}
\item \href{https://rdb.rpkg.dev/articles/codebook.html#votes-empty}{\code{votes_empty}}
\item \href{https://rdb.rpkg.dev/articles/codebook.html#votes-invalid}{\code{votes_invalid}} .
}}

\item{email}{The e-mail address of the RDB API user account to be used for authentication. A character scalar.}

\item{password}{The password of the RDB API user account to be used for authentication. A character scalar.}

\item{use_testing_server}{Whether or not to use the testing servers instead of the production servers for RDB Services API calls etc.}
}
\value{
\code{data}, invisibly.
}
\description{
Edits existing referendum entries in the  API](https://github.com/zdaarau/c2d-app/blob/master/docs/services.md#3-referendum-routes) via \href{https://github.com/zdaarau/c2d-app/blob/master/docs/services.md#3-referendum-routes}{its API}.
}
\details{
Note that adding/editing the column \code{files} is not supported, i.e. it is simply dropped from \code{data}.
}
\seealso{
Other referendum data functions:
\code{\link{add_rfrnds}()},
\code{\link{assert_vars}()},
\code{\link{count_rfrnds}()},
\code{\link{delete_rfrnds}()},
\code{\link{download_file_attachment}()},
\code{\link{rfrnd_exists}()},
\code{\link{rfrnds}()},
\code{\link{rfrnd}()},
\code{\link{search_rfrnds}()},
\code{\link{validate_rfrnds}()}
}
\concept{rfrnd}
