% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{ggplot_streamgraph}
\alias{ggplot_streamgraph}
\title{Period streamgraph}
\usage{
ggplot_streamgraph(
  data,
  by,
  period = c("week", "month", "quarter", "year", "decade", "century"),
  stacking = c("mirror", "ridge", "proportional"),
  bandwidth = 0.75,
  y_lim = NULL,
  color_palette = viridisLite::turbo,
  prune_legend = TRUE
)
}
\arguments{
\item{data}{RDB referendum data as returned by \code{\link[=rfrnds]{rfrnds()}}. A data frame that at minimum contains the column specified in \code{period} or the column \code{date} (to
compute the \link[=add_period]{period column}), plus the one(s) specified via \code{by} (if any).}

\item{by}{\code{data} column to group by before counting number of referendums.}

\item{period}{Type of period to count referendums by. One of \code{"week"},\code{"month"},\code{"quarter"},\code{"year"},\code{"decade"} or \code{"century"}.}

\item{stacking}{Stacking type. One of
\itemize{
\item \code{"mirror"} to stack absolute values symmetrically around the zero line on the x-axis.
\item \code{"ridge"} to stack absolute values from the zero line on the x-axis upwards.
\item \code{"proportional"} to stack relative values that add up to 100 \%.
}}

\item{bandwidth}{Kernel density estimation bandwidth. A numeric scalar.}

\item{y_lim}{Optional Y axis range limit. Only relevant if \code{stacking} is one of \code{"mirror"} or \code{"ridge"}. The limit applies to the upper side if
\code{stacking = "ridge"} and to both sides if \code{stacking = "mirror"}. A numeric scalar equal or greater than zero.}

\item{color_palette}{Color palette function that when called with a single integer argument returns that many color codes.}

\item{prune_legend}{Whether or not to drop \code{by} factor levels which don't occur in \code{data} from the legend. Only has an effect if \code{by} is of type factor.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2} object.
}
\description{
Creates a ggplot2 \href{https://en.wikipedia.org/wiki/Streamgraph}{streamgraph} based on \code{\link[ggstream:geom_stream]{ggstream::geom_stream()}} using the specified \code{period} as time
resolution.
}
\examples{
data_rdb <- rdb::rfrnds(quiet = TRUE)

rdb::ggplot_streamgraph(data = data_rdb,
                        by = topics_tier_1,
                        period = "year")

# you can specify a different color palette
rdb::ggplot_streamgraph(data = data_rdb,
                        by = topics_tier_1,
                        period = "year",
                        color_palette = viridisLite::viridis)

# by default, only factor levels which occur in data are included in the legend
rdb::ggplot_streamgraph(data = data_rdb,
                        by = topics_tier_1,
                        period = "decade")

# but you can include *all* factor levels in the legend if you want to
rdb::ggplot_streamgraph(data = data_rdb,
                        by = topics_tier_1,
                        period = "decade",
                        prune_legend = FALSE)
}
\seealso{
Other visualization functions:
\code{\link{plot_rfrnd_share_per_period}()},
\code{\link{plot_topic_segmentation}()},
\code{\link{plot_topic_share_per_period}()}
}
\concept{visualize}
