% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d-GEN.R
\name{hierarchize_tags}
\alias{hierarchize_tags}
\title{Hierarchize Tags}
\usage{
hierarchize_tags(x)
}
\arguments{
\item{x}{The tags to hierarchize. Either a character vector of tags or a single-row data frame containing at least the columns \code{tags_tier_1}, \code{tags_tier_2}
and \code{tags_tier_3}.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Reconstructs the hierarchical relations between the three tag variables \code{tags_tier_1}, \code{tags_tier_2} and \code{tags_tier_3}. Can also be used to simply determine
the parent tag(s) of any tag.
}
\examples{
c2d::hierarchize_tags("territorial questions")

# hierarchize the tags of all Austrian referendums
c2d::referendums(country_code = "AT") |>
  dplyr::group_by(id) |>
  dplyr::group_split() |>
  purrr::map(c2d::hierarchize_tags)
}
