% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d.gen.R
\name{infer_tags}
\alias{infer_tags}
\title{Infer higher-tier tags}
\usage{
infer_tags(tags, tier = 1L)
}
\arguments{
\item{tags}{Tags from which the corresponding tags are to be determined. A factor or character vector.}

\item{tier}{Tier of the inferred tags. Either \code{1L} or \code{2L}.}
}
\value{
A character vector.
}
\description{
Determines the top-tier (\code{tier = 1L}) or second-tier (\code{tier = 2L}) tags corresponding to \code{tags} in the
\link[=data_tags]{hierarchy}, i.e. either \code{tags} themselves or their (grand)parent tags.
}
\examples{
c2d::infer_tags(tags = c("EU", "animal protection"),
                tier = 1L)
c2d::infer_tags(tags = c("EU", "animal protection"),
                tier = 2L)

# tags of different tiers can mixed in `tags`
c2d::infer_tags(tags = c("EU", "environment"),
                tier = 2L)

# but `tags` of a higher tier than `tier` will be ignored
c2d::infer_tags(tags = "foreign policy",
                tier = 2L)
}
\seealso{
Other tag-related functions:
\code{\link{data_tags}},
\code{\link{hierarchize_tags_fast}()},
\code{\link{hierarchize_tags}()},
\code{\link{tags}()}
}
\concept{tags}
