% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras_rdb.R
\name{is_cep}
\alias{is_cep}
\title{Testa se o valor eh CEP}
\usage{
is_cep(i, fix_na = TRUE)
}
\arguments{
\item{i}{vetor de numeros CEP}

\item{fix_na}{vetor de strings a serem removidas caso existam.}
}
\description{
Via regex e limpeza de dados, verifica se i eh um cep. Olha-se apenas a estrutura
nao validamos se o cep eh valido nos correios
}
