% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras_rdb.R
\name{is_int64}
\alias{is_int64}
\title{Testa se o valor eh int64 ou lógico}
\usage{
is_int64(i)
}
\arguments{
\item{i}{vetor qualqer}
}
\value{
TRUE se o vetor i é de classe integer64
}
\description{
Testa se o valor eh int64 ou lógico
}
