% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras_rdb.R
\name{is_num}
\alias{is_num}
\title{Testa se o valor eh numero}
\usage{
is_num(i, check = TRUE, fix_na = FALSE)
}
\arguments{
\item{i}{vetor qualqer numerico}

\item{check}{se TRUE (padao) avalia o conteudo de i e infere se a variavel total eh de data}

\item{fix_na}{se tiver algum tipo de na no vetor será removido}
}
\value{
TRUE se o vetor i eh das classes numericas
}
\description{
Permite avaliar, atraves do tipo ou do conteudo, se um vetor de dados recebido eh ou nao no tipo numerico
}
