% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d.gen.R
\name{is_online}
\alias{is_online}
\title{Test C2D API availability}
\usage{
is_online(quiet = FALSE)
}
\arguments{
\item{quiet}{Whether or not to suppress printing a warning in case the API is unavailable.}
}
\value{
A logical scalar.
}
\description{
Checks if the C2D API server is online and operational.
}
\concept{api_status}
