% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d.gen.R
\name{lower_non_abbrs}
\alias{lower_non_abbrs}
\title{Lowercase non-abbreviations}
\usage{
lower_non_abbrs(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector of the same lenght as \code{x}.
}
\description{
Lowercase non-abbreviations
}
\keyword{internal}
