% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{n_rfrnds_per_period}
\alias{n_rfrnds_per_period}
\title{Count number of referendums per period}
\usage{
n_rfrnds_per_period(
  data,
  by = NULL,
  period = c("week", "month", "quarter", "year", "decade", "century"),
  fill_gaps = TRUE,
  period_floor = NULL,
  period_ceiling = NULL,
  descending = FALSE
)
}
\arguments{
\item{data}{RDB referendum data as returned by \code{\link[=rfrnds]{rfrnds()}}. A data frame that at minimum contains the column specified in \code{period} or the column \code{date} (to
compute the \link[=add_period]{period column}), plus the one(s) specified via \code{by} (if any).}

\item{by}{Optional \code{data} column(s) to group by before counting number of referendums. \link[dplyr:dplyr_tidy_select]{Tidy selections} are supported.}

\item{period}{Type of period to count referendums by. One of \code{"week"},\code{"month"},\code{"quarter"},\code{"year"},\code{"decade"} or \code{"century"}.}

\item{fill_gaps}{Whether or not to add zero-value rows to the result for \code{period} gaps in \code{data}.}

\item{period_floor}{Lower \code{period} limit up to which gaps are filled. If \code{NULL}, the lower limit is set to the minimum of \code{period} present in \code{data}. Only
relevant if \code{fill_gaps = TRUE} and \code{period} is set to a unique timespan type (\code{"year"}, \code{"decade"} or \code{"century"}).}

\item{period_ceiling}{Upper \code{period} limit up to which gaps are filled. If \code{NULL}, the upper limit is set to the maximum of \code{period} present in \code{data}. Only
relevant if \code{fill_gaps = TRUE} and \code{period} is set to a unique timespan type (\code{"year"}, \code{"decade"} or \code{"century"}).}

\item{descending}{Whether to sort the resulting table by \code{period} in descending or in ascending order.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Counts the number of RDB referendums per desired period, optionally by additional columns specified via \code{by}.
}
\details{
Note that the \code{period} types \code{"year"}, \code{"decade"} and \code{"century"} are \emph{unique} timespans, while \code{"week"}, \code{"month"} and \code{"quarter"} are \emph{recurring} timespans (e.g. every year has a week 1).
}
\examples{
rdb::rfrnds(country_code = "AT",
            quiet = TRUE) |>
  rdb::n_rfrnds_per_period()

rdb::rfrnds(country_code = "AT",
            quiet = TRUE) |>
  rdb::n_rfrnds_per_period(by = level)

# without filling gaps
rdb::rfrnds(country_code = "AT",
            quiet = TRUE) |>
  rdb::n_rfrnds_per_period(by = level,
                           fill_gaps = FALSE)

# per decade and by multiple columns
rdb::rfrnds(country_code = "AT",
            quiet = TRUE) |>
  rdb::n_rfrnds_per_period(by = c(level, type),
                           period = "decade")

# count ballot dates instead of referendums
rdb::rfrnds(country_code = "AT",
            quiet = TRUE) |>
  rdb::as_ballot_dates() |>
  rdb::n_rfrnds_per_period()
}
\seealso{
Other referendum data transformation functions:
\code{\link{as_ballot_dates}()},
\code{\link{printify_col_names}()}
}
\concept{transform}
