% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{pg_pk}
\alias{pg_pk}
\title{Get primary key column name(s) from PostgreSQL}
\usage{
pg_pk(tbl_name, schema = pg_schema, connection = connect())
}
\arguments{
\item{tbl_name}{\code{\link[base:character]{character(1)}}\cr
Table name.}

\item{schema}{\code{\link[base:character]{character(1)}}\cr
Schema name.}

\item{connection}{\code{\link[RPostgres:PqConnection-class]{PqConnection}}\cr
Database connection to a PostgreSQL DBMS as returned by \code{\link[=connect]{connect()}}.}
}
\value{
A character vector.
}
\description{
Returns the primary key column name(s) from RDB's DBMS, a PostgreSQL server.
}
\seealso{
Other PostgreSQL functions
\code{\link{exec_sql_file}()},
\code{\link{pg_add_default_data}()},
\code{\link{pg_col_metadata}()},
\code{\link{pg_col_nr_to_name}()},
\code{\link{pg_drop_tbls}()},
\code{\link{pg_has_rls}()},
\code{\link{pg_has_tbl}()},
\code{\link{pg_init_aux_tbls}()},
\code{\link{pg_init_db}()},
\code{\link{pg_init_main_tbls}()},
\code{\link{pg_post_nocodb}()},
\code{\link{pg_reset_db}()},
\code{\link{pg_role_pw}()},
\code{\link{pg_tbl_constraints}()},
\code{\link{pg_tbl_del}()},
\code{\link{pg_tbl_keep}()},
\code{\link{pg_tbl_read}()},
\code{\link{pg_tbl_update}()},
\code{\link{pg_update_seq}()},
\code{\link{read_sql_file}()},
\code{\link{split_sql_str}()}
}
\concept{pg}
\keyword{internal}
