% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{pg_tbl_keep}
\alias{pg_tbl_keep}
\title{Delete everything from PostgreSQL table except specified data}
\usage{
pg_tbl_keep(tbl, data, pk_col_name)
}
\arguments{
\item{tbl}{PostgreSQL database table from which data is to be deleted. A \code{\link[dbplyr:tbl.src_dbi]{tbl.src_dbi}} object.}

\item{data}{Table data to keep. A data frame that must contain the table's primary key column. Additional columns are ignored.}

\item{pk_col_name}{Primary key column name. A character scalar.}
}
\value{
The updated table data as a \link[tibble:tbl_df-class]{tibble}, invisibly. Note that newly inserted rows aren't included if the table's primary key column is
\verb{GENERATED ALWAYS}.
}
\description{
Delete everything from PostgreSQL table except specified data
}
\seealso{
Other PostgreSQL functions
\code{\link{exec_sql_file}()},
\code{\link{pg_col_metadata}()},
\code{\link{pg_drop_tbls}()},
\code{\link{pg_has_tbl}()},
\code{\link{pg_init_aux_tbls}()},
\code{\link{pg_init_db}()},
\code{\link{pg_init_main_tbls}()},
\code{\link{pg_pk}()},
\code{\link{pg_reset_db}()},
\code{\link{pg_tbl_del}()},
\code{\link{pg_tbl_update}()},
\code{\link{read_sql_file}()},
\code{\link{split_sql_str}()}
}
\concept{pg}
\keyword{internal}
