% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{pg_tbl_update}
\alias{pg_tbl_update}
\title{Update PostgreSQL table}
\usage{
pg_tbl_update(
  tbl,
  data,
  tbl_name,
  schema = pg_schema,
  sweep = FALSE,
  connection = connect()
)
}
\arguments{
\item{tbl}{PostgreSQL database table to be updated. A \code{\link[dbplyr:tbl.src_dbi]{tbl.src_dbi}} object.}

\item{data}{New or updated table data. A data frame that must contain at least the table's primary key column plus any additional columns with new values to
update the corresponding database fields with.}

\item{tbl_name}{Table name. One of
\itemize{
\item \code{"actors"}
\item \code{"options"}
\item \code{"legal_norms"}
\item \code{"referendum_types"}
\item \code{"referendum_types_legal_norms"}
\item \code{"referendums"}
\item \code{"referendum_titles"}
\item \code{"referendum_questions"}
\item \code{"referendum_positions"}
\item \code{"referendum_votes"}
\item \code{"referendum_sub_votes"}
\item \code{"countries"}
\item \code{"subnational_entities"}
\item \code{"municipalities"}
\item \code{"languages"} .
}}

\item{sweep}{Whether or not to also sweep the table, i.e. delete rows that are not contained in \code{data}.}

\item{connection}{Database connection to a PostgreSQL DBMS as returned by \code{\link[=connect]{connect()}}. An S4 object that inherits from \code{\link[DBI:DBIConnection-class]{DBI::DBIConnection}}.}
}
\value{
The updated table data as a \link[tibble:tbl_df-class]{tibble}, invisibly. Note that newly inserted rows aren't included if the table's primary key column is
\verb{GENERATED ALWAYS}.
}
\description{
Updates the specified PostgreSQL \code{tbl} with values from \code{data}.
}
\details{
If the \code{tbl}'s primary key column is \verb{GENERATED ALWAYS}, rows to be newly added must have \code{NA} values in the corresponding \code{data} column. This is a safety
restriction to avoid accidental insertions.
}
\seealso{
Other PostgreSQL functions
\code{\link{exec_sql_file}()},
\code{\link{pg_col_metadata}()},
\code{\link{pg_drop_tbls}()},
\code{\link{pg_has_tbl}()},
\code{\link{pg_init_aux_tbls}()},
\code{\link{pg_init_db}()},
\code{\link{pg_init_main_tbls}()},
\code{\link{pg_pk}()},
\code{\link{pg_reset_db}()},
\code{\link{pg_role_pw}()},
\code{\link{pg_tbl_constraints}()},
\code{\link{pg_tbl_del}()},
\code{\link{pg_tbl_keep}()},
\code{\link{read_sql_file}()},
\code{\link{split_sql_str}()}
}
\concept{pg}
\keyword{internal}
