% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{pg_update_seq}
\alias{pg_update_seq}
\title{Update PostgreSQL column sequence}
\usage{
pg_update_seq(
  tbl_name,
  col_name = pg_pk(tbl_name = tbl_name, schema = schema, connection = connection),
  schema = pg_schema,
  connection = connect()
)
}
\arguments{
\item{tbl_name}{\code{\link[base:character]{character(1)}}\cr
Table name.}

\item{col_name}{Name of the column whose sequence is to be updated.}

\item{schema}{\code{\link[base:character]{character(1)}}\cr
Schema name.}

\item{connection}{\code{\link[RPostgres:PqConnection-class]{PqConnection}}\cr
Database connection to a PostgreSQL DBMS as returned by \code{\link[=connect]{connect()}}.}
}
\value{
The updated sequence value as an integer.
}
\description{
Updates the current value of the sequence associated with a PostgreSQL table column to the maximum \code{col_name} value if necessary, so that the sequence
is valid to be used for inserting new rows again. Useful after manually overriding \code{col_name} values.
}
\details{
This function obtains a \href{https://www.postgresql.org/docs/current/sql-lock.html}{table-level lock} to avoid concurrency issues during the sequence update.
The sequence is only ever \emph{in}creased and never \emph{de}creased to further defend against concurrency issues in certain edge cases. Finally, \code{pg_update_seq()}
also works on empty tables where the sequence has never been called before.
}
\seealso{
Other PostgreSQL functions
\code{\link{exec_sql_file}()},
\code{\link{pg_add_default_data}()},
\code{\link{pg_col_metadata}()},
\code{\link{pg_col_nr_to_name}()},
\code{\link{pg_drop_tbls}()},
\code{\link{pg_has_rls}()},
\code{\link{pg_has_tbl}()},
\code{\link{pg_init_aux_tbls}()},
\code{\link{pg_init_db}()},
\code{\link{pg_init_main_tbls}()},
\code{\link{pg_pk}()},
\code{\link{pg_post_nocodb}()},
\code{\link{pg_reset_db}()},
\code{\link{pg_role_pw}()},
\code{\link{pg_tbl_constraints}()},
\code{\link{pg_tbl_del}()},
\code{\link{pg_tbl_keep}()},
\code{\link{pg_tbl_read}()},
\code{\link{pg_tbl_update}()},
\code{\link{pg_tbls}()},
\code{\link{read_sql_file}()},
\code{\link{split_sql_str}()}
}
\concept{pg}
\keyword{internal}
