% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{printify_var_names}
\alias{printify_var_names}
\title{Convert referendum data variables to their printable version}
\usage{
printify_var_names(var_names)
}
\arguments{
\item{var_names}{Variable name(s) present in \code{\link{data_codebook}}. A character vector.}
}
\value{
A character vector of the same length as \code{var_names}.
}
\description{
Converts referendum data variables to their ready-to-print version.
}
\examples{
rdb::printify_var_names("topics_tier_1")

# also supports unnested var names
rdb::printify_var_names("topic_tier_1")
}
\seealso{
\code{\link[=printify_col_names]{printify_col_names()}}

Other referendum metadata functions:
\code{\link{data_codebook}},
\code{\link{val_lbls}()},
\code{\link{val_scale}()},
\code{\link{var_name_unnested}()},
\code{\link{var_vals}()}
}
\concept{metadata}
