% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{purge_nocodb}
\alias{purge_nocodb}
\title{Purge NocoDB}
\usage{
purge_nocodb(
  fly_app = "rdb-nocodb",
  s3_bucket = fly_app,
  s3_endpoint = s3_endpoint_url,
  s3_access_key = pal::pkg_config_val("nocodb_s3_access_key"),
  s3_access_secret = pal::pkg_config_val("nocodb_s3_access_secret"),
  ask = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{fly_app}{Fly.io app name.}

\item{s3_bucket}{Name of the S3-compatible object storage bucket to which NocoDB's internal SQLite database is replicated to via Litestream.}

\item{s3_endpoint}{S3-compatible endpoint URL where \code{s3_bucket} is located.}

\item{s3_access_key}{S3-compatible access key ID with access to \code{s3_bucket}.}

\item{s3_access_secret}{S3-compatible access secret with access to \code{s3_bucket}.}

\item{ask}{Whether or not to ask for confirmation before proceeding. Only relevant if run \link[=interactive]{interactively}.}

\item{quiet}{Whether or not to suppress printing status output from internal processing.}
}
\value{
\code{fly_app}, invisibly.
}
\description{
Stops the \code{fly_app} NocoDB instance, purges NocoDB's internal database (the Litestream replication stored under \code{{s3_bucket}/nocodb}) and then restarts the
NocoDB instance, thereby resetting it to factory settings. Beware that this results in \strong{loss of all existing NocoDB-specific metadata} (\href{https://docs.nocodb.com/records/expand-record/#record-comment}{comments}, \href{https://docs.nocodb.com/records/expand-record/#record-audit}{audit} \href{https://docs.nocodb.com/data-sources/actions-on-data-sources/#audit-logs}{logs} etc.).
}
\examples{
\dontrun{
# purge NocoDB production instance
rdb:::purge_nocodb()

# purge NocoDB testing instance
rdb:::purge_nocodb(fly_app = "rdb-nocodb-testing")}
}
\seealso{
Other NocoDB functions
\code{\link{config_nocodb_tbls}()},
\code{\link{create_nocodb_tbl_cols}()},
\code{\link{create_nocodb_tbls}()},
\code{\link{reset_nocodb}()},
\code{\link{reset_rdb}()}

Other administration functions:
\code{\link{backup_rdb}()},
\code{\link{clean_attachment_bucket}()},
\code{\link{config_nocodb_tbls}()},
\code{\link{connect}()},
\code{\link{create_nocodb_tbl_cols}()},
\code{\link{create_nocodb_tbls}()},
\code{\link{delete_rfrnds}()},
\code{\link{dm}()},
\code{\link{notify_postgrest}()},
\code{\link{reset_nocodb}()},
\code{\link{reset_rdb}()},
\code{\link{restart_neon_ep}()},
\code{\link{restart_postgrest}()},
\code{\link{restore_rdb}()},
\code{\link{update_administrative_units}()},
\code{\link{update_countries}()},
\code{\link{update_langs}()},
\code{\link{update_municipalities}()},
\code{\link{update_nocodb_users}()},
\code{\link{update_rfrnds}()},
\code{\link{update_subnational_entities}()},
\code{\link{update_tbl}()},
\code{\link{update_topics}()}
}
\concept{admin}
\concept{nocodb}
\keyword{internal}
