% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDB-package.R
\docType{package}
\name{RDB-package}
\alias{RDB}
\alias{RDB-package}
\title{RDB: Robust Differential Abundance Test in Compositional Data}
\description{
RDB package provides an R package for robust differential abundance analysis.
}
\author{
\strong{Maintainer}: Bo Yuan \email{boyuan5@illinois.edu}

Authors:
\itemize{
  \item Shulei Wang \email{shuleiw@illinois.edu}
}

}
\keyword{internal}
