% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb-package.R
\docType{package}
\name{rdb-package}
\alias{rdb}
\alias{rdb-package}
\title{rdb: Download Data from the Referendum Database (RDB), Which Covers Direct Democratic Votes Worldwide}
\description{
rdb provides access to the Referendum Database (RDB) from R. This database aims to record all direct democratic votes worldwide and is operated by the Centre for Democracy Studies Aarau (ZDA) at the University of Zurich, Switzerland.
}
\section{Package configuration}{


Some of rdb's functionality is controlled via package-specific global configuration which can either be set via \href{https://rdrr.io/r/base/options.html}{R options} or \href{https://en.wikipedia.org/wiki/Environment_variable}{environment variables} (the former take precedence). This configuration includes:\tabular{llll}{
   \strong{Description} \tab \strong{R option} \tab \strong{Environment variable} \tab \strong{Default value} \cr
   RDB PostgreSQL (neon.tech) host address. Defaults to a \href{https://neon.tech/docs/introduction/read-replicas}{read replica} which has no write access to the database. \tab \code{rdb.pg_host} \tab \code{R_RDB_PG_HOST} \tab \code{"ep-frosty-flower-a28keh10.eu-central-1.aws.neon.tech"} \cr
   RDB PostgreSQL username for database operations performed via this package. To be able to edit the database, the user must have sufficient privileges. Defaults to the read-only user \code{r_anon}. \tab \code{rdb.pg_user} \tab \code{R_RDB_PG_USER} \tab \code{"r_anon"} \cr
   Password for \code{pg_user}. Defaults to the password for the read-only user \code{r_anon}. \tab \code{rdb.pg_password} \tab \code{R_RDB_PG_PASSWORD} \tab \code{"WTYUeDtTm0UV4nXzTZ7morHSdM0wQh0p"} \cr
   File path to a UTF-8 encoded CSV file containing the columns \code{role} and \code{password} describing additional roles on the RDB PostgreSQL server. Required for \code{\link[=pg_init_db]{pg_init_db()}} and others. \tab \code{rdb.pg_roles_csv_file} \tab \code{R_RDB_PG_ROLES_CSV_FILE} \tab  \cr
   \href{https://neon.tech/docs/manage/api-keys}{API key} for the RDB neon.tech project \code{neon_rdb}. Required for \code{\link[=restart_neon_ep]{restart_neon_ep()}}. \tab \code{rdb.neon_api_key} \tab \code{R_RDB_NEON_API_KEY} \tab  \cr
   E-mail address of the \href{https://admin.rdb.vote/}{RDB NocoDB server}'s super admin user. Required for \code{\link[=reset_nocodb]{reset_nocodb()}}. \tab \code{rdb.nocodb_email} \tab \code{R_RDB_NOCODB_EMAIL} \tab  \cr
   Password of the \href{https://admin.rdb.vote/}{RDB NocoDB server}'s super admin user Required for \code{\link[=reset_nocodb]{reset_nocodb()}}. \tab \code{rdb.nocodb_password} \tab \code{R_RDB_NOCODB_PASSWORD} \tab  \cr
   File path to a UTF-8 encoded CSV file containing the columns \code{name}, \code{email} and \code{password} describing additional user accounts to create on the \href{https://admin.rdb.vote/}{RDB NocoDB server} via \code{\link[=reset_nocodb]{reset_nocodb()}}. \tab \code{rdb.nocodb_user_account_csv_file} \tab \code{R_RDB_NOCODB_USER_ACCOUNT_CSV_FILE} \tab  \cr
   ID of an access key granting read/write access to the S3-compatible object store holding the \code{rdb-attachments} bucket. Required for various \href{https://rdb.rpkg.dev/reference/#administration}{administration functions}. \tab \code{rdb.nocodb_s3_access_key} \tab \code{R_RDB_NOCODB_S3_ACCESS_KEY} \tab  \cr
   Secret of an access key granting read/write access to the S3-compatible object store holding the \code{rdb-attachments} bucket. Required for various \href{https://rdb.rpkg.dev/reference/#administration}{administration functions}. \tab \code{rdb.nocodb_s3_access_secret} \tab \code{R_RDB_NOCODB_S3_ACCESS_SECRET} \tab  \cr
   Fly.io \href{https://fly.io/docs/flyctl/tokens/}{API token} granting access to the RDB organization. Required for \code{\link[=restart_postgrest]{restart_postgrest()}}. \tab \code{rdb.fly_io_api_token} \tab \code{R_RDB_FLY_IO_API_TOKEN} \tab  \cr
   Maximal timespan to preserve the package's \href{https://pkgpins.rpkg.dev/}{pkgpins} cache. Cache entries older than this will be deleted upon package loading. \tab \code{rdb.global_max_cache_age} \tab \code{R_RDB_GLOBAL_MAX_CACHE_AGE} \tab \code{"30 days"} \cr
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://rdb.rpkg.dev}
  \item \url{https://gitlab.com/zdaarau/rpkgs/rdb}
  \item Report bugs at \url{https://gitlab.com/zdaarau/rpkgs/rdb/issues}
}

}
\author{
\strong{Maintainer}: Salim Brüggemann \email{salim-b@pm.me} (\href{https://orcid.org/0000-0002-5329-5987}{ORCID})

Authors:
\itemize{
  \item Uwe Serdült \email{uwe.serduelt@zda.uzh.ch} (\href{https://orcid.org/0000-0002-2383-3158}{ORCID})
}

Other contributors:
\itemize{
  \item  Centre for Democracy Studies Aarau (ZDA) \email{info@zdaarau.ch} [copyright holder]
}

}
\keyword{internal}
