% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb-package.R
\docType{package}
\name{rdb-package}
\alias{rdb}
\alias{rdb-package}
\title{rdb: Download Data from the Referendum Database (RDB), Which Covers Direct Democratic Votes Worldwide}
\description{
rdb provides access to the Referendum Database (RDB) from R. This database aims to record all direct democratic votes worldwide and is operated by the Centre for Democracy Studies Aarau (ZDA) at the University of Zurich, Switzerland.
}
\section{Package configuration}{


Some of rdb's functionality is controlled via package-specific global configuration which can either be set via \href{https://rdrr.io/r/base/options.html}{R options} or \href{https://en.wikipedia.org/wiki/Environment_variable}{environment variables} (the former take precedence). This configuration includes:\tabular{llll}{
   \strong{Description} \tab \strong{R option} \tab \strong{Environment variable} \tab \strong{Default value} \cr
   NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API authentication token} for the \href{https://admin.rdb.vote/}{RDB NocoDB server}. Used as default for the \verb{*nococdb*()} family of functions. \tab \code{rdb.nocodb_api_token} \tab \code{R_RDB_NOCODB_API_TOKEN} \tab  \cr
   File path to a UTF-8 encoded CSV file containing the columns \code{name}, \code{email} and \code{password} describing additional user accounts to create on the \href{https://admin.rdb.vote/}{RDB NocoDB server} via \code{rdb::reset_nocodb()}. \tab \code{rdb.nocodb_user_account_csv_file} \tab \code{R_RDB_NOCODB_USER_ACCOUNT_CSV_FILE} \tab  \cr
   RDB PostgreSQL (neon.tech) host address. Defaults to a \href{https://neon.tech/docs/introduction/read-replicas}{read replica} which has no write access to the database. \tab \code{rdb.pg_host} \tab \code{R_RDB_PG_HOST} \tab \code{"ep-frosty-flower-a28keh10.eu-central-1.aws.neon.tech"} \cr
   RDB PostgreSQL username. To be able to edit the database, the user must have sufficient privileges. Defaults to the read-only user \code{r_anon}. \tab \code{rdb.pg_user} \tab \code{R_RDB_PG_USER} \tab \code{"r_anon"} \cr
   Password for the specified RDB PostgreSQL user. Defaults to the password for user \code{r_anon}. \tab \code{rdb.pg_password} \tab \code{R_RDB_PG_PASSWORD} \tab \code{"WTYUeDtTm0UV4nXzTZ7morHSdM0wQh0p"} \cr
   Maximal timespan to preserve the package's \href{https://pkgpins.rpkg.dev/}{pkgpins} cache. Cache entries older than this will be deleted upon package loading. \tab \code{rdb.global_max_cache_age} \tab \code{R_RDB_GLOBAL_MAX_CACHE_AGE} \tab \code{"30 days"} \cr
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://rdb.rpkg.dev}
  \item \url{https://gitlab.com/zdaarau/rpkgs/rdb}
  \item Report bugs at \url{https://gitlab.com/zdaarau/rpkgs/rdb/issues}
}

}
\author{
\strong{Maintainer}: Salim Brüggemann \email{salim-b@pm.me} (\href{https://orcid.org/0000-0002-5329-5987}{ORCID})

Authors:
\itemize{
  \item Uwe Serdült \email{uwe.serduelt@zda.uzh.ch} (\href{https://orcid.org/0000-0002-2383-3158}{ORCID})
}

Other contributors:
\itemize{
  \item  Centre for Democracy Studies Aarau (ZDA) \email{info@zdaarau.ch} [copyright holder]
}

}
\keyword{internal}
