% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras_rdb.R
\name{rdb_limpa_cep}
\alias{rdb_limpa_cep}
\title{Limpa CEP}
\usage{
rdb_limpa_cep(cep, fix_na = TRUE)
}
\arguments{
\item{cep}{vetor de números ou strings de CEP}

\item{fix_na}{padroes de NA removidos do número ou string i}
}
\description{
Via regex e limpeza de dados, verifica-se a estrutura do vetor informado, removendo tudo que não for
número e completando com zeros a esquerda casos com menos de 8 digitos. Se o CEP for da forma
99999-999, o traço é removido e o número retorna limpo.
não validamos se o cep é valido nos correios.
}
