% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{read_tbl}
\alias{read_tbl}
\title{Read in RDB table}
\usage{
read_tbl(
  tbl_name = tbl_metadata$name,
  connection = connect(),
  disconnect = TRUE
)
}
\arguments{
\item{tbl_name}{Table name. One of \code{"actors"}, \code{"options"}, \code{"legal_norms"}, \code{"referendum_types"}, \code{"referendums"}, \code{"referendum_titles"}, \code{"referendum_questions"}, \code{"referendum_positions"}, \code{"referendum_votes"}, \code{"referendum_sub_votes"}, \code{"countries"}, \code{"subnational_entities"}, \code{"municipalities"} and \code{"languages"}.}

\item{connection}{Database connection to \link[=connect]{RDB's PostgreSQL DBMS}. An S4 object that inherits from \code{\link[DBI:DBIConnection-class]{DBI::DBIConnection}}.}

\item{disconnect}{Whether or not to \link[DBI:dbDisconnect]{terminate} the database connection when finished.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Fetches a table from the \link[=connect]{RDB DBMS}'s \code{public} schema and returns it as a \link[tibble:tbl_df-class]{tibble}.
}
\examples{
rdb::read_tbl("countries")
}
\seealso{
Other data functions:
\code{\link{assert_vars}()},
\code{\link{delete_rfrnds}()},
\code{\link{download_file_attachment}()},
\code{\link{rfrnd_types}()},
\code{\link{rfrnds}()},
\code{\link{rfrnds_old}()},
\code{\link{validate_rfrnds}()}
}
\concept{data}
