% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{read_tbl}
\alias{read_tbl}
\title{Read in RDB table}
\usage{
read_tbl(
  tbl_name = tbl_metadata$name,
  connection = connect(),
  disconnect = TRUE
)
}
\arguments{
\item{tbl_name}{Table name. One of
\itemize{
\item \code{"actors"}
\item \code{"options"}
\item \code{"legal_norms"}
\item \code{"referendum_types"}
\item \code{"referendums"}
\item \code{"referendum_types_legal_norms"}
\item \code{"referendum_types_referendums"}
\item \code{"referendum_titles"}
\item \code{"referendum_questions"}
\item \code{"referendum_positions"}
\item \code{"referendum_votes"}
\item \code{"referendum_sub_votes"}
\item \code{"supranational_entities"}
\item \code{"countries"}
\item \code{"subnational_entities"}
\item \code{"municipalities"}
\item \code{"languages"} .
}}

\item{connection}{Database connection to a PostgreSQL DBMS as returned by \code{\link[=connect]{connect()}}. An S4 object that inherits from \code{\link[DBI:DBIConnection-class]{DBI::DBIConnection}}.}

\item{disconnect}{Whether or not to \link[DBI:dbDisconnect]{terminate} the database connection when finished.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Fetches a table from the \link[=connect]{RDB DBMS}'s \code{public} schema and returns it as a \link[tibble:tbl_df-class]{tibble}.
}
\examples{
rdb::read_tbl("countries")
}
\seealso{
Other data functions:
\code{\link{assert_vars}()},
\code{\link{delete_rfrnds}()},
\code{\link{download_file_attachments}()},
\code{\link{rfrnd_types}()},
\code{\link{rfrnds}()},
\code{\link{rfrnds_old}()},
\code{\link{validate_rfrnds}()}
}
\concept{data}
