% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{reset_nocodb}
\alias{reset_nocodb}
\title{Reset NocoDB}
\usage{
reset_nocodb(
  hostname = nocodb_hostname,
  b2_bucket = s3_bucket_attachments,
  b2_region = s3_region,
  ask = TRUE,
  quiet = TRUE
)
}
\arguments{
\item{hostname}{NocoDB server \href{https://en.wikipedia.org/wiki/Hostname}{hostname}. A character scalar.}

\item{b2_bucket}{Backblaze B2 bucket name used to store NocoDB's file attachments.}

\item{b2_region}{Identifier of the Backblaze B2 region where the \code{b2_bucket} resides.}

\item{ask}{Whether or not to ask for confirmation before deleting the existing RDB NocoDB base. Only relevant if run \link[=interactive]{interactively}.}

\item{quiet}{Whether or not to suppress printing status output from internal processing.}
}
\value{
\code{hostname}, invisibly.
}
\description{
Completely re-creates the \code{Main} base and adds the RDB team's user accounts on the specified NocoDB server.
Beware that this results in \strong{loss of all existing NocoDB-specific metadata} (\href{https://docs.nocodb.com/records/expand-record/#record-comment}{comments}, \href{https://docs.nocodb.com/records/expand-record/#record-audit}{audit} \href{https://docs.nocodb.com/data-sources/actions-on-data-sources/#audit-logs}{logs} etc.).
}
\seealso{
Other NocoDB functions
\code{\link{config_nocodb_tbls}()},
\code{\link{create_nocodb_tbls}()},
\code{\link{purge_nocodb}()},
\code{\link{reset_rdb}()}

Other administration functions:
\code{\link{clean_rfrnd_attachments}()},
\code{\link{config_nocodb_tbls}()},
\code{\link{connect}()},
\code{\link{create_nocodb_tbls}()},
\code{\link{delete_rfrnds}()},
\code{\link{dm}()},
\code{\link{notify_postgrest}()},
\code{\link{purge_nocodb}()},
\code{\link{reset_rdb}()},
\code{\link{restart_neon_ep}()},
\code{\link{restart_postgrest}()},
\code{\link{update_countries}()},
\code{\link{update_langs}()},
\code{\link{update_municipalities}()},
\code{\link{update_rfrnds}()},
\code{\link{update_subnational_entities}()},
\code{\link{update_tbl}()},
\code{\link{update_topics}()}
}
\concept{admin}
\concept{nocodb}
\keyword{internal}
