% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{reset_nocodb}
\alias{reset_nocodb}
\title{Reset NocoDB}
\usage{
reset_nocodb(
  origin = pal::pkg_config_val("nocodb_origin"),
  ask = TRUE,
  save_user_mapping = TRUE,
  quiet = TRUE
)
}
\arguments{
\item{origin}{\code{\link[base:character]{character(1)}}\cr
NocoDB server \href{https://developer.mozilla.org/docs/Glossary/Origin}{origin}.}

\item{ask}{Whether or not to ask for confirmation before deleting the existing RDB NocoDB base. Only relevant if run \link[=interactive]{interactively}.}

\item{save_user_mapping}{Whether or not to add NocoDB's current user data to the \code{nocodb_users} table via \code{\link[=update_nocodb_users]{update_nocodb_users()}} before resetting it.}

\item{quiet}{\code{\link[base:logical]{logical(1)}}\cr
Whether or not to suppress printing status output from internal processing.}
}
\value{
\code{origin}, invisibly.
}
\description{
Completely re-creates the connection settings and \code{Main} base and adds the RDB team's user accounts on the specified
NocoDB server. Beware that this results in \strong{loss of all existing NocoDB-specific metadata} (\href{https://docs.nocodb.com/records/expand-record/#record-comment}{comments}, \href{https://docs.nocodb.com/records/expand-record/#record-audit}{audit} \href{https://docs.nocodb.com/data-sources/actions-on-data-sources/#audit-logs}{logs} etc.).
}
\seealso{
Other NocoDB functions
\code{\link{config_nocodb_tbls}()},
\code{\link{create_nocodb_tbl_cols}()},
\code{\link{create_nocodb_tbls}()},
\code{\link{purge_nocodb}()},
\code{\link{reset_rdb}()}

Other administration functions:
\code{\link{backup_rdb}()},
\code{\link{clean_attachment_bucket}()},
\code{\link{config_nocodb_tbls}()},
\code{\link{connect}()},
\code{\link{create_nocodb_tbl_cols}()},
\code{\link{create_nocodb_tbls}()},
\code{\link{delete_rfrnds}()},
\code{\link{dm}()},
\code{\link{notify_postgrest}()},
\code{\link{purge_nocodb}()},
\code{\link{reset_rdb}()},
\code{\link{restart_neon_ep}()},
\code{\link{restart_postgrest}()},
\code{\link{restore_rdb}()},
\code{\link{update_administrative_units}()},
\code{\link{update_countries}()},
\code{\link{update_langs}()},
\code{\link{update_municipalities}()},
\code{\link{update_nocodb_users}()},
\code{\link{update_rfrnds}()},
\code{\link{update_subnational_entities}()},
\code{\link{update_tbl}()},
\code{\link{update_topics}()}
}
\concept{admin}
\concept{nocodb}
\keyword{internal}
