% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{reset_rdb}
\alias{reset_rdb}
\title{Reset RDB}
\usage{
reset_rdb(
  hostname = nocodb_hostname,
  connection = connect(),
  init_db = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{hostname}{NocoDB server \href{https://en.wikipedia.org/wiki/Hostname}{hostname}. A character scalar.}

\item{connection}{Database connection to \link[=connect]{RDB's PostgreSQL DBMS}. An S4 object that inherits from \code{\link[DBI:DBIConnection-class]{DBI::DBIConnection}}.}

\item{init_db}{Whether or not to first initialize the RDB PostgreSQL database using \code{\link[=pg_init_db]{pg_init_db()}}.}

\item{quiet}{Whether or not to suppress printing status output from internal processing.}
}
\value{
\code{nocodb_hostname}, invisibly.
}
\description{
Resets the whole RDB PostgreSQL database including all RDB PostgreSQL tables and the RDB NocoDB server.
}
\seealso{
Other administration functions:
\code{\link{config_nocodb_tbls}()},
\code{\link{connect}()},
\code{\link{create_nocodb_tbls}()},
\code{\link{delete_rfrnds}()},
\code{\link{dm}()},
\code{\link{notify_postgrest}()},
\code{\link{reset_nocodb}()},
\code{\link{update_countries}()},
\code{\link{update_langs}()},
\code{\link{update_municipalities}()},
\code{\link{update_rfrnds}()},
\code{\link{update_subnational_entities}()},
\code{\link{update_tbl}()}

Other NocoDB functions
\code{\link{config_nocodb_tbls}()},
\code{\link{create_nocodb_tbls}()},
\code{\link{reset_nocodb}()}
}
\concept{admin}
\concept{nocodb}
