% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{reset_rdb}
\alias{reset_rdb}
\title{Reset RDB}
\usage{
reset_rdb(
  hostname_nocodb = nocodb_hostname,
  hostname_pg = pal::pkg_config_val(key = "pg_host", pkg = this_pkg),
  reset_db = FALSE,
  ask = TRUE,
  disconnect = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{hostname_nocodb}{NocoDB server hostname. A character scalar.}

\item{hostname_pg}{PostgreSQL server hostname. A character scalar.}

\item{reset_db}{Whether or not to first fully recreate the RDB PostgreSQL database using \code{\link[=pg_reset_db]{pg_reset_db()}} and \code{\link[=pg_init_db]{pg_init_db()}}. If \code{FALSE}, only the RDB
PostgreSQL tables are recreated.}

\item{ask}{Whether or not to ask for confirmation before proceeding. Only relevant if run \link[=interactive]{interactively}.}

\item{disconnect}{Whether or not to \link[DBI:dbDisconnect]{terminate} the database connection when finished.}

\item{quiet}{Whether or not to suppress printing status output from internal processing.}
}
\value{
\code{nocodb_hostname}, invisibly.
}
\description{
Resets the whole RDB PostgreSQL database including all RDB PostgreSQL tables and the RDB NocoDB server.
}
\seealso{
Other administration functions:
\code{\link{config_nocodb_tbls}()},
\code{\link{connect}()},
\code{\link{create_nocodb_tbls}()},
\code{\link{delete_rfrnds}()},
\code{\link{dm}()},
\code{\link{notify_postgrest}()},
\code{\link{reset_nocodb}()},
\code{\link{restart_postgrest}()},
\code{\link{update_countries}()},
\code{\link{update_langs}()},
\code{\link{update_municipalities}()},
\code{\link{update_rfrnds}()},
\code{\link{update_subnational_entities}()},
\code{\link{update_tbl}()}

Other NocoDB functions
\code{\link{config_nocodb_tbls}()},
\code{\link{create_nocodb_tbls}()},
\code{\link{reset_nocodb}()}
}
\concept{admin}
\concept{nocodb}
