% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{restart_postgrest}
\alias{restart_postgrest}
\title{Restart RDB PostgREST server}
\usage{
restart_postgrest(
  api_token = pal::pkg_config_val(key = "fly_io_api_token", pkg = this_pkg),
  quiet = FALSE
)
}
\arguments{
\item{quiet}{Whether or not to suppress printing status output from internal processing.}
}
\value{
\code{NULL}, invisibly
}
\description{
Restarts the RDB PostgREST server VM (a \href{https://fly.io/docs/}{Fly.io} app), which causes a \href{https://postgrest.org/en/latest/references/schema_cache.html#schema-cache-reloading}{schema cache reload}.
}
\details{
\code{restart_postgrest()} serves as an alternative to \code{\link[=notify_postgrest]{notify_postgrest()}} when the PostgREST server is not \code{LISTEN}ing on the \code{pgrst} channel (e.g. because it
is not \href{https://postgrest.org/en/v12/references/listener.html#listener-on-read-replicas}{connected to a read-write PostgreSQL instance}).
}
\seealso{
Other PostgREST functions
\code{\link{notify_postgrest}()}

Other administration functions:
\code{\link{clean_attachment_bucket}()},
\code{\link{config_nocodb_tbls}()},
\code{\link{connect}()},
\code{\link{create_nocodb_tbl_cols}()},
\code{\link{create_nocodb_tbls}()},
\code{\link{delete_rfrnds}()},
\code{\link{dm}()},
\code{\link{notify_postgrest}()},
\code{\link{purge_nocodb}()},
\code{\link{reset_nocodb}()},
\code{\link{reset_rdb}()},
\code{\link{restart_neon_ep}()},
\code{\link{update_administrative_units}()},
\code{\link{update_countries}()},
\code{\link{update_langs}()},
\code{\link{update_municipalities}()},
\code{\link{update_rfrnds}()},
\code{\link{update_subnational_entities}()},
\code{\link{update_tbl}()},
\code{\link{update_topics}()}
}
\concept{admin}
\concept{postgrest}
\keyword{internal}
