% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{rfrnd}
\alias{rfrnd}
\title{Get a single referendum's data}
\usage{
rfrnd(
  id,
  incl_archive = FALSE,
  tidy = TRUE,
  use_testing_server = pal::pkg_config_val(key = "use_testing_server", pkg = this_pkg)
)
}
\arguments{
\item{id}{Referendum's unique \href{https://rdb.rpkg.dev/articles/codebook.html#id}{identifier}.}

\item{incl_archive}{Whether or not to include an \code{archive} column containing data from an earlier, obsolete state of the Referendum Database (RDB).}

\item{tidy}{Whether or not to tidy the referendum data, i.e. apply various data cleansing tasks and add additional variables. If \code{FALSE}, the raw MongoDB
referendum data will only be modified just enough to be able to return it as a \link[tibble:tbl_df-class]{tibble}. Note that untidy data doesn't conform to the
\link[=data_codebook]{codebook} (i.a. different variable names).}

\item{use_testing_server}{Whether or not to use the testing servers instead of the production servers for RDB Services API calls etc.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Downloads a single referendum's data from the Referendum Database (RDB). See the \code{\link{codebook}} for a detailed description of all variables.
}
\examples{
rdb::rfrnd(id = "5bbbe26a92a21351232dd73f")
}
\seealso{
Other referendum data functions:
\code{\link{add_rfrnds}()},
\code{\link{assert_vars}()},
\code{\link{count_rfrnds}()},
\code{\link{delete_rfrnds}()},
\code{\link{download_file_attachment}()},
\code{\link{edit_rfrnds}()},
\code{\link{rfrnd_exists}()},
\code{\link{rfrnds}()},
\code{\link{rfrnds_old}()},
\code{\link{search_rfrnds}()},
\code{\link{validate_rfrnds}()}
}
\concept{rfrnd}
