% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{s3_auth}
\alias{s3_auth}
\title{Authenticate S3-compatible object store}
\usage{
s3_auth(
  s3_endpoint = s3_endpoint_url,
  s3_access_key = pal::pkg_config_val("nocodb_s3_access_key"),
  s3_access_secret = pal::pkg_config_val("nocodb_s3_access_secret")
)
}
\arguments{
\item{s3_endpoint}{S3-compatible endpoint URL.}

\item{s3_access_key}{S3-compatible access key ID.}

\item{s3_access_secret}{S3-compatible access secret.}
}
\value{
An object of class \code{\link[s3fs:S3FileSystem]{s3fs::S3FileSystem}}, invisibly.
}
\description{
Executes \code{\link[s3fs:s3_file_system]{s3fs::s3_file_system()}} to authenticate against the specified S3-compatible object storage provider.
}
\concept{s3}
\keyword{internal}
