% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{split_sql_str}
\alias{split_sql_str}
\title{Split SQL string into statements}
\usage{
split_sql_str(.text, .connection = connect(), ...)
}
\arguments{
\item{.text}{SQL statements to split as a character scalar.}

\item{.connection}{Database connection as returned by \code{\link[=connect]{connect()}}. An S4 object that inherits from \code{\link[DBI:DBIConnection-class]{DBI::DBIConnection}}.}

\item{...}{Further (named) arguments used as temporary variables available for substitution while interpolating \code{.text} via \code{\link[DBI:sqlInterpolate]{DBI::sqlInterpolate()}}.}
}
\value{
A list of \code{\link[DBI:SQL]{DBI::SQL}} objects.
}
\description{
Splits a string of SQL into individual SQL statements and interpolates them using \code{\link[DBI:sqlInterpolate]{DBI::sqlInterpolate()}}. Useful to prepare the contents of an \code{.sql} script
to be able to feed it to \code{\link[DBI:dbSendStatement]{DBI::dbSendStatement()}} (which only accepts a single statement per invocation).
}
\seealso{
Other PostgreSQL functions
\code{\link{pg_col_metadata}()},
\code{\link{pg_has_tbl}()},
\code{\link{pg_init_aux_tbls}()},
\code{\link{pg_init_db}()},
\code{\link{pg_init_main_tbls}()},
\code{\link{pg_pk}()},
\code{\link{pg_tbl_del}()},
\code{\link{pg_tbl_keep}()},
\code{\link{pg_tbl_update}()}
}
\concept{pg}
\keyword{internal}
