% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{update_countries}
\alias{update_countries}
\title{Update RDB \code{countries} table}
\usage{
update_countries(sweep = TRUE, connection = connect(), disconnect = TRUE)
}
\arguments{
\item{sweep}{Whether or not to also sweep the table, i.e. delete rows that are not contained in \code{data}.}

\item{connection}{Database connection to a PostgreSQL DBMS as returned by \code{\link[=connect]{connect()}}. An S4 object that inherits from \code{\link[DBI:DBIConnection-class]{DBI::DBIConnection}}.}

\item{disconnect}{Whether or not to \link[DBI:dbDisconnect]{terminate} the database connection when finished.}
}
\value{
The newly written \code{countries} table data as a \link[tibble:tbl_df-class]{tibble}, invisibly.
}
\description{
Updates the RDB's \code{countries} table with the data from \code{\link{data_iso_3166_1}} and \code{\link{data_iso_3166_3}}.
}
\seealso{
Other administration functions:
\code{\link{config_nocodb_tbls}()},
\code{\link{connect}()},
\code{\link{create_nocodb_tbls}()},
\code{\link{delete_rfrnds}()},
\code{\link{dm}()},
\code{\link{notify_postgrest}()},
\code{\link{reset_nocodb}()},
\code{\link{reset_rdb}()},
\code{\link{update_langs}()},
\code{\link{update_municipalities}()},
\code{\link{update_rfrnds}()},
\code{\link{update_subnational_entities}()},
\code{\link{update_tbl}()}
}
\concept{admin}
