% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{update_rfrnds}
\alias{update_rfrnds}
\title{Update RDB \code{referendums} table}
\usage{
update_rfrnds(data, sweep = FALSE, connection = connect(), disconnect = TRUE)
}
\arguments{
\item{data}{New or updated referendum data. A \link[tibble:tbl_df-class]{tibble} that must contain a \href{https://rdb.rpkg.dev/articles/codebook.html#id}{id} column
identifying the referendums to be updated plus any additional columns containing the new values to update the corresponding database fields with.}

\item{sweep}{Whether or not to also sweep the table, i.e. delete rows that are not contained in \code{data}.}

\item{connection}{Database connection to \link[=connect]{RDB's PostgreSQL DBMS}. An S4 object that inherits from \code{\link[DBI:DBIConnection-class]{DBI::DBIConnection}}.}

\item{disconnect}{Whether or not to \link[DBI:dbDisconnect]{terminate} the database connection when finished.}
}
\value{
\code{data}, invisibly.
}
\description{
Updates the RDB's \code{referendums} table with \code{data}. By default, rows not contained in \code{data} are kept.
}
\seealso{
Other administration functions:
\code{\link{config_ncdb}()},
\code{\link{connect}()},
\code{\link{delete_rfrnds}()},
\code{\link{dm}()},
\code{\link{update_countries}()},
\code{\link{update_langs}()},
\code{\link{update_municipalities}()},
\code{\link{update_subnational_entities}()},
\code{\link{update_tbl}()}
}
\concept{admin}
