% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{update_tbl}
\alias{update_tbl}
\title{Update RDB table}
\usage{
update_tbl(
  data,
  tbl_name = tbl_metadata$name,
  sweep = FALSE,
  connection = connect(),
  disconnect = TRUE
)
}
\arguments{
\item{data}{New or updated table data. A data frame that must contain at least the table's primary key column plus any additional columns with new values to
update the corresponding database fields with.}

\item{tbl_name}{Table name. One of
\itemize{
\item \code{"actors"}
\item \code{"options"}
\item \code{"legal_norms"}
\item \code{"referendum_types"}
\item \code{"referendum_types_legal_norms"}
\item \code{"referendums"}
\item \code{"referendum_titles"}
\item \code{"referendum_questions"}
\item \code{"referendum_positions"}
\item \code{"referendum_votes"}
\item \code{"referendum_sub_votes"}
\item \code{"countries"}
\item \code{"subnational_entities"}
\item \code{"municipalities"}
\item \code{"languages"} .
}}

\item{sweep}{Whether or not to also sweep the table, i.e. delete rows that are not contained in \code{data}.}

\item{connection}{Database connection to \link[=connect]{RDB's PostgreSQL DBMS}. An S4 object that inherits from \code{\link[DBI:DBIConnection-class]{DBI::DBIConnection}}.}

\item{disconnect}{Whether or not to \link[DBI:dbDisconnect]{terminate} the database connection when finished.}
}
\value{
\code{data}, invisibly.
}
\description{
Updates the specified RDB table with the provided data.
}
\seealso{
Other administration functions:
\code{\link{config_nocodb_tbls}()},
\code{\link{connect}()},
\code{\link{create_nocodb_tbls}()},
\code{\link{delete_rfrnds}()},
\code{\link{dm}()},
\code{\link{notify_postgrest}()},
\code{\link{reset_nocodb}()},
\code{\link{reset_rdb}()},
\code{\link{update_countries}()},
\code{\link{update_langs}()},
\code{\link{update_municipalities}()},
\code{\link{update_rfrnds}()},
\code{\link{update_subnational_entities}()}
}
\concept{admin}
