% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{url_admin_portal}
\alias{url_admin_portal}
\title{Assemble RDB admin portal URL}
\usage{
url_admin_portal(...)
}
\arguments{
\item{...}{Optional path components added to the base URL.}
}
\value{
A character scalar.
}
\description{
Assemble RDB admin portal URL
}
\examples{
rdb:::url_admin_portal("referendum/5bbbfd7b92a21351232e46b5")
}
\seealso{
Other URL assembly functions:
\code{\link{url_api_old}()},
\code{\link{url_codebook}()},
\code{\link{url_ncdb}()},
\code{\link{url_website}()}
}
\concept{url_assembly}
\keyword{internal}
