% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d.gen.R
\name{url_api}
\alias{url_api}
\title{Assemble C2D Services API URL}
\usage{
url_api(
  ...,
  use_testing_server = pal::pkg_config_val(key = "use_testing_server", pkg = this_pkg)
)
}
\arguments{
\item{...}{Optional path components added to the base URL.}

\item{use_testing_server}{Whether or not to use the testing servers instead of the production servers for C2D Services API calls etc.}
}
\value{
A character scalar.
}
\description{
Assemble C2D Services API URL
}
\examples{
c2d:::url_api("health")
}
\keyword{internal}
