% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{url_website}
\alias{url_website}
\title{Assemble website URL}
\usage{
url_website(...)
}
\arguments{
\item{...}{Optional path components added to the base URL.}
}
\value{
A character scalar.
}
\description{
Assemble website URL
}
\examples{
rdb:::url_website("referendum/CH/5bbc04f692a21351232e5a01")
}
\seealso{
Other URL assembly functions:
\code{\link{url_admin_portal}()},
\code{\link{url_api_old}()},
\code{\link{url_codebook}()}
}
\concept{url_assembly}
\keyword{internal}
