% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2d.gen.R
\name{val_lbls}
\alias{val_lbls}
\title{Get the set of possible \emph{value labels} of a referendum data variable}
\usage{
val_lbls(v_name, incl_affixes = TRUE)
}
\arguments{
\item{v_name}{A variable name. Must be one of the column names of \code{\link{data_codebook}}.}

\item{incl_affixes}{Whether or not to add the corresponding \code{value_label_prefix} and \code{value_label_suffix} to the returned labels.}
}
\value{
A character vector. Of length \code{0} if \code{v_name}'s values are not restricted to a predefined set or no value labels are defined in the
\link[=data_codebook]{codebook}.
}
\description{
Returns a character vector of value labels of a specific \code{\link[=rfrnds]{rfrnds()}} column, in the same order as \code{\link[=v_vals]{v_vals()}}, or of length \code{0} if \code{v_name}'s values are
not restricted to a predefined set or no value labels are defined in the \link[=data_codebook]{codebook}.
}
\examples{
c2d::val_lbls("result",
              incl_affixes = FALSE)
c2d::val_lbls("result")

# Convert the labels to sentence case with trailing dot
c2d::val_lbls("result") |> pal::sentenceify()
}
\seealso{
Other referendum metadata functions:
\code{\link{data_codebook}},
\code{\link{printify_v_names}()},
\code{\link{v_vals}()}
}
\concept{metadata}
