% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{val_scale}
\alias{val_scale}
\title{Get \emph{value scale} of referendum data variable}
\usage{
val_scale(var_name)
}
\arguments{
\item{var_name}{Variable name for which the value scale is to be returned. Must be one of the column names of \code{\link{data_codebook}}.}
}
\value{
A character scalar.
}
\description{
Returns the value scale of a specific \code{\link[=rfrnds]{rfrnds()}} column.
}
\examples{
rdb::val_scale("level")
rdb::val_scale("topics_tier_1")
}
\seealso{
Other referendum metadata functions:
\code{\link{data_codebook}},
\code{\link{printify_var_names}()},
\code{\link{val_lbls}()},
\code{\link{var_name_unnested}()},
\code{\link{var_vals}()}
}
\concept{metadata}
