% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.gen.R
\name{var_name_unnested}
\alias{var_name_unnested}
\title{Get unnested variable names}
\usage{
var_name_unnested(var_names)
}
\arguments{
\item{var_names}{A character vector of RDB referendum data variable name(s). They must be either present in \code{\link{data_codebook}} or added by one of the \href{https://rdb.rpkg.dev/reference/#section-augmentation}{data augmentation functions}.}
}
\value{
A character vector of the same length as \code{var_names}.
}
\description{
Returns the unnested analogue(s) of the specified variable name(s), which result from \link[=unnest_var]{unnesting}. For variable names that do \emph{not} refer to
nested list columns, \code{var_names} is simply returned as-is.
}
\examples{
rdb::var_name_unnested("inst_object_revision_modes")
rdb::var_name_unnested(paste0("topics_tier_", 1:3))
}
\seealso{
\code{\link[=unnest_var]{unnest_var()}}

Other referendum metadata functions:
\code{\link{data_codebook}},
\code{\link{printify_var_names}()},
\code{\link{val_lbls}()},
\code{\link{val_scale}()},
\code{\link{var_vals}()}
}
\concept{metadata}
